% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-visualize-basics.R
\name{add_exposure_summary_to_1D_plot}
\alias{add_exposure_summary_to_1D_plot}
\title{Deprecated: add_exposure_summary_to_1D_plot}
\usage{
add_exposure_summary_to_1D_plot(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
A list of ggplot layers.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{add_exposure_summary_to_1D_plot()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0. Please use \code{\link[=plot_categories]{plot_categories()}} instead.
}
\seealso{
\code{\link[=plot_categories]{plot_categories()}}
}
\keyword{internal}
