% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-information-from-stanfit.R
\name{add_ibbu_stanfit_draws}
\alias{add_ibbu_stanfit_draws}
\title{Add MCMC draws of IBBU parameters to a tibble.}
\usage{
add_ibbu_stanfit_draws(
  fit,
  which = "posterior",
  draws = NULL,
  summarize = FALSE,
  wide = FALSE,
  nest = TRUE,
  category = "category",
  group = "group"
)
}
\arguments{
\item{fit}{mv-ibbu-stanfit object.}

\item{which}{Should parameters for the prior, posterior, or both be added? (default: `"posterior"`)}

\item{draws}{Vector with specific draw(s) to be returned, or `NULL` if all draws are to be returned. (default: `NULL`)}

\item{summarize}{Should the mean of the draws be returned instead of all of the draws? (default: `FALSE`)}

\item{wide}{Should all parameters be returned in one row? (default: `FALSE`)}

\item{nest}{Should the category mean vectors and scatter matrices be nested into one cell each, or should each element
be stored in a separate cell? (default: `TRUE`)}

\item{category}{Name of the category variable. (default: "category")}

\item{group}{Name of the group variable. (default: "group")}
}
\value{
tibble with post-warmup (posterior) MCMC draws of the prior/posterior parameters of the IBBU model
(\code{kappa, nu, m, S, lapse_rate}). \code{kappa} and \code{nu} are the pseudocounts that determine the strength of the beliefs
into the mean and covariance matrix, respectively. \code{m} is the mean of the multivariate normal distribution over category
means mu. \code{S} is the scatter matrix that determines both the covariance of the category means mu, and the
Inverse Wishart distribution over category covariance matrices Sigma.

The expected value of the category mean mu is \code{m}. The expected value of the category covariance matrix Sigma
is \code{S / (nu - D - 1)}, where \code{D} is the dimension of the multivariate Gaussian category. For details,
\insertCite{@see @murphy2012 p. 134;textual}{MVBeliefUpdatr}.
}
\description{
Add MCMC draws of all parameters from incremental Bayesian belief-updating (IBBU) to a tibble. Both wide
(`wide=TRUE`) or long format (`wide=FALSE`) can be chosen as output. By default all post-warmup draws are
returned, but if `summarize=TRUE` then just the mean of each parameter is returned instead.
}
\details{
By default, the category means and scatter matrices are nested, rather than each of their elements being
stored separately (`nest=TRUE`).
}
\examples{
TBD
}
\references{
\insertRef{murphy2012}{MVBeliefUpdatr}
}
\seealso{
TBD
}
\keyword{TBD}
