% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{as.mvg_ibbu_stanfit}
\alias{as.mvg_ibbu_stanfit}
\title{Make an mvg_ibbu_stanfit}
\usage{
as.mvg_ibbu_stanfit(stanfit, input)
}
\arguments{
\item{stanfit}{stanfit object}

\item{input}{Input for a call to rstan prepared for one of the acceptable
stan programs. See \code{\link{compose_data}}.}
}
\value{
An mvg_ibbu_stanfit object
}
\description{
Combines a stanfit object and its input into an mvg_ibbu_stanfit object.
}
\examples{
TBD
}
\seealso{
\code{\link[rstan]{stanfit}}
}
