% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mvg-ibbu-stanfit.R
\name{as.NIW_ibbu_stanfit}
\alias{as.NIW_ibbu_stanfit}
\title{Make an NIW_ibbu_stanfit}
\usage{
as.NIW_ibbu_stanfit(stanfit, input, transform = NULL)
}
\arguments{
\item{stanfit}{stanfit object}

\item{input}{Input for a call to rstan prepared for one of the acceptable
stan programs. See \code{\link{compose_data}}.}

\item{transform}{Optionally, a list of transform (and corresponding untransform) functions that were
applied to the cues prior while creating the input.}
}
\value{
NIW_ibbu_stanfit object
}
\description{
Combines a \code{\link[rstan]{stanfit}} object and its input into an \code{\link{NIW_ibbu_stanfit}} object.
}
\examples{
TBD
}
\seealso{
\code{\link[rstan]{stanfit}}
}
