% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-model-list.R
\name{as_model_list}
\alias{as_model_list}
\title{Helper constructor to create an MVBU_ModelList}
\usage{
as_model_list(models, model_labels = NULL, metadata = list())
}
\arguments{
\item{models}{A list of S7 cognitive model objects or \code{MVBU_StanfitPosterior} objects.}

\item{model_labels}{Optional character vector of names or step labels for each model.}

\item{metadata}{Optional named list of general attributes.}
}
\value{
An \link{MVBU_ModelList} object.
}
\description{
Helper constructor to create an MVBU_ModelList
}
