% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-phase2-migration.R
\name{as_s7_exemplar_model}
\alias{as_s7_exemplar_model}
\title{Coerce legacy exemplar model-like object to S7 model object}
\usage{
as_s7_exemplar_model(x, category = "category", decision_rule = "sampling")
}
\description{
Coerce legacy exemplar model-like object to S7 model object
}
\keyword{internal}
