% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-phase2-migration.R
\name{as_s7_niw_model_distribution}
\alias{as_s7_niw_model_distribution}
\title{Coerce a legacy NIW inferred/fit-like object to S7 model-distribution object}
\usage{
as_s7_niw_model_distribution(x = NULL, group_label = "")
}
\description{
Coerce a legacy NIW inferred/fit-like object to S7 model-distribution object
}
\keyword{internal}
