% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-asserts.R
\name{assert_NIW_belief}
\alias{assert_NIW_belief}
\title{Deprecated: assert_NIW_belief}
\usage{
assert_NIW_belief(x, category = "category", verbose = F, strict = F)
}
\arguments{
\item{x}{Object to test.}

\item{category}{Name of the category field to inspect.}

\item{verbose}{Logical. Whether to emit additional diagnostics while testing the object.}

\item{strict}{Logical. Whether to require the object to be an NIW belief strictly, rather than allowing an ideal adaptor.}
}
\value{
Invisibly TRUE if the assertion passes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{assert_NIW_belief()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0. Please use \link{NIW_IdealAdaptor} or S7 assertions instead.
}
\seealso{
\link{NIW_IdealAdaptor}
}
\keyword{internal}
