% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-model.R
\name{check_compatibility_between_input_and_model}
\alias{check_compatibility_between_input_and_model}
\title{Functions that are shared across the various types of likelihood and model types.}
\usage{
check_compatibility_between_input_and_model(x, model)
}
\description{
Functions that are shared across the various types of likelihood and model types.
}
