% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-NIW-IA-stanfit.R
\name{contains_draws}
\alias{contains_draws}
\alias{contains_draws.stanfit}
\alias{contains_draws.ideal_adaptor_stanfit}
\title{Check whether a stanfit object contains samples}
\usage{
contains_draws(x, ...)

\method{contains_draws}{stanfit}(x)

\method{contains_draws}{ideal_adaptor_stanfit}(x)
}
\arguments{
\item{x}{A \link[rstan:stanfit-class]{rstan::stanfit} object.}
}
\value{
A logical.
}
\description{
Check whether a stanfit object contains samples
}
\seealso{
TBD
}
\keyword{TBD}
