% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R, R/S7-stanfit-diagnostics.R
\name{control_params}
\alias{control_params}
\alias{control_params,MVBU_Stanfit-method}
\title{Extract control parameters of the NUTS sampler}
\usage{
control_params(x, pars = NULL, ...)

## S7 method for class <MVBU_Stanfit>
control_params(x, pars = NULL, ...)
}
\arguments{
\item{x}{An S7 fit object or stanfit object.}

\item{pars}{Optional parameter names.}

\item{...}{Additional arguments.}
}
\value{
A named list of control parameters.
}
\description{
Extract control parameters of the NUTS sampler
}
