% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit-input.R
\name{control_staninput}
\alias{control_staninput}
\title{Specify control parameters for new_ideal_adaptor_stanfit_input()}
\usage{
control_staninput(
  tau_scale = 5,
  L_omega_eta = 1,
  split_loglik_per_observation = 0,
  transform_type = c("identity", "center", "standardize", "PCA whiten", "ZCA whiten")[3]
)
}
\arguments{
\item{tau_scale}{A vector of scales for the Cauchy priors for each cue's standard deviations. Used in
both the prior for m_0 and the prior for S_0. (default: vector of \code{5}s, assuming that the data are standardized).}

\item{L_omega_eta}{A vector of etas of the LKJ prior for the correlations of the covariance matrix of \code{mu_0}. Only used for
models with multivariate categories (e.g., NIW_ideal_adaptor). (default: \code{1},
which corresponds to a uniform prior of correlation matrices)}

\item{split_loglik_per_observation}{Optionally, split the log likelihood per observation. This can be helpful of leave-one-out
estimation in order to avoid high Pareto k, but it also makes the stored stanfit object much larger. (default: \code{0})}

\item{transform_type}{An affine transformation that can be applied to the data. See \code{type} in \code{\link{get_affine_transform}}
for details. (default: "standardize", which standardizes each cue separately)}
}
\value{
A list of control parameters that can be passed to \code{\link{new_ideal_adaptor_stanfit_input}}.
}
\description{
This function is used to specify control parameters for the \code{new_ideal_adaptor_stanfit_input()} function, and to provide
reasonable defaults for any of the unspecified parameters.
}
