% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{cov2ss}
\alias{cov2ss}
\title{Get centered sum-of-squares matrix from covariance matrix.}
\usage{
cov2ss(cov, n)
}
\arguments{
\item{cov}{A covariance matrix.}

\item{n}{Number of observations that have gone into the sum-of-square matrix.}
}
\value{
A square matrix.
}
\description{
Get centered sum-of-square matrix from covariance matrix.
}
\examples{
TBD
}
\seealso{
\code{\link{cov2ss}}
}
\keyword{TBD}
