% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-class-stanfit.R
\name{is.ideal_adaptor_stanfit}
\alias{is.ideal_adaptor_stanfit}
\alias{ideal_adaptor_stanfit}
\title{Deprecated: is.ideal_adaptor_stanfit}
\usage{
is.ideal_adaptor_stanfit(x, verbose = FALSE)

ideal_adaptor_stanfit(
  data = data.frame(),
  staninput = NULL,
  stanvars = NULL,
  backend = "rstan",
  save_pars = NULL,
  stan_args = list(),
  stanfit = NULL,
  basis = NULL,
  transform_information = NULL,
  criteria = list(),
  file = NULL,
  version = NULL,
  metadata = list()
)
}
\arguments{
\item{x}{Object to be checked.}

\item{verbose}{Currently being ignored.}

\item{data}{A \code{data.frame} containing the data used to fit the model.}

\item{staninput}{A Stan input object containing the data handed to rstan
through \code{\link{make_staninput}}. Contains at least two components:
\code{transformed} and \code{untransformed}.}

\item{stanvars}{A \code{stanvars} object or \code{NULL}. Default: \code{NULL}.}

\item{backend}{The name of the backend used to fit the model (character).
Default: \code{"rstan"}.}

\item{save_pars}{Optional storage for saved parameter names. Default:
\code{NULL}.}

\item{stan_args}{Named list of additional control arguments that were passed
to the Stan backend directly. Default: \code{list()}.}

\item{stanfit}{An object of class \code{\link[rstan:stanfit-class]{stanfit}}
containing posterior draws. Default: \code{NULL}.}

\item{basis}{An object that contains a small subset of the Stan data created
at fitting time, needed to process new data. Default: \code{NULL}.}

\item{transform_information}{An object of type
\code{\link{MVBU_TransformInformation}} or \code{NULL}. Default: \code{NULL}.}

\item{criteria}{An empty \code{list} for adding model fit criteria after
estimation of the model. Default: \code{list()}.}

\item{file}{Optional name of a file in which the model object was stored in
or loaded from. Default: \code{NULL}.}

\item{version}{The versions of \pkg{MVBeliefUpdatr} and \pkg{rstan} with
which the model was fitted. Default: \code{NULL}.}

\item{metadata}{List containing auxiliary information including
\code{label_information}. Default: \code{list()}.}
}
\value{
A logical.

An object of class \code{MVBU_Stanfit}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{is.ideal_adaptor_stanfit()} is deprecated; use
\code{S7::S7_inherits(x, IdealAdaptorStanfit)} instead.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{ideal_adaptor_stanfit()} is deprecated; use the S7 constructor
\code{\link[=MVBU_Stanfit]{MVBU_Stanfit()}} or \code{\link[=fit_ideal_adaptor]{fit_ideal_adaptor()}} instead.
}
\keyword{internal}
