% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-basics.R
\name{get_sum_of_squares_from_df}
\alias{get_sum_of_squares_from_df}
\alias{get_sum_of_uncentered_squares_from_df}
\alias{get_sum_of_centered_squares_from_df}
\title{Deprecated: Get sum of squares from a data frame or matrix}
\usage{
get_sum_of_squares_from_df(
  data,
  variables = NULL,
  center = TRUE,
  verbose = FALSE
)

get_sum_of_uncentered_squares_from_df(data, variables = NULL, verbose = FALSE)

get_sum_of_centered_squares_from_df(data, variables = NULL, verbose = FALSE)
}
\arguments{
\item{data}{A \code{tibble}, \code{data.frame}, or \code{matrix}.}

\item{variables}{Only required if data is not already a \code{matrix}.}

\item{center}{Logical; if \code{TRUE}, compute centered sum of squares.}

\item{verbose}{Logical; produce verbose output.}
}
\value{
A matrix.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_sum_of_squares_from_df()} and its aliases are deprecated as of MVBeliefUpdatr 0.1.0
and will be removed in 0.2.0. Please use \code{\link[=get_sufficient_category_statistics]{get_sufficient_category_statistics()}} instead.
}
\keyword{internal}
