% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-phase2-migration.R
\name{.as_s7_mnix_ideal_adaptor}
\alias{.as_s7_mnix_ideal_adaptor}
\title{Coerce legacy MNIX ideal adaptor-like object to S7 model object}
\usage{
.as_s7_mnix_ideal_adaptor(x, category = "category", decision_rule = "sampling")
}
\description{
Coerce legacy MNIX ideal adaptor-like object to S7 model object
}
\keyword{internal}
