% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-phase2-migration.R
\name{.as_s7_mvg_ideal_observer}
\alias{.as_s7_mvg_ideal_observer}
\title{Coerce legacy MVG ideal observer-like object to S7 model object}
\usage{
.as_s7_mvg_ideal_observer(x, category = "category", decision_rule = "sampling")
}
\description{
Coerce legacy MVG ideal observer-like object to S7 model object
}
\keyword{internal}
