% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit.R
\name{.check_stanfit_file}
\alias{.check_stanfit_file}
\title{Normalize a fit file path to include the .rds suffix.}
\usage{
.check_stanfit_file(file)
}
\arguments{
\item{file}{File path or base name.}
}
\value{
A character scalar with the normalized file path.
}
\description{
Normalize a fit file path to include the .rds suffix.
}
\keyword{internal}
