% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit.R
\name{.contains_draws}
\alias{.contains_draws}
\title{Check whether a Stanfit object contains posterior draws.}
\usage{
.contains_draws(x, ...)
}
\arguments{
\item{x}{Object to inspect.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A logical scalar.
}
\description{
Check whether a Stanfit object contains posterior draws.
}
\keyword{internal}
