% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit.R
\name{.file_refit_options}
\alias{.file_refit_options}
\title{Return the supported refit-policy options for cached Stanfits.}
\usage{
.file_refit_options()
}
\value{
A character vector of supported options.
}
\description{
Return the supported refit-policy options for cached Stanfits.
}
\keyword{internal}
