% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-core-classes.R
\name{.get_model_family_registration}
\alias{.get_model_family_registration}
\title{Get the registration information for a model family.}
\usage{
.get_model_family_registration(family)
}
\arguments{
\item{family}{Model family name.}
}
\value{
A list with the registered class names.
}
\description{
Get the registration information for a model family.
}
\keyword{internal}
