% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit.R
\name{.make_parnames}
\alias{.make_parnames}
\title{Build parameter names for a rectangular grid of indices.}
\usage{
.make_parnames(prefix, ...)
}
\arguments{
\item{prefix}{Character prefix for the parameter names.}

\item{...}{Arguments passed to \code{expand.grid}.}
}
\value{
A character vector of parameter names.
}
\description{
Build parameter names for a rectangular grid of indices.
}
\keyword{internal}
