% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-core-classes.R
\name{.mvbu_normalize_sigma_noise}
\alias{.mvbu_normalize_sigma_noise}
\title{Normalize perceptual noise covariance matrix.}
\usage{
.mvbu_normalize_sigma_noise(Sigma_noise, cue_labels)
}
\description{
Normalize perceptual noise covariance matrix.
}
\keyword{internal}
