% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-core-classes.R
\name{.new_niw_model_distribution}
\alias{.new_niw_model_distribution}
\title{Construct a NIW model distribution object}
\usage{
.new_niw_model_distribution(
  cache = list(),
  metadata = list(),
  group_label = ""
)
}
\description{
Construct a NIW model distribution object
}
\keyword{internal}
