% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit.R
\name{.read_ideal_adaptor_stanfit}
\alias{.read_ideal_adaptor_stanfit}
\title{Read a cached ideal adaptor Stanfit from disk.}
\usage{
.read_ideal_adaptor_stanfit(file)
}
\arguments{
\item{file}{File path to the cached fit.}
}
\value{
A cached fit object or \code{NULL} if none is available.
}
\description{
Read a cached ideal adaptor Stanfit from disk.
}
\keyword{internal}
