% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit.R
\name{.rename_pars}
\alias{.rename_pars}
\title{Rename fitted parameter names to their transformed equivalents.}
\usage{
.rename_pars(x, include_original_pars = FALSE)
}
\arguments{
\item{x}{An ideal adaptor Stanfit object.}

\item{include_original_pars}{Whether to preserve original parameter names.}
}
\value{
The updated Stanfit object.
}
\description{
Rename fitted parameter names to their transformed equivalents.
}
\keyword{internal}
