% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit.R
\name{.stanfit_needs_refit}
\alias{.stanfit_needs_refit}
\title{Check if cached \code{ideal_adaptor_stanfit} can be used.}
\usage{
.stanfit_needs_refit(
  x,
  current_version = get_current_versions(),
  data = NULL,
  staninput = NULL,
  silent = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{x}{Old \code{ideal_adaptor_stanfit} object (e.g., loaded from file).}

\item{current_version}{Current version of relevant packages. (default: will be automatically
obtained from current packages).}

\item{data}{New data to check consistency of factor level names. (default: \code{NULL}))}

\item{staninput}{New Stan data (result of a call to \code{\link{make_staninput}}).
Pass \code{NULL} to avoid this data check. (default: \code{NULL}))}

\item{silent}{Logical. If \code{TRUE}, no messages will be given. (default: \code{FALSE}))}

\item{verbose}{Logical. If \code{TRUE} detailed report of the differences
is printed to the console. (default: \code{FALSE}))}
}
\value{
A boolean indicating whether a refit is needed.
}
\description{
Checks whether a given cached fit can be used without refitting when
\code{file_refit = "on_change"} is used.
}
\details{
fit differs from the given data and code.
}
\keyword{internal}
