% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-visualize-basics.R
\name{ellipse.pmap}
\alias{ellipse.pmap}
\title{Deprecated: ellipse.pmap}
\usage{
\method{ellipse}{pmap}(x, centre, level, ...)
}
\arguments{
\item{x}{Covariance matrix.}

\item{centre}{Center vector.}

\item{level}{Probability level.}

\item{...}{Additional arguments.}
}
\value{
Ellipse coordinates matrix.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{ellipse.pmap()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0.
}
\keyword{internal}
