% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-objects.R
\name{example_data}
\alias{example_data}
\title{Example data for different types of models.}
\usage{
example_data(model_type, example = 1, ..., Ns = 20)
}
\description{
Example data for different types of models.
}
