% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-objects.R
\name{example_exemplar_model}
\alias{example_exemplar_model}
\alias{example_MVG_ideal_observer}
\alias{example_NIW_ideal_adaptor}
\title{Example exemplar model.}
\usage{
example_exemplar_model(
  example = 1,
  categories = c("/b/", "/p/"),
  cues = c("VOT", "f0_semitones", "vowel_duration")[1:example],
  prior = rep(1/length(categories), length(categories)),
  lapse_rate = 0.05,
  lapse_bias = prior,
  Sigma_noise = matrix(rep(0, length(cues)^2), nrow = length(cues), dimnames = list(cues,
    cues)),
  verbose = T
)

example_MVG_ideal_observer(
  example = 1,
  categories = c("/b/", "/p/"),
  cues = c("VOT", "f0_semitones", "vowel_duration")[1:example],
  prior = rep(1/length(categories), length(categories)),
  lapse_rate = 0.05,
  lapse_bias = prior,
  Sigma_noise = matrix(rep(0, length(cues)^2), nrow = length(cues), dimnames = list(cues,
    cues)),
  verbose = T
)

example_NIW_ideal_adaptor(
  example = 1,
  categories = c("/b/", "/p/"),
  cues = c("VOT", "f0_semitones", "vowel_duration")[1:example],
  prior = rep(1/length(categories), length(categories)),
  lapse_rate = 0.05,
  lapse_bias = prior,
  Sigma_noise = matrix(rep(0, length(cues)^2), nrow = length(cues), dimnames = list(cues,
    cues)),
  verbose = T
)
}
\description{
Example exemplar model.

Example MVG ideal observers.

Example NIW ideal adaptors.
}
