% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-objects.R
\name{example_MVG_ideal_observer}
\alias{example_MVG_ideal_observer}
\title{Example MVG ideal observers.}
\usage{
example_MVG_ideal_observer(
  example = 1,
  categories = c("/b/", "/p/"),
  cues = c("VOT", "f0_semitones", "vowel_duration")[1:example],
  prior = rep(1/length(categories), length(categories)),
  lapse_rate = 0.05,
  lapse_bias = prior,
  Sigma_noise = matrix(rep(0, length(cues)^2), nrow = length(cues), dimnames = list(cues,
    cues))
)
}
\description{
Example MVG ideal observers.
}
