% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-visualize-basics.R
\name{facet_or_animate}
\alias{facet_or_animate}
\title{Deprecated: facet_or_animate}
\usage{
facet_or_animate(
  p,
  facet_rows_by,
  facet_cols_by,
  facet_wrap_by,
  animate_by,
  animation_follow
)
}
\arguments{
\item{p}{A \code{ggplot} object.}

\item{facet_rows_by}{Row facet variable.}

\item{facet_cols_by}{Col facet variable.}

\item{facet_wrap_by}{Wrap facet variable.}

\item{animate_by}{Animation variable.}

\item{animation_follow}{Logical whether animation follows data.}
}
\value{
A \code{ggplot} or animated object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{facet_or_animate()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0.
}
\keyword{internal}
