% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_my.R
\name{get_aggregates_from_grouped_data_as_list_of_lists}
\alias{get_aggregates_from_grouped_data_as_list_of_lists}
\alias{get_aggregates_from_grouped_data_as_list_of_arrays}
\title{Get aggregates from grouped data as list of lists or list of arrays}
\usage{
get_aggregates_from_grouped_data_as_list_of_lists(
  data,
  groups,
  cols,
  fill = NA,
  verbose = F,
  ...
)

get_aggregates_from_grouped_data_as_list_of_arrays(
  data,
  groups,
  cols,
  fill = NA,
  simplify = as.list(rep(T, length(list(...)))),
  verbose = F,
  ...
)
}
\arguments{
\item{data}{`tibble` or `data.frame` with the data.}

\item{groups}{Character vector of column names that contain information about which observations form a group. These columns must
factors.}

\item{cols}{Character vector of column names for which aggregate values should be obtained. This can be more than one column, but only one
aggregate will be returned *across* columns, i.e., the aggregate will be calculated for each unique group in the data.}

\item{fill}{What value should be substituted for `groups` for which there is no data (see ...)? (default: `NA`)}

\item{...}{A named list of functions (statistics) of `cues` to calculate for each unique `group` in the data.}
}
\value{
A named list of length `...` (names are the names of the functions that have been computed). The list will be
  sorted by the `group` variables (in the order they are presented), in ascending order of the levels of those variables.
  The elements of the list will be lists.
  Missing values---resulting from combinations of `group` variables for which there is no data---will be filled
  with the value provided in `fill` and coerced into the same structure as all other outputs for that function
  (e.g., if the function f results in a 2-element vector for all combinations of `group` variables for which
  there is data, then unobserved data results in a 2-element vector of `fill` values).
}
\description{
Get aggregates from grouped data. Each function is calculated for each unique combination
of values in the columns `category` and (optionally) `group`.
}
\keyword{TBD}
