% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-MVG-IO.R
\name{get_categorization_from_MVG_ideal_observer}
\alias{get_categorization_from_MVG_ideal_observer}
\title{Get categorization from an ideal observer}
\usage{
get_categorization_from_MVG_ideal_observer(
  x,
  model,
  decision_rule,
  noise_treatment = if (decision_rule == "sampling") "sample" else "marginalize",
  lapse_treatment = if (decision_rule == "sampling") "sample" else "marginalize",
  simplify = F
)
}
\arguments{
\item{x}{A vector of observations.}

\item{model}{An \code{\link[=is.MVG_ideal_observer]{MVG_ideal_observer}} object.}

\item{decision_rule}{Must be one of "criterion", "proportional", or "sampling".}

\item{noise_treatment}{Determines whether and how multivariate Gaussian noise is considered during categorization.
See \code{\link[=get_MVG_likelihood]{get_MVG_likelihood}}. (default: "sample" if decision_rule is "sample"; "marginalize" otherwise).}

\item{lapse_treatment}{Determines whether and how lapses will be treated. Can be "no_lapses", "sample" or "marginalize".
If "sample", whether a trial is lapsing or not will be sampled for each observations. If a trial is sampled to be
a lapsing trial the lapse biases are used as the posterior for that trial. If "marginalize", the posterior probability
will be adjusted based on the lapse formula lapse_rate * lapse_bias + (1 - lapse_rate) * posterior probability from
perceptual model. (default: "sample" if decision_rule is "sample"; "marginalize" otherwise).}

\item{simplify}{Should the output be simplified, and just the label of the selected category be returned? This
option is only available for the criterion and sampling decision rules. (default: `FALSE`)}
}
\value{
Either a tibble of observations with posterior probabilities for each category (in long format), or a
character vector indicating the chosen category in the same order as the observations in x (if simplify = `TRUE`).
}
\description{
Categorize a single observation based on an MVG ideal observer The decision rule can be specified to be either the
criterion choice rule, proportional matching (Luce's choice rule), or the sampling-based interpretation of
Luce's choice rule.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
