% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-information-from-NIW-belief.R
\name{get_categorization_from_NIW_belief}
\alias{get_categorization_from_NIW_belief}
\title{Get categorization from an NIW belief}
\usage{
get_categorization_from_NIW_belief(x, belief, decision_rule, simplify = F)
}
\arguments{
\item{x}{An observation.}

\item{belief}{An \code{\link[=is.NIW_belief]{NIW_belief}} object.}

\item{decision_rule}{Must be one of "criterion", "proportional", or "sampling".}

\item{simplify}{Should the output be simplified, and just the label of the selected category be returned? This
option is only availale for the criterion and sampling decision rules. (default: `FALSE`)}
}
\value{
Either a vector of posterior probabilities of the same length as the number of categories in the NIW
belief object, or a character indicating the chosen category (if simplify = `TRUE`).
}
\description{
Categorize a single observation based on an NIW belief. The decision rule can be specified to be either the
criterion choice rule, proportional matching (Luce's choice rule), or the sampling-based interpretation of
Luce's choice rule.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
