% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit-methods.R
\name{get_categorization_function}
\alias{get_categorization_function}
\title{Get categorization function from a model object}
\usage{
get_categorization_function(
  x,
  lapse_treatment = c("no_lapses", "sample", "marginalize")[3],
  groups = get_group_labels(x, include_prior = TRUE),
  ...
)
}
\arguments{
\item{x}{A model object.}

\item{lapse_treatment}{Should the consequences of attentional lapses be
included in the categorization function ("marginalize") or not
("no_lapses")? (default: "marginalize")}

\item{groups}{Character vector of groups to include.}

\item{...}{Optionally, additional arguments handed to \code{\link{get_draws}}.}
}
\value{
A tibble with categorization functions per group and draw.
}
\description{
Returns a categorization function that can be used to categorize a set of
cues into categories.
}
