% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-fit.R
\name{get_categorization_function_from_grouped_ibbu_draws}
\alias{get_categorization_function_from_grouped_ibbu_draws}
\title{Get categorization function from grouped IBBU draws}
\usage{
get_categorization_function_from_grouped_ibbu_draws(fit, ...)
}
\description{
Convenience function intended for internal use.
}
