% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-belief.R
\name{get_category_labels_from_NIW_belief}
\alias{get_category_labels_from_NIW_belief}
\title{Get category labels from NIW belief object}
\usage{
get_category_labels_from_NIW_belief(x)
}
\arguments{
\item{x}{An NIW belief object.}
}
\description{
Get the unique labels for all categories from an NIW belief object.
}
