% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R
\name{get_category_posterior}
\alias{get_category_posterior}
\title{Get posterior category probabilities for one or more observations.}
\usage{
get_category_posterior(x, new_data, ...)
}
\arguments{
\item{x}{A cognitive model object.}

\item{new_data}{A numeric matrix of observations or a list of matrices.}
}
\value{
A matrix of posterior category probabilities with one row per observation.
}
\description{
Get posterior category probabilities for one or more observations.
}
