% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R
\name{get_category_posterior_prediction}
\alias{get_category_posterior_prediction}
\title{Get both posterior category probabilities and category predictions.}
\usage{
get_category_posterior_prediction(x, new_data, ...)
}
\arguments{
\item{x}{A cognitive model object.}

\item{new_data}{A numeric matrix of observations or a list of matrices.}
}
\value{
A list containing the posterior matrix and the category prediction.
}
\description{
Get both posterior category probabilities and category predictions.
}
