% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-information-from-stanfit.R
\name{get_category_statistic}
\alias{get_category_statistic}
\title{Get category mean mu or covariance matrix sigma from an NIW belief MCMC object.}
\usage{
get_category_statistic(x, grouping.vars = NULL, statistic = c("mu", "Sigma"))
}
\arguments{
\item{x}{\code{\link{NIW_ibbu_stanfit}} or NIW belief MCMC object.}

\item{statistic}{Which category statistic should be returned? `mu` for category mean or `Sigma` for category
covariance matrix, or `c("mu", "Sigma")` for both. (default: both)}

\item{category}{Character vector with categories (or category) for which category statistics are to be
returned.  If `NULL` then all categories are included. (default: `NULL`)}

\item{group}{Character vector with groups (or group) for which category statistics are to be
returned. If `NULL` then all groups are included. (default: `NULL`)}
}
\description{
Returns the category means mu and/or category covariance matrix Sigma for the exposure data from an NIW
IBBU stanfit or NIW belief MCMC object.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
