% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-staninput.R
\name{get_category_statistics_as_list_of_arrays}
\alias{get_category_statistics_as_list_of_arrays}
\title{Get category statistics from data as list of arrays}
\usage{
get_category_statistics_as_list_of_arrays(
  data,
  group = "group",
  category = "category",
  cues,
  fill = as.list(rep(0, length(list(...)))),
  simplify = as.list(rep(T, length(list(...)))),
  verbose = F,
  ...
)
}
\arguments{
\item{data}{\code{tibble} or \code{data.frame} with the data. Each row should be an observation of a category,
and contain information about the category label, the cue values of the observation, and optionally grouping variables.}

\item{group}{Name of column that contains information about which observations form a group. This column must be
a factor.}

\item{category}{Name of column that contains the category label for the exposure data. This column must be a factor.
Can be \code{NULL} for unsupervised updating (not yet implemented). (default: "category")}

\item{cues}{Names of columns with cue values.}

\item{fill}{A value to fill in for missing combinations of \code{group} and \code{category} for which there is no data. (default: \code{0})}

\item{simplify}{A list of logicals of the same length as \code{...} indicating whether the array resulting from each function
should be simplified? See \code{\link{to_array}} for more detail. (default: \code{TRUE} for all functions)}

\item{...}{A named list of functions (statistics) of \code{cues} to calculate for each unique combination of \code{group} and
\code{category}.}
}
\value{
A named list of length \code{...} (names are the names of the functions that have been computed). The list will be
sorted first by \code{group} and then by \code{category}, in ascending order of the levels of those variables. The elements of
the list will be arrays.
Missing values---resulting from combinations of \code{group} and \code{category} for which there is no data---will be filled
with \code{fill} values coerced into the same structure as all other outputs for that function (e.g., if the function f results
in a 2-element vector for all combinations of \code{group} and \code{category} for which there is data, then unobserved data
results in a 2-element vector of \code{fill} values).
}
\description{
Convenience function that obtains category statistics (aggregate functions) over \code{cues} in a data frame, and prepares them to be
used as input to a MVBeliefUpdatr Stan program. Each function is calculated for each unique combination of values in the columns
\code{category} and (optionally) \code{group}. Missing values will be automatically filled since Stan can't handle \code{NA} inputs.
}
\keyword{TBD}
