% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-class.R
\name{get_class}
\alias{get_class}
\title{Deprecated: get_class}
\usage{
get_class(x)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_class()} was deprecated in MVBeliefUpdatr 0.1.0 and will be removed in 0.2.0.
Please use \code{\link[=get_representation_type]{get_representation_type()}} or S7 classes instead.
}
\seealso{
\code{\link[=get_representation_type]{get_representation_type()}}
}
\keyword{internal}
