% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get-constructors.R
\name{get_constructor}
\alias{get_constructor}
\alias{get_category_constructor}
\alias{get_group_constructor}
\alias{get_cue_constructor}
\alias{get_cue2_constructor}
\title{Deprecated: get_constructor}
\usage{
get_constructor(x, variable = NULL)

get_category_constructor(x)

get_group_constructor(x)

get_cue_constructor(x)

get_cue2_constructor(x)
}
\arguments{
\item{x}{An \code{\link{IdealAdaptorStanfit}} or \code{\link{MVBU_Stanfit}} object.}

\item{variable}{Either "category", "group", "cue", or "cue2". If set to
\code{NULL} then a list of all constructors is returned. (default: \code{NULL})}
}
\value{
A constructor function, a list of constructor functions, or \code{NULL}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_constructor()}, \code{get_category_constructor()}, \code{get_group_constructor()},
\code{get_cue_constructor()}, and \code{get_cue2_constructor()} are deprecated in favor of
\code{\link{get_category_labels}}, \code{\link{get_group_labels}}, and
\code{\link{get_cue_labels}}.
}
\keyword{internal}
