% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R, R/S7-stanfit-posterior.R,
%   R/S7-model-list.R, R/S7-methods.R, R/S7-stanfit-methods.R
\name{get_cue_labels}
\alias{get_cue_labels}
\alias{get_cue_labels,MVBU_StanfitPosterior-method}
\alias{get_cue_labels,MVBU_ModelList-method}
\alias{get_cue_labels,MVBU_Object-method}
\alias{get_cue_labels,MVBU_Stanfit-method}
\title{Get cue labels from a representation, template, or model}
\usage{
get_cue_labels(x, indices = NULL, ...)

## S7 method for class <MVBU_StanfitPosterior>
get_cue_labels(x, indices = NULL, ...)

## S7 method for class <MVBU_ModelList>
get_cue_labels(x, indices = NULL, ...)

## S7 method for class <MVBU_Object>
get_cue_labels(x, indices = NULL, ...)

## S7 method for class <MVBU_Stanfit>
get_cue_labels(x, indices = NULL, ...)
}
\arguments{
\item{x}{A representation, representation template, or cognitive model.}

\item{indices}{An optional integer vector of indices to subset the returned labels.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A character vector of cue labels.
}
\description{
Get cue labels from a representation, template, or model
}
