% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-MCMC.R
\name{get_cue_labels_fom_MCMC}
\alias{get_cue_labels_fom_MCMC}
\title{Get cue labels from an NIW belief MCMC object [DEPRECATED: integrated with get_cue_labels_from_model]}
\usage{
get_cue_labels_fom_MCMC(x, indices = NULL)
}
\arguments{
\item{x}{mv_ibbu_stanfit object.}

\item{indeces}{A vector of indices that should be turned into the original cue labels corresponding to those
indices, or `NULL` if all cue labels should be returned. (default: `NULL`)}
}
\value{
A character vector.
}
\description{
Returns the cue labels for the specified indices extracted from the m column of the NIW belief MCMC object.
If no indices are provided, then all cue labels are returned.
}
