% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-information-from-NIW-belief.R
\name{get_cue_labels_from_NIW_belief}
\alias{get_cue_labels_from_NIW_belief}
\title{Get cue labels from NIW belief object}
\usage{
get_cue_labels_from_NIW_belief(x)
}
\arguments{
\item{x}{An NIW belief object.}
}
\description{
Get the names for all cues from an NIW belief object.
}
