% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-visualize-basics.R
\name{get_default_colors}
\alias{get_default_colors}
\title{Deprecated: get_default_colors}
\usage{
get_default_colors(var, levels)
}
\arguments{
\item{var}{Variable name (\code{"category"} or \code{"group"}).}

\item{levels}{Character vector of factor levels.}
}
\value{
Character vector of color values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_default_colors()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0.
}
\keyword{internal}
