% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-visualize-basics.R
\name{get_default_linetypes}
\alias{get_default_linetypes}
\title{Deprecated: get_default_linetypes}
\usage{
get_default_linetypes(var, levels)
}
\arguments{
\item{var}{Variable name.}

\item{levels}{Character vector of factor levels.}
}
\value{
Integer vector of linetypes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_default_linetypes()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0.
}
\keyword{internal}
