% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-basics.R
\name{get_default_colors}
\alias{get_default_colors}
\alias{get_default_shapes}
\alias{get_default_linetypes}
\title{Get default colors, linetype, etc.}
\usage{
get_default_colors(var, levels)

get_default_shapes(var, levels)

get_default_linetypes(var, levels)
}
\arguments{
\item{var}{Variable for which default color/linetype is requested. Can be either \code{"category"} or
\code{"group"}.}

\item{levels}{Character vector of levels of the variable.}
}
\value{
Character vector of color/linetype values in the order specified by \code{levels}.
}
\description{
Get default colors, linetype, etc.
}
