% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{get_expected_category_statistic_from_stanfit}
\alias{get_expected_category_statistic_from_stanfit}
\alias{get_expected_mu_from_stanfit}
\alias{get_expected_sigma_from_stanfit}
\title{Get expected category mean mu or covariance matrix sigma}
\usage{
get_expected_category_statistic_from_stanfit(
  x,
  categories = get_category_levels_from_stanfit(x),
  groups = get_group_levels_from_stanfit(x, include_prior = TRUE),
  statistic = c("mu", "Sigma"),
  untransform_cues = TRUE
)

get_expected_mu_from_stanfit(x, ...)

get_expected_sigma_from_stanfit(x, ...)
}
\arguments{
\item{x}{An \code{\link[=is.NIW_ideal_adaptor_stanfit]{mv_ibbu_stanfit}} object.}

\item{categories}{Character vector with categories for which category statistics are to be
returned. (default: all categories)}

\item{groups}{Character vector with groups for which category statistics are to be returned.
(default: all groups)}

\item{statistic}{Which category statistic should be returned? `mu` for category mean or `Sigma` for category
covariance matrix, or `c("mu", "Sigma")` for both. (default: both)}

\item{untransform_cues}{Should m_0 and S_0 be transformed back into the original cue space? (default: `TRUE`)}
}
\value{
If just one group and category was requested, a vector (for the mean) or matrix (for the covariance
matrix). If more than one group or category was requested, a tibble with one row for each unique combination
of group and category.
}
\description{
Returns the expected value of posterior marginal distribution over category means mu and/or
category covariance matrix Sigma, marginalized over all MCMC samples.
}
\details{
Each MCMC samples' expected value for the category mean \code{E[mu] = m_n}
(i.e, the posterior/updated mean of the multivariate Normal over category means \code{mu}).
Marginalizing across all MCMC samples (representing uncertainty in the true value of
\code{m_n}), we get \code{E[E[mu]] = mean(m_n)}.

Each MCMC samples' expected value for the category covariance matrix
\code{E[Sigma] = S_n / (nu_n - D - 1)}, where \code{S_n} is the posterior/updated scatter matrix,
\code{nu_n} is the posterior/updated pseudocount representing the strength of the posterior/updated
beliefs over category covariance matrices sigma (i.e., the inverse-Wishart), and \code{D} is
the dimension of the multivariate Normal. Marginalizing across all MCMC samples
(representing uncertainty in the true value of \code{S_n}), we get
\code{E[E[Sigma]] = mean(S_n / (nu_n - D - 1))}.
}
\references{
\insertRef{murphy2012}{MVBeliefUpdatr}
}
\seealso{
TBD
}
\keyword{TBD}
