% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIW.R
\name{get_expected_mu}
\alias{get_expected_mu}
\alias{get_expected_sigma}
\title{Get expected category mean mu or covariance matrix sigma}
\usage{
get_expected_mu(x, category, group)

get_expected_sigma(x, category, group)
}
\description{
Returns the expected value of posterior marginal distribution over category means mu or
category covariance matrix Sigma, marginalized over all MCMC samples.
}
\details{
Each MCMC samples' expected value \code{E[mu] = M_n}
(i.e, the posterior/updated mean of the mulativariate Normal over category means \code{mu}).
Marginalizing across all MCMC samples (representing uncertainty in the true value of
\code{M_n}), we get \deq{E\[E\[mu\]\] = mean(M_n)}.

Each MCMC samples' expected value
\code{E[Sigma] = S_n / (nu_n - D - 1)}, where \code{S_n} is the posterior/updated scatter matrix,
\code{nu_n} is the posterior/updated pseudocount representing the strength of the posterior/updated
beliefs over category covariance matrices sigma (i.e., the inverse-Wishart), and \code{D} is
the dimension of the multivariate Normal. Marginalizing across all MCMC samples
(representing uncertainty in the true value of \code{S_n}), we get
\code{E[E[Sigma]] = mean(S_n / (nu_n - D - 1))}.
}
\examples{
TBD
}
\references{
Murphy, K. P. (2012). Machine learning: a probabilistic perspective. MIT press.
}
\seealso{
TBD
}
\keyword{TBD}
