% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-niw-nix-conversions.R
\name{get_expected_mu_from_m}
\alias{get_expected_mu_from_m}
\alias{get_m_from_expected_mu}
\title{Convert Between Location Hyperparameter and Expected Mean Vector}
\usage{
get_expected_mu_from_m(m)

get_m_from_expected_mu(mu)
}
\arguments{
\item{m}{Mean vector \eqn{m} (numeric vector, matrix, or list).}

\item{mu}{Expected category mean vector \eqn{\mu} (numeric vector, matrix, or list).}
}
\value{
Expected category mean \eqn{\mu = m}, or location parameter \eqn{m = \mu}.
}
\description{
Functions to convert between the prior/posterior location hyperparameter \eqn{m}
and the expected category mean vector \eqn{\mu} under a Normal-Inverse-Wishart (NIW)
or Normal-Inverse-Chi-Squared (NIX) distribution \insertCite{@see @murphy2012 p. 134}{MVBeliefUpdatr}.
}
\details{
For conjugate NIW and NIX models, the expected mean \eqn{\mu} is identically equal
to the location parameter \eqn{m}:
\deqn{\mu = m}
}
\references{
\insertRef{murphy2012}{MVBeliefUpdatr}
}
\seealso{
\code{\link{get_expected_mu}}, \code{\link{get_expected_Sigma_from_S}}
}
