% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R, R/S7-stanfit-methods.R
\name{get_exposure_data}
\alias{get_exposure_data}
\alias{get_exposure_data,MVBU_Stanfit-method}
\title{Get the exposure data from an S7 stanfit or stanfit-input object.}
\usage{
get_exposure_data(x, ...)

## S7 method for class <MVBU_Stanfit>
get_exposure_data(x, groups = get_group_labels(x, include_prior = FALSE), ...)
}
\arguments{
\item{x}{An S7 stanfit or stanfit-input object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A tibble containing the exposure data.
}
\description{
Get the exposure data from an S7 stanfit or stanfit-input object.
}
\seealso{
\code{\link{get_data}}, \code{\link{get_test_data}}
}
