% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{get_data}
\alias{get_data}
\alias{get_data.ideal_adaptor_stanfit}
\alias{get_exposure_data}
\alias{get_exposure_data.ideal_adaptor_stanfit}
\alias{get_test_data}
\alias{get_test_data.ideal_adaptor_stanfit}
\title{Get the exposure and/or test data from an ideal adaptor stanfit.}
\usage{
get_data(x, ...)

\method{get_data}{ideal_adaptor_stanfit}(
  x,
  groups = get_group_levels(x, include_prior = FALSE),
  .rename_to_MVB_default = FALSE,
  ...
)

get_exposure_data(x, ...)

\method{get_exposure_data}{ideal_adaptor_stanfit}(x, groups = get_group_levels(x, include_prior = FALSE), ...)

get_test_data(x, ...)

\method{get_test_data}{ideal_adaptor_stanfit}(
  x,
  groups = get_group_levels(x, include_prior = FALSE),
  .from_staninput = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{\link{ideal_adaptor_stanfit}} object.}

\item{...}{additional arguments to \code{\link{get_staninput}}.}

\item{groups}{Character vector of groups for which test data is requested. Typically, the levels of these factors
are automatically added to the fit during the creation of the fit. If necessary, however, it is possible to use
\link[tidybayes:recover_types]{tidybayes::recover_types} on the stanfit object to add or change these levels later.
(default: all categories/groups will be selected)}

\item{.rename_to_MVB_default}{Should the data columns be renamed to the default names used internally in
\code{MVBeliefUpdatr}? his option is included primarily for internal use. (default: \code{FALSE})}

\item{.from_staninput}{Should the data be extracted from the staninput (rather than data) object stored in the
stanfit object? This returns the data in a somewhat different format. This option is included primarily for
internal use. (default: \code{FALSE})}
}
\value{
A \code{tibble} in which each row is a test token. Columns include the cues
and the response counts (one column per category) for all test tokens and all groups.
}
\description{
Returns the exposure and/or test data used during the creation of the \link[rstan:stanfit]{rstan::stanfit}.
object.
}
\seealso{
TBD
}
\keyword{TBD}
