% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R, R/S7-stanfit-posterior.R,
%   R/S7-methods.R, R/S7-stanfit-methods.R
\name{get_group_labels}
\alias{get_group_labels}
\alias{get_group_labels,MVBU_StanfitPosterior-method}
\alias{get_group_labels,MVBU_Object-method}
\alias{get_group_labels,MVBU_Stanfit-method}
\title{Get group labels from a model, stanfit, or stanfit-input object}
\usage{
get_group_labels(x, indices = NULL, include_prior = FALSE, ...)

## S7 method for class <MVBU_StanfitPosterior>
get_group_labels(x, indices = NULL, include_prior = FALSE, ...)

## S7 method for class <MVBU_Object>
get_group_labels(x, indices = NULL, include_prior = FALSE, ...)

## S7 method for class <MVBU_Stanfit>
get_group_labels(x, indices = NULL, include_prior = FALSE, ...)
}
\arguments{
\item{x}{A model, stanfit, or stanfit-input object.}

\item{indices}{An optional integer vector of indices to subset the returned labels.}

\item{include_prior}{Logical; whether to include the prior group label. Default: \code{FALSE}.
Only applicable if \code{x} is an \code{\link{MVBU_Stanfit}} or \code{\link{MVBU_StanfitPosterior}}
object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A character vector of group labels.
}
\description{
Get group labels from a model, stanfit, or stanfit-input object
}
\seealso{
\code{\link{get_cue_labels}}, \code{\link{get_category_labels}}, \code{\link{get_labels}}
}
