% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{get_input_from_stanfit}
\alias{get_input_from_stanfit}
\title{Get the input data from an NIW ideal adaptor stanfit.}
\usage{
get_input_from_stanfit(fit)
}
\arguments{
\item{fit}{\code{\link{NIW_ideal_adaptor_stanfit}} object.}
}
\value{
A list with element names and structures determined by the type of stanfit model.
}
\description{
Returns the inputs handed to \code{stan} or \code{sampling} during the creation of the \code{stanfit}
object.
}
\seealso{
TBD
}
\keyword{TBD}
