% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-information-from-stanfit.R
\name{get_ibbu_stanfit_exposure_category_statistic}
\alias{get_ibbu_stanfit_exposure_category_statistic}
\alias{get_ibbu_stanfit_exposure_mean}
\alias{get_ibbu_stanfit_exposure_Sigma}
\title{Get category mean mu or covariance matrix sigma of exposure data for IBBU}
\usage{
get_ibbu_stanfit_exposure_category_statistic(
  x,
  category = NULL,
  group = NULL,
  statistic = c("mu", "Sigma")
)

get_ibbu_stanfit_exposure_mean(x, category, group)

get_ibbu_stanfit_exposure_Sigma(x, category, group)
}
\arguments{
\item{x}{An mv_ibbu_stanfit or NIW belief MCMC object.}

\item{category}{Character vector with categories (or category) for which category statistics are to be
returned.  If `NULL` then all categories are included. (default: `NULL`)}

\item{group}{Character vector with groups (or group) for which category statistics are to be
returned. If `NULL` then all groups are included. (default: `NULL`)}

\item{statistic}{Which category statistic should be returned? `mu` for category mean or `Sigma` for category
covariance matrix, or `c("mu", "Sigma")` for both. (default: both)}
}
\value{
If just one group and category was requested, a vector (for the mean) or matrix (for the covariance
matrix). If more than one group or category was requested, a tibble with one row for each unique combination
of group and category.
}
\description{
Returns the category means mu and/or category covariance matrix Sigma for the exposure data for an incremental
Bayesian belief-updating (IBBU) model from an IBBU stanfit or NIW belief MCMC object.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
