% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R, R/S7-stanfit-posterior.R,
%   R/S7-methods.R, R/S7-stanfit-methods.R
\name{get_labels}
\alias{get_labels}
\alias{get_labels,MVBU_StanfitPosterior-method}
\alias{get_labels,MVBU_Object-method}
\alias{get_labels,MVBU_Stanfit-method}
\title{Get all labels from a representation, template, model, stanfit, or fit-input object}
\usage{
get_labels(x, ...)

## S7 method for class <MVBU_StanfitPosterior>
get_labels(x, ...)

## S7 method for class <MVBU_Object>
get_labels(x, ...)

## S7 method for class <MVBU_Stanfit>
get_labels(x, ...)
}
\arguments{
\item{x}{A representation, representation template, cognitive model, stanfit,
or fit-input object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A named list of character vectors with elements \code{cue},
\code{category}, \code{response_category}, and \code{group}.
}
\description{
Get all labels from a representation, template, model, stanfit, or fit-input object
}
\seealso{
\code{\link{set_labels}}, \code{\link{get_cue_labels}},
\code{\link{get_category_labels}},
\code{\link{get_response_category_labels}},
\code{\link{get_group_labels}}
}
