% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get-info-from-model.R
\name{get_lapse_biases_from_model}
\alias{get_lapse_biases_from_model}
\title{Deprecated: get_lapse_biases_from_model}
\usage{
get_lapse_biases_from_model(model, categories = NULL)
}
\arguments{
\item{model}{A model object.}

\item{categories}{Optional vector of category labels to subset.}
}
\value{
Numeric vector of lapse biases.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_lapse_biases_from_model()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0. Please use \code{\link[=get_lapse_bias]{get_lapse_bias()}} instead.
}
\seealso{
\code{\link[=get_lapse_bias]{get_lapse_bias()}}
}
\keyword{internal}
