% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-MVG-IO.R
\name{get_likelihood}
\alias{get_likelihood}
\alias{get_likelihood_from_MVG}
\title{Get likelihood}
\usage{
get_likelihood(x, mu, Sigma, log = T)

get_likelihood_from_MVG(
  x,
  model,
  log = T,
  category = "category",
  category.label = NULL,
  wide = FALSE
)
}
\arguments{
\item{x}{Observations. Can be a vector with k elements for a single observation or a matrix with k
columns and n rows, in which case each row of the matrix is taken to be one observation. If x is a
tibble with k columns or a list of vectors of length k, it is reduced into a matrix with k columns.}

\item{mu}{The category mean mu. Should be a
matrix or vector of length k.}

\item{Sigma}{The category covariance matrix Sigma. Should be a square k x k matrix.}

\item{log}{Should the log-transformed density be returned (`TRUE`)? (default: `TRUE`)}
}
\description{
Get likelihood of observations x given the MVG parameters mu and Sigma. This is the density of
a multivariate normal distribution.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
