% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-visualize-basics.R
\name{get_limits}
\alias{get_limits}
\title{Deprecated: get_limits}
\usage{
get_limits(data, measure, by = NULL, hdi.prob = 0.99, min = NULL, max = NULL)
}
\arguments{
\item{data}{A \code{tibble} or \code{data.frame}.}

\item{measure}{Variable name.}

\item{by}{Optional grouping variable.}

\item{hdi.prob}{HDI probability mass.}

\item{min}{Optional minimum.}

\item{max}{Optional maximum.}
}
\value{
Numeric vector of length 2.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_limits()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0.
}
\keyword{internal}
