% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R, R/S7-stanfit-posterior.R,
%   R/S7-model-list.R
\name{get_metadata}
\alias{get_metadata}
\alias{get_metadata,MVBU_StanfitPosterior-method}
\alias{get_metadata,MVBU_ModelList-method}
\title{Get metadata from an object}
\usage{
get_metadata(x, ...)

## S7 method for class <MVBU_StanfitPosterior>
get_metadata(x, ...)

## S7 method for class <MVBU_ModelList>
get_metadata(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A list containing metadata associated with the object.
}
\description{
Get metadata from an object
}
