% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R, R/S7-stanfit-posterior.R
\name{get_model_family}
\alias{get_model_family}
\alias{get_model_family,MVBU_StanfitPosterior-method}
\title{Get the model family from an object}
\usage{
get_model_family(x, ...)

## S7 method for class <MVBU_StanfitPosterior>
get_model_family(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A character string representing the model family.
}
\description{
Get the model family from an object
}
