% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R, R/S7-stanfit-methods.R,
%   R/S7-stanfit-input-methods.R
\name{get_model_type}
\alias{get_model_type}
\alias{get_model_type,MVBU_Stanfit-method}
\alias{get_model_type,IdealAdaptorStaninput-method}
\alias{get_model_type,IdealAdaptorStanfitInput-method}
\title{Get the model type of a cognitive model or model distribution}
\usage{
get_model_type(x, ...)

## S7 method for class <MVBU_Stanfit>
get_model_type(x, ...)

## S7 method for class <IdealAdaptorStaninput>
get_model_type(x, ...)

## S7 method for class <IdealAdaptorStanfitInput>
get_model_type(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A character scalar identifying the model family/type.
}
\description{
Get the model type of a cognitive model or model distribution
}
