% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get-info-from-NIW-IA.R
\name{get_NIW_categorization_function}
\alias{get_NIW_categorization_function}
\alias{get_categorization_function_from_NIW_ideal_adaptor}
\title{Deprecated: get_NIW_categorization_function}
\usage{
get_NIW_categorization_function(
  ms,
  Ss,
  kappas,
  nus,
  priors = rep(1/length(ms), length(ms)),
  lapse_rate = 0,
  lapse_biases = rep(1/length(ms), length(ms)),
  Sigma_noise = NULL,
  noise_treatment = .infer_noise_treatment(Sigma_noise),
  lapse_treatment = if (lapse_rate > 0) "marginalize" else "no_lapses"
)

get_categorization_function_from_NIW_ideal_adaptor(model, ...)
}
\arguments{
\item{ms}{Means of the multivariate normal distributions over category means.}

\item{Ss}{Scatter matrices of the inverse Wishart distribution over category covariance matrices.}

\item{kappas}{Strength of the beliefs into the distribution over category means.}

\item{nus}{Strength of the beliefs into the distribution over category covariance matrices.}

\item{priors}{Vector of categories' prior probabilities.}

\item{lapse_rate}{A lapse rate for the categorization responses.}

\item{lapse_biases}{A lapse bias for the categorization responses.}

\item{Sigma_noise}{A noise matrix.}

\item{noise_treatment}{Noise treatment.}

\item{lapse_treatment}{Lapse treatment.}

\item{model}{A model object.}

\item{...}{Additional arguments.}
}
\value{
A categorization function.

A categorization function.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_NIW_categorization_function()} is deprecated. Use \code{\link{get_category_posterior_function}} instead.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_categorization_function_from_NIW_ideal_adaptor()} is deprecated. Use \code{\link{get_category_posterior_function}} instead.
}
\keyword{internal}
