% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R, R/S7-stanfit-methods.R
\name{get_number_of_draws}
\alias{get_number_of_draws}
\alias{get_random_draw_indices}
\alias{get_number_of_draws,any-method}
\alias{get_random_draw_indices,any-method}
\title{Get number of post-warmup MCMC samples from stanfit}
\usage{
get_number_of_draws(fit, ...)

get_random_draw_indices(fit, ndraws = NULL, ...)

## S7 method for class <any>
get_number_of_draws(fit, ...)

## S7 method for class <any>
get_random_draw_indices(fit, ndraws = NULL, ...)
}
\arguments{
\item{fit}{A \code{\link[rstan:stanfit-class]{stanfit}} object or \code{\link{MVBU_Stanfit}} object.}

\item{...}{Additional arguments passed to methods.}

\item{ndraws}{Number of indices to be returned.}
}
\description{
Get number of post-warmup MCMC samples from stanfit

Get indices for random MCMC draws from stanfit
}
