% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-visualize-basics.R
\name{get_plot_limits}
\alias{get_plot_limits}
\title{Deprecated: get_plot_limits}
\usage{
get_plot_limits(plot)
}
\arguments{
\item{plot}{A \code{ggplot} object.}
}
\value{
List with elements \code{x} and \code{y}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_plot_limits()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0.
}
\keyword{internal}
