% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIW.R
\name{get_posterior_predictive}
\alias{get_posterior_predictive}
\title{Get posterior predictive}
\usage{
get_posterior_predictive(x, M, S, kappa, nu, log = T)
}
\arguments{
\item{x}{Input (observation).}

\item{M}{The mean of the multivariate Normal distribution of the category mean mu.}

\item{S}{The scatter matrix of the inverse-Wishart distribution over the category covariance
matrix Sigma.}

\item{kappa}{The strength of the beliefs over the category mean (pseudocounts).}

\item{nu}{The strength of the beliefs over the category covariance matrix (pseudocounts).}

\item{log}{Should the log-transformed density be returned (`TRUE`)? (default: `TRUE`)}
}
\description{
Get posterior predictive of observation x given the NIW parameters M, S, kappa, and nu. This is
a multivariate Student-T distribution (Murphy, 2012, p. 135).
}
\examples{
TBD
}
\references{
Murphy, K. P. (2012). Machine learning: a probabilistic perspective. MIT press.
}
\seealso{
TBD
}
\keyword{TBD}
