% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get-info-from-NIW-belief.R
\name{get_posterior_predictive_from_NIW_belief}
\alias{get_posterior_predictive_from_NIW_belief}
\alias{get_posterior_predictives_from_NIW_beliefs}
\title{Deprecated: get_posterior_predictive_from_NIW_belief}
\usage{
get_posterior_predictive_from_NIW_belief(
  x,
  model,
  noise_treatment = "no_noise",
  log = TRUE,
  category = "category",
  category.label = NULL,
  wide = FALSE
)

get_posterior_predictives_from_NIW_beliefs(
  x,
  model,
  noise_treatment = "no_noise",
  log = TRUE,
  category = "category",
  category.label = NULL,
  grouping.var = NULL,
  wide = FALSE
)
}
\arguments{
\item{x}{Observations.}

\item{model}{An NIW belief or adaptor model.}

\item{noise_treatment}{Noise treatment.}

\item{log}{Logical; whether log likelihood is returned.}

\item{category}{Category column name.}

\item{category.label}{Category labels.}

\item{wide}{Logical; whether wide format is returned.}

\item{grouping.var}{Grouping variable name.}
}
\value{
Posterior predictive data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_posterior_predictive_from_NIW_belief()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0. Please use \code{\link[=likelihood]{likelihood()}} instead.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_posterior_predictives_from_NIW_beliefs()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0. Please use \code{\link[=likelihood]{likelihood()}} instead.
}
\seealso{
\code{\link[=likelihood]{likelihood()}}
}
\keyword{internal}
