% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R, R/S7-stanfit-posterior.R,
%   R/S7-methods.R, R/S7-stanfit-input-methods.R
\name{get_response_category_labels}
\alias{get_response_category_labels}
\alias{get_response_category_labels,MVBU_StanfitPosterior-method}
\alias{get_response_category_labels,MVBU_Object-method}
\alias{get_response_category_labels,IdealAdaptorStanfitInput-method}
\title{Get response category labels from an object}
\usage{
get_response_category_labels(x, indices = NULL, ...)

## S7 method for class <MVBU_StanfitPosterior>
get_response_category_labels(x, indices = NULL, ...)

## S7 method for class <MVBU_Object>
get_response_category_labels(x, indices = NULL, ...)

## S7 method for class <IdealAdaptorStanfitInput>
get_response_category_labels(x, indices = NULL, ...)
}
\arguments{
\item{x}{An object (e.g., \code{\link{IdealAdaptorStanfitInput}},
\code{\link{MVBU_Stanfit}}, or \code{\link{MVBU_StanfitPosterior}}).}

\item{indices}{An optional integer vector of indices to subset the returned labels.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A character vector of response category labels.
}
\description{
Get response category labels from an object
}
\seealso{
\code{\link{get_category_labels}}, \code{\link{get_group_labels}}, \code{\link{get_cue_labels}}
}
