% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIW.R
\name{get_S_from_expected_Sigma}
\alias{get_S_from_expected_Sigma}
\title{Get Scatter matrix S from expected category covariance Sigma and pseudocount nu}
\usage{
get_S_from_expected_Sigma(Sigma, nu)
}
\description{
Get Scatter matrix S from expected category covariance Sigma and pseudocount nu
}
\seealso{
\code{\link{get_Sigma_from_S}}
}
