% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R, R/S7-stanfit-methods.R
\name{get_staninput}
\alias{get_staninput}
\alias{get_staninput,any-method}
\alias{get_staninput,MVBU_Stanfit-method}
\title{Get the stored Stan input from an S7 fit or fit-input object}
\usage{
get_staninput(x, ...)

## S7 method for class <any>
get_staninput(x, ...)

## S7 method for class <MVBU_Stanfit>
get_staninput(x, ...)
}
\arguments{
\item{x}{An S7 fit or fit-input object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
The stored S7 Stan input object.
}
\description{
Get the stored Stan input from an S7 fit or fit-input object
}
