% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-staninput.R
\name{get_statistics_as_list_of_arrays}
\alias{get_statistics_as_list_of_arrays}
\title{Get statistics from a data set}
\usage{
get_statistics_as_list_of_arrays(data, group, category, cues, verbose = F, ...)
}
\arguments{
\item{data}{`tibble` or `data.frame` with the data. Each row should be an observation of a category,
and contain information about the category label, the cue values of the observation, and optionally grouping variables.}

\item{group}{Name of column that contains information about which observations form a group. This column must be
a factor.}

\item{category}{Name of column that contains the category label for the exposure data. This column must be a factor.
Can be `NULL` for unsupervised updating (not yet implemented). (default: "category")}

\item{cues}{Names of columns with cue values.}

\item{...}{A named list of functions (statistics) of `cues` to calculate for each unique combination of `group` and
`category`.}
}
\value{
A named list of length `...` (names are the names of the functions that have been computed). The list will be
sorted first by `group` and then by `category`, in ascending order of the levels of those variables.
}
\description{
Get statistics from data. Calculates the means and sum of squares matrices for the specified cues for
any combination of groups (optional) and categories, and returns them as a list.
}
\keyword{TBD}
