% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose-input-for-stan.R
\name{get_sufficient_statistics_as_list_of_arrays}
\alias{get_sufficient_statistics_as_list_of_arrays}
\title{Get sufficient statistics from a data set}
\usage{
get_sufficient_statistics_as_list_of_arrays(
  data,
  cues,
  category,
  group,
  verbose = F,
  ...
)
}
\arguments{
\item{data}{`tibble` or `data.frame` with the data. Each row should be an observation of a category,
and contain information about the category label, the cue values of the observation, and optionally grouping variables.}

\item{cues}{Names of columns with cue values.}

\item{category}{Name of column that contains the category label for the exposure data. Can be `NULL` for unsupervised updating
(not yet implemented). (default: "category")}

\item{group}{Name of column that contains information about which observations form a group. This could be individual
subjects or conditions in an experiment. The latter is more efficient, but should only be used if exposure is
identical for every individual within the group. Test does not have to be identical for every individual within
the same group. For example, one can group multiple groups of subjects that have received the same exposure
but were tested on different test tokens.}

\item{test}{`tibble` or `data.frame` with the test data. Each row should be an observation, and contain information
about the cue values of the test stimulus and the participant's response.}
}
\value{
A list.
}
\description{
Get sufficient statistics from data. Calculates the means and sum of squares matrices for the specified cues for
any combination of groups (optional) and categories, and returns them as a list.
}
\keyword{TBD}
