% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{get_sum_of_squares_from_df}
\alias{get_sum_of_squares_from_df}
\alias{get_sum_of_uncentered_squares_from_df}
\alias{get_sum_of_centered_squares_from_df}
\title{Get sum of uncentered squares}
\usage{
get_sum_of_squares_from_df(data, variables = NULL, center = T, verbose = F)

get_sum_of_uncentered_squares_from_df(data, variables = NULL, verbose = F)

get_sum_of_centered_squares_from_df(data, variables = NULL, verbose = F)
}
\arguments{
\item{data}{A `tibble`, `data.frame`, or `matrix`. If data is a `tibble` or `data.frame`, the columns for
specified variables are extracted and (together) converted into a matrix with as many colums as there are
variables.}

\item{variables}{Only required if data is not already a `matrix`.}
}
\value{
A matrix.
}
\description{
Get sum of uncentered squares. This quantity is a sufficient statistic for, for example, multivariate Gaussian
belief-updating under an Normal-Inverse-Wishart prior.
}
\keyword{TBD}
