% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-stanfit.R
\name{get_test_data_from_NIW_ideal_adaptor_stanfit}
\alias{get_test_data_from_NIW_ideal_adaptor_stanfit}
\title{Get the test data from an NIW ideal adaptor stanfit.}
\usage{
get_test_data_from_NIW_ideal_adaptor_stanfit(fit)
}
\arguments{
\item{x}{\code{\link{NIW_ideal_adaptor_stanfit}} object.}
}
\value{
A \code{tibble} in which each row is a test token. Columns include the cues
and the response counts (one column per category) for all test tokens and all groups.
}
\description{
Returns the test data used during the creation of the \code{\link[rstan]{stanfit}}.
object.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
