% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-2025.R
\name{get_test_data_from_stanfit}
\alias{get_test_data_from_stanfit}
\title{Legacy wrapper for get_test_data.}
\usage{
get_test_data_from_stanfit(...)
}
\description{
Deprecated. Use \code{\link{get_test_data}} instead.

Deprecated. Use get_test_data() instead.
}
\keyword{internal}
