% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R, R/S7-stanfit-methods.R
\name{get_transform_function}
\alias{get_transform_function}
\alias{get_untransform_function}
\alias{get_transform_function,any-method}
\alias{get_untransform_function,any-method}
\title{Get the transform/untransform function from an object.}
\usage{
get_transform_function(x, ...)

get_untransform_function(x, ...)

## S7 method for class <any>
get_transform_function(x, ...)

## S7 method for class <any>
get_untransform_function(x, ...)
}
\arguments{
\item{x}{An S7 fit, fit-input, or transform-information object.}

\item{...}{Additional arguments.}
}
\value{
A function.
}
\description{
Get the transform/untransform function from an object.
}
