% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-2025.R
\name{get_transform_information_from_stanfit}
\alias{get_transform_information_from_stanfit}
\title{Deprecated: get_transform_information_from_stanfit}
\usage{
get_transform_information_from_stanfit(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_transform_information_from_stanfit()} was deprecated in MVBeliefUpdatr 0.1.0
and will be removed in 0.2.0. Please use \code{\link[=get_transform_information]{get_transform_information()}} instead.
}
\seealso{
\code{\link[=get_transform_information]{get_transform_information()}}
}
\keyword{internal}
