% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{h95}
\alias{h95}
\title{Hillenbrand et al. (1995) Vowel Acoustic Data}
\format{
A tibble with 1,668 rows and 9 variables:
\describe{
\item{speaker}{Speaker identifier as a factor.}
\item{sex}{Speaker sex as a factor with levels \code{"female"} and \code{"male"}.}
\item{type}{Speaker demographic group as a factor with levels \code{"men"},
\code{"women"}, \code{"boys"}, and \code{"girls"}.}
\item{vowel}{Standardized Unicode IPA vowel label in phonemic slashes as a factor
with 12 levels: \code{/æ/}, \code{/ɝ/}, \code{/ɑ/}, \code{/eɪ/}, \code{/ɛ/},
\code{/i/}, \code{/ɪ/}, \code{/oʊ/}, \code{/ɔ/}, \code{/u/}, \code{/ʊ/}, \code{/ʌ/}.}
\item{f0}{Fundamental frequency in Hertz (Hz).}
\item{f1}{First formant frequency in Hertz (Hz).}
\item{f2}{Second formant frequency in Hertz (Hz).}
\item{f3}{Third formant frequency in Hertz (Hz).}
\item{duration}{Vowel duration in milliseconds (ms).}
}
}
\source{
Standardized from \code{phonTools::h95} \insertCite{barreda2015}{MVBeliefUpdatr}.
}
\usage{
data(h95)
}
\description{
Acoustic measurements of 12 American English vowels produced by men, women,
boys, and girls in /h-V-d/ syllables. This dataset is a standardized copy of
the dataset bundled in the \pkg{phonTools} package \insertCite{barreda2015}{MVBeliefUpdatr},
originally published by \insertCite{hillenbrand1995;textual}{MVBeliefUpdatr}.
}
\references{
\insertRef{hillenbrand1995}{MVBeliefUpdatr}

\insertRef{barreda2015}{MVBeliefUpdatr}
}
\seealso{
\code{\link{pb52}}, \code{\link{swehvd}}, \code{\link{mixer6}}
}
\keyword{data}
