% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-NIW-IA-stanfit.R
\docType{class}
\name{ideal_adaptor_stanfit-class}
\alias{ideal_adaptor_stanfit-class}
\alias{ideal_adaptor_stanfit}
\alias{NIW_ideal_adaptor_stanfit}
\title{An S7 class for ideal_adaptor stanfit objects that use one of the ideal_adaptor Stan programs.}
\usage{
ideal_adaptor_stanfit(
  data = data.frame(),
  staninput = ideal_adaptor_staninput(),
  stanvars = NULL,
  backend = "rstan",
  save_pars = NULL,
  stan_args = list(),
  stanfit = NULL,
  basis = NULL,
  transform_information = NULL,
  criteria = list(),
  file = NULL,
  version = get_current_versions(),
  labels = list()
)
}
\description{
An S7 class for ideal_adaptor stanfit objects that use one of the ideal_adaptor Stan programs.
}
\details{
See \code{methods(class = "ideal_adaptor_stanfit")} for an overview of available methods.
}
