% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-NIW-IA-stanfit.R
\name{ideal_adaptor_stanfit}
\alias{ideal_adaptor_stanfit}
\title{Constructor for ideal_adaptor_stanfit (S7)}
\usage{
ideal_adaptor_stanfit(
  data = data.frame(),
  staninput = ideal_adaptor_staninput(),
  stanvars = NULL,
  backend = "rstan",
  save_pars = NULL,
  stan_args = list(),
  stanfit = NULL,
  basis = NULL,
  transform_information = NULL,
  criteria = list(),
  file = NULL,
  version = get_current_versions(),
  labels = list()
)
}
\arguments{
\item{data}{data.frame}

\item{staninput}{ideal_adaptor_staninput or list; if list, will be coerced}

\item{stanvars}{backend stanvars or NULL}

\item{backend}{character backend name}

\item{save_pars}{optional}

\item{stan_args}{list of stan args}

\item{stanfit}{rstan stanfit or NULL}

\item{basis}{optional}

\item{transform_information}{transform_information}

\item{criteria}{list}

\item{file}{character path or NULL/"".}

\item{version}{package version info (from get_current_versions())}

\item{labels}{list}
}
\description{
Constructor for ideal_adaptor_stanfit (S7)
}
