% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-NIW-IA-stanfit.R
\docType{class}
\name{ideal_adaptor_staninput-class}
\alias{ideal_adaptor_staninput-class}
\alias{ideal_adaptor_staninput}
\title{S7 class for processed Stan input used by ideal_adaptor_stanfit}
\usage{
ideal_adaptor_staninput(transformed = list(), untransformed = list())
}
\description{
S7 class for processed Stan input used by ideal_adaptor_stanfit
}
