% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit.R
\docType{class}
\name{IdealAdaptorStanfit-class}
\alias{IdealAdaptorStanfit-class}
\alias{IdealAdaptorStanfit}
\title{An S7 class for ideal-adaptor Stan fit objects.}
\usage{
IdealAdaptorStanfit(
  data = data.frame(),
  staninput = NULL,
  stanvars = NULL,
  backend = "rstan",
  save_pars = NULL,
  stan_args = list(),
  stanfit = NULL,
  basis = NULL,
  transform_information = NULL,
  criteria = list(),
  file = NULL,
  version = NULL,
  labels = list()
)
}
\description{
An S7 class for ideal-adaptor Stan fit objects.
}
