% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit-input.R
\docType{class}
\name{IdealAdaptorStanfitInput-class}
\alias{IdealAdaptorStanfitInput-class}
\alias{IdealAdaptorStanfitInput}
\title{An S7 class for ideal-adaptor fit-input objects.}
\usage{
IdealAdaptorStanfitInput(
  data = data.frame(),
  staninput = NULL,
  transform_information = NULL
)
}
\description{
An S7 class for ideal-adaptor fit-input objects.
}
