% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-staninput.R
\docType{class}
\name{IdealAdaptorStaninput-class}
\alias{IdealAdaptorStaninput-class}
\alias{IdealAdaptorStaninput}
\title{S7 class for ideal-adaptor Stan input.}
\usage{
IdealAdaptorStaninput(values = list())
}
\description{
S7 class for ideal-adaptor Stan input.
}
