% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-NIW-IA-stanfit.R
\name{infer_NIW_ideal_adaptor}
\alias{infer_NIW_ideal_adaptor}
\title{Infer NIW ideal adaptor}
\usage{
infer_NIW_ideal_adaptor(
  exposure = NULL,
  test = NULL,
  cues = NULL,
  category = NULL,
  response = NULL,
  group = NULL,
  group.unique = NULL,
  center.observations = TRUE,
  scale.observations = FALSE,
  pca.observations = FALSE,
  pca.cutoff = 1,
  lapse_rate = NULL,
  mu_0 = NULL,
  Sigma_0 = NULL,
  tau_scale = NULL,
  L_omega_eta = 1,
  split_loglik_per_observation = 0,
  staninput = NULL,
  untransform_fit = FALSE,
  sample = TRUE,
  file = NULL,
  file_refit = "never",
  file_compress = T,
  stanvars = NULL,
  backend = "rstan",
  save_pars = NULL,
  basis = NULL,
  init = NULL,
  control = NULL,
  silent = 1,
  stan_model_args = list(),
  stanmodel = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{exposure}{`tibble` or `data.frame` with the exposure data. Each row should be an observation of a category,
and contain information about the category label, the cue values of the observation, and optionally grouping variables.}

\item{test}{`tibble` or `data.frame` with the test data. Each row should be an observation, and contain information
about the cue values of the test stimulus and the participant's response.}

\item{cues}{Names of columns with cue values. Must exist in both exposure and test data.}

\item{category}{Name of column in exposure data that contains the category label. Can be \code{NULL} for unsupervised updating
(not yet implemented). (default: "category")}

\item{response}{Name of column in test data that contains participants' responses. (default: "response")}

\item{group}{Name of column that contains information about which observations form a group. Typically, this is
a variable identifying subjects/participants. Must exist in both exposure and test data. (default: "group")}

\item{group.unique}{Name of column that uniquely identifies each group with identical exposure. This could be a
variable indicating the different conditions in an experiment. Using group.unique is optional, but can be
substantially more efficient if many groups share the same exposure. To ignore, set to \code{NULL}. (default: \code{NULL})}

\item{center.observations}{Should the data be centered based on cues' means during exposure? Note that the cues' means
used for centering are calculated after aggregating the data to all unique combinations specified by \code{group.unique}.
These means are only expected to be the same as the standard deviations over the entire exposure data if the exposure data
are perfectly balanced with regard to \code{group.unique}. Centering will not affect the inferred correlation
or covariance matrices but it will affect the absolute position of the inferred means. The relative position of the inferred
means remains unaffected.
If \code{TRUE} and \code{mu_0} is specified, \code{mu_0} will also be centered (\code{Sigma_0} is not affected by centering and thus not changed).
(default: \code{TRUE})}

\item{scale.observations}{Should the data be standardized based on cues' standard deviation during exposure? Note that the
cues' standard deviations used for scaling are calculated after aggregating the data to all unique combinations specified
by \code{group.unique}. These standard deviations are only expected to be the same as the standard deviations over the entire
exposure data if the exposure data are perfectly balanced with regard to \code{group.unique}.
Scaling will not affect the inferred correlation matrix,
but it will affect the inferred covariance matrix because it affects the inferred standard deviations. It will also
affect the absolute position of the inferred means. The relative position of the inferred means remains unaffected.
If \code{TRUE} and \code{mu_0} and \code{Sigma_0} are specified, \code{mu_0} and \code{Sigma_0} will also be scaled.
(default: `FALSE`)}

\item{pca.observations}{Should the data be transformed into orthogonal principal components? (default: \code{FALSE})}

\item{pca.cutoff}{Determines which principal components are handed to the MVBeliefUpdatr Stan program: all
components necessary to explain at least the pca.cutoff of the total variance. (default: .95) Ignored if
\code{pca.observation = FALSE}. (default: 1)}

\item{lapse_rate, mu_0, Sigma_0}{Optionally, lapse rate, prior expected category means (\code{mu_0}) and/or prior expected
category covariance matrices (\code{Sigma_0}) for all categories. Lapse rate should be a number between 0 and 1. For \code{mu_0}
and \code{Sigma_0}, each should be a list, with each element being the expected mean/covariance matrix for a specific
category prior to updating. Elements of \code{mu_0} and \code{Sigma_0} should be ordered in the same order as the levels of the
category variable in \code{exposure} and \code{test}. These prior expected means and covariance matrices could be
estimated, for example, from phonetically annotated speech recordings (see \code{\link{make_MVG_from_data}}
for a convenient way to do so). Internally, m_0 is then set to \code{mu_0} (so that the expected value of the prior
distribution of means is mu_0) and S_0 is set so that the expected value of the inverse-Wishart is \code{Sigma_0} given nu_0.
Importantly, \strong{Sigma_0 should be convolved with perceptual noise (i.e., add perceptual noise covariance matrix to
the category variability covariance matrices when you specify \code{Sigma_0})} since the stancode for the inference of the
NIW ideal adaptor does \emph{not} infer category and noise variability separately.}

\item{tau_scale}{A vector of scales for the Cauchy priors for each cue's standard deviations. Used in
both the prior for m_0 and the prior for S_0. (default: vector of 5s if scale.observations = TRUE, SD of each cue otherwise).}

\item{L_omega_eta}{A vector of etas of the LKJ prior for the correlations of the covariance matrix of \code{mu_0}. (default: 1,
which corresponds to a uniform prior of correlation matrices)}

\item{split_loglik_per_observation}{Optionally, split the log likelihood per observation. This can be helpful of leave-one-out
estimation in order to avoid high Pareto k, but it also makes the stored stanfit object much larger. (default: 0)}

\item{staninput}{A list of the type that would be returned by \code{\link{make_staninput}}.
This list can be provided *instead* of the arguments required by \code{make_staninput}.}

\item{untransform_fit}{Logical flag indicating whether the samples of the model should be transformed back
into the original cue space by applying the untransform function. (default: \code{FALSE})}

\item{sample}{Should the model be fit and sampled from?}

\item{file}{Either NULL or a character string. In the latter case, the fitted model object is saved
via saveRDS in a file named after the string supplied in file. The .rds extension is added automatically.
If the file already exists, the model from that file will be loaded and returned instead of refitting the model.
As existing files won't be overwritten, you have to manually remove the file in order to refit and save the
model under an existing file name. The file name is stored in the \code{NIW_ideal_adaptor_stanfit} object
for later use. (default: \code{NULL})}

\item{file_refit}{Modifies when the fit stored via the \code{file} argument
is re-used. For \code{"never"} (default) the fit is always loaded if it
exists and fitting is skipped. For \code{"always"} the model is always
refitted. If set to \code{"on_change"}, model will be refit if data passed to
Stan differ from
what is stored in the file. Refit will not be triggered for changes in
additional parameters of the fit (e.g., initial values, number of iterations,
control arguments, ...).}

\item{file_compress}{Logical or a character string, specifying one of the
compression algorithms supported by \code{\link{saveRDS}}. If the
\code{file} argument is provided, this compression will be used when saving
the fitted model object.}

\item{backend}{Character string naming the package to use as the backend for
fitting the Stan model. Options are \code{"rstan"} (the default) or
\code{"cmdstanr"}. Details on the
\pkg{rstan} and \pkg{cmdstanr} packages are available at
\url{https://mc-stan.org/rstan/} and \url{https://mc-stan.org/cmdstanr/},
respectively.}

\item{silent}{Verbosity level between \code{0} and \code{2}.
If \code{1} (the default), most of the
informational messages of compiler and sampler are suppressed.
If \code{2}, even more messages are suppressed. The actual
sampling progress is still printed. Set \code{refresh = 0} to turn this off
as well. If using \code{backend = "rstan"} you can also set
\code{open_progress = FALSE} to prevent opening additional progress bars.}

\item{stan_model_args}{A \code{list} of further arguments passed to
\code{\link[rstan:stan_model]{rstan::stan_model}} for \code{backend =
"rstan"} or to \code{cmdstanr::cmdstan_model} for \code{backend =
"cmdstanr"}, which allows to change how models are compiled.}

\item{stanmodel}{Name of stanmodel that should be used. Overrides any default selection.}

\item{...}{Additional parameters are passed to \code{\link[rstan]{sampling}}}
}
\value{
An object of class \code{NIW_ideal_adaptor_stanfit} with the fitted stan model. In interpreting the
  inferred parameters, it should be kept in mind that the \emph{inferred} scatter matrix S_0 includes
  variability from internal perceptual and/or external environmental noise, \emph{in addition} to the motor
  noise that is reflected in production data. This also implies that, \strong{if \code{Sigma_0} is provided
  by the user it should be convolved with perceptual noise}. This is particularly important if the data you're
  fitting contains test phases without exposure (e.g., pre-exposure tests). Make sure to read the notes about
  the \code{Sigma_0} argument in the help page on \code{\link{make_staninput}}. Use
  \code{methods(class = "NIW_ideal_adaptor_stanfit")} for an overview on available methods.
}
\description{
Infers a posterior distribution of NIW_ideal _adaptors from the input data using rstan/stan. The function can take
two types of inputs: an staninput list, as prepared by \code{\link{make_staninput}},
or the exposure and test data, the names of the cues, category, and response columns (and optionally group and/or block columns).
}
\seealso{
\code{\link{is.NIW_ideal_adaptor_stanfit}} for information about NIW_ideal_adaptor_stanfit objects,
\code{\link{get_draws}} to draw samples from the stanfit.
}
