% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-visualize-basics.R
\name{inv_symlog}
\alias{inv_symlog}
\title{Deprecated: inv_symlog}
\usage{
inv_symlog(x, C = 0)
}
\arguments{
\item{x}{Numeric vector to invert.}

\item{C}{Scaling constant. Defaults to 0.}
}
\value{
Inverse transformed numeric vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{inv_symlog()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0. Please use \code{\link[scales:pseudo_log_trans]{scales::pseudo_log_trans()}} instead.
}
\seealso{
\code{\link[scales:pseudo_log_trans]{scales::pseudo_log_trans()}}
}
\keyword{internal}
