% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-class-exemplar-model.R
\name{is.exemplar_model}
\alias{is.exemplar_model}
\title{Deprecated: is.exemplar_model}
\usage{
is.exemplar_model(x, group = NULL, verbose = F, tolerance = MVBU_PROB_TOL)
}
\arguments{
\item{x}{Object to be checked.}

\item{group}{Name of one or more group variables, each unique combination of which describes an exemplar model. (default: NULL)}

\item{verbose}{Should verbose output be provided? (default: \code{TRUE})}

\item{tolerance}{Probability tolerance.}
}
\value{
A logical.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{is.exemplar_model()} was deprecated in MVBeliefUpdatr 0.1.0 and will be removed in 0.2.0.
Please use S7 validators and \link{Exemplar_Model} instead.
}
\seealso{
\link{Exemplar_Model}
}
\keyword{internal}
