% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-class-exemplars.R
\name{is.exemplars}
\alias{is.exemplars}
\title{Deprecated: is.exemplars}
\usage{
is.exemplars(x, group = NULL, verbose = F)
}
\arguments{
\item{x}{Object to be checked.}

\item{group}{Name of one or more group variables, each unique combination of which describes a set of exemplars. (default: NULL)}

\item{verbose}{Logical. If \code{TRUE}, emits diagnostics.}
}
\value{
A logical.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{is.exemplars()} was deprecated in MVBeliefUpdatr 0.1.0 and will be removed in 0.2.0.
Please use S7 validators and \link{Exemplar_Model} instead.
}
\seealso{
\link{Exemplar_Model}
}
\keyword{internal}
