% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit.R
\name{is.ideal_adaptor_stanfit}
\alias{is.ideal_adaptor_stanfit}
\title{Is this an NIW ideal adaptor stanfit?}
\usage{
is.ideal_adaptor_stanfit(x, verbose = FALSE)
}
\arguments{
\item{x}{Object to be checked.}

\item{verbose}{Currently being ignored.}
}
\value{
A logical.
}
\description{
Check whether \code{x} is of class \code{\link{ideal_adaptor_stanfit}}.
}
