% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-NIW-IA-stanfit.R
\name{is.ideal_adaptor_stanfit_input}
\alias{is.ideal_adaptor_stanfit_input}
\title{Is this a list of NIW ideal adaptor stanfit inputs?}
\usage{
is.ideal_adaptor_stanfit_input(x, verbose = FALSE)
}
\description{
Is this a list of NIW ideal adaptor stanfit inputs?
}
