% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-NIW-IA-stanfit.R
\name{is.ideal_adaptor_staninput}
\alias{is.ideal_adaptor_staninput}
\title{Is this a list of NIW ideal adaptor stanfit inputs?}
\usage{
is.ideal_adaptor_staninput(x)
}
\value{
A logical.
}
\description{
Check whether \code{x} is of class \code{\link{ideal_adaptor_stanfit}}.
}
\seealso{
TBD
}
\keyword{TBD}
