% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{is.model}
\alias{is.model}
\title{Is this a model?}
\usage{
is.model(x, verbose = F, tolerance = 1e-05)
}
\arguments{
\item{x}{Object to be checked.}

\item{verbose}{Should verbose output be provided? (default: `TRUE`)}
}
\value{
A logical.
}
\description{
Check whether \code{x} is a model with lapse rates, biases, priors, and perceptual noise.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
