% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{is.mv_ibbu_input}
\alias{is.mv_ibbu_input}
\title{Is this a list of MV IBBU inputs?}
\usage{
is.mv_ibbu_input(x)
}
\value{
A logical.
}
\description{
Check whether \code{x} is of class \code{mv_ibbu_input}.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
