% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{is.mv_ibbu_MCMC}
\alias{is.mv_ibbu_MCMC}
\title{Is this a tibble of MV IBBU draws?}
\usage{
is.mv_ibbu_MCMC(x, is.nested = T, is.long = T)
}
\value{
A logical.
}
\description{
Check whether \code{x} is a tibble of post-warmup draws of parameters obtained from incremental
Bayesian belief-updating (IBBU).
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
