% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-class-MVG.R
\name{is.MVG}
\alias{is.MVG}
\title{Deprecated: is.MVG}
\usage{
is.MVG(x, group = NULL, category = "category", is.long = T, verbose = F)
}
\arguments{
\item{x}{Object to be checked.}

\item{group}{Name of one or more group variables, each unique combination of which describes an MVG. (default: NULL)}

\item{category}{DEPRECATED Name of the category variable. (default: "category")}

\item{is.long}{Is this check assessing whether the ideal observer is in long format (\code{TRUE}) or wide format (\code{FALSE})?
(default: \code{TRUE})}

\item{verbose}{Logical. If \code{TRUE}, emits diagnostics.}
}
\value{
A logical.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{is.MVG()} was deprecated in MVBeliefUpdatr 0.1.0 and will be removed in 0.2.0.
Please use S7 validators and \link{MVG_IdealObserver} instead.
}
\seealso{
\link{MVG_IdealObserver}
}
\keyword{internal}
