% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-class-MVG-IO.R
\name{is.MVG_ideal_observer}
\alias{is.MVG_ideal_observer}
\title{Deprecated: is.MVG_ideal_observer}
\usage{
is.MVG_ideal_observer(
  x,
  group = NULL,
  category = "category",
  is.long = T,
  with.lapse = if (with.lapse_bias) T else F,
  with.lapse_bias = F,
  verbose = F,
  tolerance = MVBU_PROB_TOL
)
}
\arguments{
\item{x}{Object to be checked.}

\item{group}{Name of one or more group variables, each unique combination of which describes an MVG_ideal_observer. (default: NULL)}

\item{category}{Name of the category variable. (default: "category")}

\item{is.long}{Is this check assessing whether the ideal observer is in long format (\code{TRUE}) or wide format (\code{FALSE})?
(default: \code{TRUE})}

\item{with.lapse}{Does this ideal observer have a lapse rate? (default: \code{FALSE})}

\item{with.lapse_bias}{Does this ideal observer have a lapse bias? (default: \code{FALSE})}

\item{verbose}{Should verbose output be provided? (default: \code{TRUE})}
}
\value{
A logical.
}
\description{
Deprecated. Use the S7-based validators and constructors for ideal observers with MVG categories instead.
}
\details{
Check whether \code{x} is an ideal observer with \link[=is.MVG]{multivariate Gaussian (MVG) categories}. Optionally, one can also check whether a lapse rate
and lapse bias is part of the ideal observer.
}
\seealso{
TBD
}
\keyword{internal}
