% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mvg-ibbu-stanfit.R
\name{is.NIW_ibbu_input}
\alias{is.NIW_ibbu_input}
\title{Is this a list of NIW IBBU inputs?}
\usage{
is.NIW_ibbu_input(x)
}
\value{
A logical.
}
\description{
Check whether \code{x} is of class \code{NIW_ibbu_input}.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
