% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mvg-ibbu-stanfit.R
\name{is.NIW_ibbu_stanfit}
\alias{is.NIW_ibbu_stanfit}
\title{Is this an NIW IBBU stanfit?}
\usage{
is.NIW_ibbu_stanfit(x)
}
\value{
A logical.
}
\description{
Check whether \code{x} is of class \code{NIW_ibbu_stanfit}.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
