% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-NIW-IA-stanfit.R
\name{is.NIW_ideal_adaptor_MCMC}
\alias{is.NIW_ideal_adaptor_MCMC}
\title{Is this a tibble of MCMC draws of an NIW ideal adaptor?}
\usage{
is.NIW_ideal_adaptor_MCMC(
  x,
  is.nested = T,
  is.long = T,
  with.prior = F,
  with.lapse = if (with.lapse_bias) T else F,
  with.lapse_bias = F
)
}
\value{
A logical.
}
\description{
Check whether \code{x} is a tibble of post-warmup draws of parameters obtained from incremental
conjugate Bayesian belief-updating (IBBU) over a Normal-Inverse-Wishart (NIW) prior.
}
\seealso{
TBD
}
\keyword{TBD}
