% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-NIW-belief.R
\name{is.NIW_belief}
\alias{is.NIW_belief}
\alias{is.NIW_belief_w_lapse}
\alias{is.NIW_belief_w_bias}
\alias{is.NIW_beliefs}
\alias{is.NIW_beliefs_w_lapse}
\alias{is.NIW_beliefs_w_bias}
\title{Is this an NIW belief object?}
\usage{
is.NIW_belief(
  x,
  category = "category",
  is.long = T,
  with.lapse = if (with.bias) T else F,
  with.bias = F
)

is.NIW_belief_w_lapse(x, category = "category", is.long = T, with.bias = F)

is.NIW_belief_w_bias(x, category = "category", is.long = T)

is.NIW_beliefs(
  x,
  category = "category",
  is.long = T,
  with.lapse = if (with.bias) T else F,
  with.bias = F
)

is.NIW_beliefs_w_lapse(x, category = "category", is.long = T, with.bias = F)

is.NIW_beliefs_w_bias(x, category = "category", is.long = T)
}
\arguments{
\item{x}{Object to be checked.}

\item{category}{Name of the category variable. (default: "category")}

\item{is.long}{Is this check assessing whether the belief is in long format (`TRUE`) or wide format (`FALSE`)?
(default: `TRUE`)}

\item{with.lapse}{Does this belief have a lapse rate? (default: `FALSE`)}

\item{with.bias}{Does this belief have a bias? (default: `FALSE`)}
}
\value{
A logical.
}
\description{
Check whether \code{x} is an NIW belief/set of NIW beliefs. Optionally, one can also check whether a lapse rate
and bias is part of the belief.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
