% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-objects.R
\name{lift_MVG_ideal_observer_to_NIW_ideal_adaptor}
\alias{lift_MVG_ideal_observer_to_NIW_ideal_adaptor}
\title{Turn an MVG_ideal_observer into an NIW_ideal_adaptor}
\usage{
lift_MVG_ideal_observer_to_NIW_ideal_adaptor(
  x,
  group = NULL,
  category = "category",
  kappa,
  nu,
  verbose = F
)
}
\arguments{
\item{x}{An MVG_ideal_observer object.}

\item{group}{Optionally, a grouping structure can be specified. If group structure is not NULL, one
NIW belief or ideal adaptor will be derived for each level of \code{group_structure}. (default: NULL)}

\item{category}{Name of variable in \code{data} that contains the category information. (default: "category")}

\item{kappa}{The strength of the beliefs over the category mean (pseudocounts). (default: same as nu)}

\item{nu}{The strength of the beliefs over the category covariance matrix (pseudocounts). (default: number of
cues + 2)}

\item{verbose}{If true provides more information. (default: FALSE)}
}
\value{
A tibble that is an NIW_ideal_adaptor object.
}
\description{
Make an ideal adaptor out of an ideal observer, so that the *expected* category mean and category covariance
matrix of the ideal adaptor match the ideal observers category mean and category covariance matrix.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
