% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-NIW-IA-stanfit.R
\name{loo.ideal_adaptor_stanfit}
\alias{loo.ideal_adaptor_stanfit}
\title{loo for NIW ideal adaptor stanfit}
\usage{
\method{loo}{ideal_adaptor_stanfit}(
  x,
  pars = "log_lik",
  ...,
  save_psis = FALSE,
  cores = getOption("mc.cores", 1)
)
}
\arguments{
\item{x}{An \code{\link{ideal_adaptor_stanfit}} object.}
}
\value{
A \code{loo} object.
}
\description{
\code{loo} method for \code{\link{ideal_adaptor_stanfit}} objects.
}
