% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit-methods.R
\name{loo.MVBU_Stanfit}
\alias{loo.MVBU_Stanfit}
\title{loo for MVBU stanfit objects}
\usage{
\method{loo}{MVBU_Stanfit}(
  x,
  pars = "log_lik",
  ...,
  save_psis = FALSE,
  cores = getOption("mc.cores", 1)
)
}
\arguments{
\item{x}{An \code{\link{MVBU_Stanfit}} object.}

\item{pars}{Parameter name containing the pointwise log likelihood. (default: \code{"log_lik"})}

\item{...}{Additional arguments passed to \code{loo::loo.array}.}

\item{save_psis}{Logical scalar passed to \code{loo::loo.array}. (default: \code{FALSE})}

\item{cores}{Number of cores for parallel execution. (default: \code{getOption("mc.cores", 1)})}
}
\value{
A \code{loo} object.
}
\description{
\code{loo} method for \code{\link{MVBU_Stanfit}} objects. For details,
see \link[loo:loo]{loo::loo}.
}
