% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-NIW-IA.R
\name{loo.NIW_ideal_adaptor_stanfit}
\alias{loo.NIW_ideal_adaptor_stanfit}
\title{loo NIW ideal adaptor stanfit}
\usage{
\method{loo}{NIW_ideal_adaptor_stanfit}(
  x,
  pars = "log_lik",
  ...,
  save_psis = FALSE,
  cores = getOption("mc.cores", 1)
)
}
\arguments{
\item{x}{An \code{\link{NIW_ideal_adaptor_stanfit}} object.}
}
\description{
\code{loo} method for \code{\link{NIW_ideal_adaptor_stanfit}} objects.
}
