% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{make_MV_exposure_data}
\alias{make_MV_exposure_data}
\title{Make multivariate Gaussian exposure data.}
\usage{
make_MV_exposure_data(
  Ns,
  means,
  Sigmas,
  category.labels = NULL,
  cue.labels = NULL,
  keep.parameters = F
)
}
\arguments{
\item{Ns}{A list of integers, with each number specifying the number of observations to be drawn from the corresponding
Gaussian.}

\item{means}{List of mean vectors, each specifying the mean of a multivariate Gaussian.}

\item{category.labels}{List of category names, each specifying the category label of a multivariate Gaussian. If \code{NULL}
(default) then Gaussians will be numbered from 1:N.}

\item{cue.labels}{List of cue names. If \code{NULL} (default) then the cues will be numbered cue1, cue2, ...}

\item{keep.parameters}{Should the parameters handed to this function be included in the output? (default: FALSE)}

\item{sigmas}{List of covariance matrices, each specifying the covariance of a multivariate Gaussian.}
}
\value{
A tibble.
}
\description{
Returns a tibble of observations drawn from multivariate Gaussians, with one observation per row. Each row
provides the category label and cue values. If \code{keep.parameters = T} then the parameters (\code{N, mean, sigma})
are also returned.
}
\details{
The input is expected to be lists of parameters which the n-th element of each list specifying the n-th Gaussian and the
number of observations to be drawn from that Gaussian.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
