% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{make_NIW_prior_from_data}
\alias{make_NIW_prior_from_data}
\title{Make NIW prior from data.}
\usage{
make_NIW_prior_from_data(
  data,
  group = NULL,
  category = "category",
  cues,
  kappa = NA,
  nu = NA,
  lapse = NA
)
}
\arguments{
\item{data}{The tibble or data.frame from which to construct the prior.}

\item{category}{Name of variable in \code{data} that contains the category information. (default: "category")}

\item{cues}{Name(s) of variables in \code{data} that contain the cue information.}

\item{kappa}{The strength of the beliefs over the category mean (pseudocounts).}

\item{nu}{The strength of the beliefs over the category covariance matrix (pseudocounts).}

\item{lapse}{Optionally specify a lapse rate. (default: \code{NA})}
}
\value{
A tibble that is an NIW_belief object.
}
\description{
Constructs an NIW_belief object, representing Normal-Inverse Wishart (NIW) parameters for all categories found in
the data. This object can be used as a prior for functions like \code{\link{update_NIW_beliefs()}}.
}
\details{
Currently, \code{make_NIW_prior_from_data()} does not infer kappa or nu, nor does it fit hierarchical data. Rather
the function simply estimates the category mean and covariance matrix from the sample (\code{data}), assumes them
to be the expected category mean (mu) and covariance (Sigma), and derives the M and S parameters
of the NIW from mu and Sigma based on the user-provided kappa and nu. That means M = mu and S = Sigma * (nu - D -1),
where D is the dimensionality of the data.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
