% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-basics.R
\name{make_vector_column}
\alias{make_vector_column}
\title{Deprecated: Combine columns into a vector column}
\usage{
make_vector_column(data, cols, vector_col, .keep = "all")
}
\arguments{
\item{data}{A tibble or data.frame.}

\item{cols}{A character vector of variable names to combine.}

\item{vector_col}{Name of the new vector-valued column.}

\item{.keep}{A tidyselect option passed to dplyr::mutate.}
}
\value{
Same as \code{data}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{make_vector_column()} is deprecated as of MVBeliefUpdatr 0.1.0 and will be removed in 0.2.0.
}
\keyword{internal}
