% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-staninput.R
\docType{class}
\name{MVBU-Staninput-class}
\alias{MVBU-Staninput-class}
\alias{MVBU_Staninput}
\title{S7 base class for processed Stan input}
\usage{
MVBU_Staninput(values = list())
}
\description{
S7 base class for processed Stan input
}
\section{Slots}{

\describe{
\item{\code{values}}{A named list containing the Stan input values.}
}}

