% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-transform-information.R
\docType{class}
\name{MVBU-TransformInformation-class}
\alias{MVBU-TransformInformation-class}
\alias{MVBU_TransformInformation}
\title{S7 base class for transformation information}
\usage{
MVBU_TransformInformation(
  transform.parameters = list(),
  transform.function = function(x) x,
  untransform.function = function(x) x
)
}
\arguments{
\item{transform.parameters}{A list of sufficient parameters for the
transform. Default: \code{list()}.}

\item{transform.function}{Function to transform parameters. Default:
identity.}

\item{untransform.function}{Function to untransform parameters. Default:
identity.}
}
\description{
S7 base class for transformation information
}
