% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit-posterior.R
\docType{class}
\name{MVBU_StanfitPosterior}
\alias{MVBU_StanfitPosterior}
\title{S7 class for Stanfit posterior parameter draws}
\usage{
MVBU_StanfitPosterior(draws = list(), metadata = list())
}
\arguments{
\item{draws}{A list containing \code{raw} and \code{summary} posterior
draws. Default: \code{list()}.}

\item{metadata}{A list containing auxiliary information including
\code{label_information} (cues, categories, group), \code{model_family},
\code{original_variable_names}, \code{sufficient_statistics}, and
attached data. Default: \code{list()}.}
}
\description{
Represents a posterior distribution over a cognitive model, as obtained
for example from an \link{MVBU_Stanfit} object. Holds MCMC parameter draw
distributions without carrying sampling chains, diagnostics, or raw model
source structures, and is thus more lightweight than the \link{MVBU_Stanfit}
class.
}
