% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-staninput.R
\docType{class}
\name{MVBU_Staninput}
\alias{MVBU_Staninput}
\alias{IdealAdaptorStaninput}
\title{MVBeliefUpdatr Stan Input Classes and Constructors}
\usage{
MVBU_Staninput(values = list())

IdealAdaptorStaninput(values = list())
}
\arguments{
\item{values}{A named list containing the Stan input values.
Default: \code{list()}.}
}
\description{
MVBeliefUpdatr Stan Input Classes and Constructors
}
