% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-classes.R, R/s7-foundation.R
\name{MVG-IdealObserver-class}
\alias{MVG-IdealObserver-class}
\alias{MVG_IdealObserver}
\title{MVG Ideal Observer Model Class}
\usage{
MVG_IdealObserver(
  category_likelihood = MVBU_CategoryRepresentationTemplate(),
  decision_rule = character(0),
  category_prior = integer(0),
  lapse_rate = integer(0),
  lapse_bias = integer(0),
  metadata = list()
)

MVG_IdealObserver(
  category_likelihood = MVBU_CategoryRepresentationTemplate(),
  decision_rule = character(0),
  category_prior = integer(0),
  lapse_rate = integer(0),
  lapse_bias = integer(0),
  metadata = list()
)
}
\description{
Cognitive model class for multivariate Gaussian observers.
Inherits all base cognitive-model machinery and is paired with
\link{MVG_CategoryRepresentation} templates.

Cognitive model class for multivariate Gaussian observers.
Inherits all base cognitive-model machinery and is paired with
\link{MVG_CategoryRepresentation} templates.
}
\seealso{
\link{NIW_IdealAdaptor}, \link{MVG_IdealObserverDistribution}
}
\keyword{internal}
