% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mvg-ibbu-stanfit.R
\docType{class}
\name{mvg_ibbu_stanfit-class}
\alias{mvg_ibbu_stanfit-class}
\alias{mvg_ibbu_stanfit}
\title{An S4 class for stanfit objects that use one of the mvg_ibbu stan programs.}
\description{
An S4 class for stanfit objects that use one of the mvg_ibbu stan programs.
}
\section{Slots}{

\describe{
\item{\code{input_data}}{A list}

\item{\code{labels}}{A list}
}}

