% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-class.R
\name{new_category_representation}
\alias{new_category_representation}
\title{Construct a base category representation object}
\usage{
new_category_representation(
  category_labels,
  cue_labels,
  category_likelihood_function = NULL,
  metadata = list()
)
}
\arguments{
\item{category_labels}{Character vector of category label(s).}

\item{cue_labels}{Character vector of cue label(s).}

\item{category_likelihood_function}{Optional category likelihood function.
Default: \code{NULL}.}

\item{metadata}{Optional list of metadata. Default: \code{list()}.}
}
\value{
An \link{MVBU_CategoryRepresentation} object.
}
\description{
Construct a base category representation object
}
\keyword{internal}
