% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-class.R
\name{new_category_representation_template}
\alias{new_category_representation_template}
\title{Construct an MVBU Category Representation Template}
\usage{
new_category_representation_template(representations, metadata = list())
}
\arguments{
\item{representations}{A named or unnamed list of
\code{MVBU_CategoryRepresentation} objects.}

\item{metadata}{Optional metadata list for template-level properties.
Defaults to \code{list()}.}
}
\value{
An object of class \code{MVBU_CategoryRepresentationTemplate}.
}
\description{
Construct an MVBU Category Representation Template
}
