% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-classes.R
\name{new_exemplar_model}
\alias{new_exemplar_model}
\title{Construct an Exemplar model object}
\usage{
new_exemplar_model(
  category_template = NULL,
  decision_rule = "sampling",
  category_prior = NULL,
  lapse_rate = 0,
  lapse_bias = NULL,
  Sigma_noise = NULL,
  noise_treatment = "no_noise",
  lapse_treatment = "no_lapses",
  metadata = list()
)
}
\description{
Construct an Exemplar model object
}
\keyword{internal}
