% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-classes.R, R/s7-foundation.R
\name{new_exemplar_model_distribution}
\alias{new_exemplar_model_distribution}
\title{Construct an exemplar model distribution object}
\usage{
new_exemplar_model_distribution(
  cache = list(),
  metadata = list(),
  group_label = ""
)

new_exemplar_model_distribution(
  cache = list(),
  metadata = list(),
  group_label = ""
)
}
\description{
Construct an exemplar model distribution object

Construct an exemplar model distribution object
}
\keyword{internal}
