% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-core-classes.R
\name{new_model_distribution}
\alias{new_model_distribution}
\title{Construct a general model distribution object.}
\usage{
new_model_distribution(
  model_family,
  cache = list(),
  metadata = list(),
  group_label = ""
)
}
\arguments{
\item{model_family}{Model-family name.}

\item{cache}{Optional cache list.}

\item{metadata}{Optional metadata list.}

\item{group_label}{Optional group label.}
}
\value{
An S7 model-distribution object.
}
\description{
Construct a general model distribution object.
}
