% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-core-classes.R
\name{new_muvg_model_distribution}
\alias{new_muvg_model_distribution}
\title{Construct a MUVG model distribution object.}
\usage{
new_muvg_model_distribution(
  cache = list(),
  metadata = list(),
  group_label = ""
)
}
\arguments{
\item{cache}{Optional cache list.}

\item{metadata}{Optional metadata list.}

\item{group_label}{Optional group label.}
}
\value{
A MUVG S7 model-distribution object.
}
\description{
Construct a MUVG model distribution object.
}
