% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mvg-ibbu-stanfit.R
\docType{class}
\name{NIW_ibbu_stanfit-class}
\alias{NIW_ibbu_stanfit-class}
\alias{NIW_ibbu_stanfit}
\title{An S4 class for stanfit objects that use one of the NIW_ibbu stan programs.}
\description{
An S4 class for stanfit objects that use one of the NIW_ibbu stan programs.
}
\section{Slots}{

\describe{
\item{\code{input_data}}{list containing the data handed to rstan through \code{compose_data} function.}

\item{\code{transform_functions}}{list containing elements transform.function and untransform.function.}

\item{\code{labels}}{list}
}}

