% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-NIW-IA-stanfit.R
\docType{class}
\name{NIW_ideal_adaptor_stanfit-class}
\alias{NIW_ideal_adaptor_stanfit-class}
\alias{NIW_ideal_adaptor_stanfit}
\title{An S4 class for stanfit objects that use one of the NIW_ideal_adaptor stan programs.}
\description{
An S4 class for stanfit objects that use one of the NIW_ideal_adaptor stan programs.
}
\section{Slots}{

\describe{
\item{\code{input_data}}{list containing the data handed to rstan through \code{compose_data} function.}

\item{\code{transform_information}}{list containing elements transform.parameters, transform.function, and
untransform.function.}

\item{\code{labels}}{list}
}}

