% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-plot-expected-categories.R
\name{plot_expected_categories.ideal_adaptor_stanfit}
\alias{plot_expected_categories.ideal_adaptor_stanfit}
\title{Deprecated: plot_expected_categories.ideal_adaptor_stanfit}
\usage{
plot_expected_categories.ideal_adaptor_stanfit(
  model,
  type = "density",
  cues = get_cue_labels(model),
  ...
)
}
\arguments{
\item{model}{Model object.}

\item{type}{Plot type ("density" or "contour").}

\item{cues}{Cue names to plot.}

\item{...}{Arguments passed to \code{\link[=plot_categories]{plot_categories()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{plot_expected_categories.ideal_adaptor_stanfit()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0. Please use \code{\link[=plot_categories]{plot_categories()}} instead.
}
\seealso{
\code{\link[=plot_categories]{plot_categories()}}
}
\keyword{internal}
