% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-NIW-IA.R
\name{plot_expected_categories_density_1D}
\alias{plot_expected_categories_density_1D}
\alias{plot_expected_categories_density1D}
\title{Plot expected univariate (1D) category likelihoods}
\usage{
plot_expected_categories_density_1D(
  x,
  data.exposure = NULL,
  data.test = NULL,
  facet_rows_by = NULL,
  facet_cols_by = NULL,
  facet_wrap_by = NULL,
  animate_by = NULL,
  animation_follow = F,
  xlim,
  ylim = NULL,
  x.expand = c(0, 0),
  category.ids = NULL,
  category.labels = NULL,
  category.colors = NULL,
  category.linetypes = NULL,
  ...
)

plot_expected_categories_density1D(
  x,
  data.exposure = NULL,
  data.test = NULL,
  facet_rows_by = NULL,
  facet_cols_by = NULL,
  facet_wrap_by = NULL,
  animate_by = NULL,
  animation_follow = F,
  xlim,
  ylim = NULL,
  x.expand = c(0, 0),
  category.ids = NULL,
  category.labels = NULL,
  category.colors = NULL,
  category.linetypes = NULL,
  ...
)
}
\arguments{
\item{x}{An \code{\link{NIW_belief}} or \code{\link{NIW_beliefs}} object.}

\item{data.exposure}{Optional \code{tibble} or \code{data.frame} that contains exposure data to be plotted. (default: `NULL`)}

\item{data.test}{Optional \code{tibble} or \code{data.frame} that contains test data to be plotted. (default: `NULL`)}

\item{facet_rows_by, facet_cols_by, facet_wrap_by, animate_by}{Which group variables, if any, should be used for faceting and/or
animation? (defaults: `NULL`)}

\item{animation_follow}{Should the animation follow the data (zoom in and out)? (default: `FALSE`)}

\item{xlim, ylim}{Limits for the x- and y-axis.}

\item{category.ids}{Vector of category IDs to be plotted or leave `NULL` to plot all groups. (default: `NULL`)}

\item{category.labels}{Vector of group labels of same length as `category.ids` or `NULL` to use defaults. (default: `NULL`)}

\item{category.colors}{Vector of colors of same length as category.ids or `NULL` to use defaults. (default: `NULL`)}

\item{category.linetypes}{Vector of linetypes of same length as category.ids or `NULL` to use defaults. (default: `NULL`)
Currently being ignored.}

\item{...}{additional arguments to geom_line.}

\item{levels}{Levels of the confidence ellipses. (default: .5, .66, .8, .9., and .95)}
}
\value{
ggplot object.
}
\description{
Plot univariate Gaussian categories expected given NIW belief(s). One NIW belief describes the uncertainty about the
category statistics of all categories. This includes the m (the mean of category means \eqn{\mu}), S (the scattermatrix),
kappa (the strength of the belief in m) and nu (the strength of the belief in S). For the univariate case, m and S are
scalars \insertCite{@see @murphy2012 p. 136}{MVBeliefUpdatr}.
}
\details{
It is possible to hand more than one NIW belief to this function, and to facet or animate by variables that uniquely
identify the different beliefs. For example, one can plot
different priors for different talkers (grouping by talker), or different posteriors for different exposure conditions
(grouping by exposure condition), the incremental updating of NIW beliefs (grouping by observations), or any combinations
of these.
}
\references{
\insertRef{murphy2012}{MVBeliefUpdatr}
}
\seealso{
TBD
}
\keyword{TBD}
