% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-forward-ibbu.R
\name{plot_expected_categories_contour2D}
\alias{plot_expected_categories_contour2D}
\title{Plot expected bivariate (2D) categories.}
\usage{
plot_expected_categories_contour2D(
  x,
  grouping.var = NULL,
  panel.group = F,
  animate.group = F,
  levels = c(1/2, 2/3, 4/5, 9/10, 19/20),
  data.exposure = NULL,
  data.test = NULL,
  plot.exposure = F,
  plot.test = F,
  category.ids = NULL,
  category.labels = NULL,
  category.colors = NULL,
  category.linetypes = NULL
)
}
\arguments{
\item{x}{NIW belief object.}

\item{grouping.var}{Grouping variable that is used to create separate instances of the categories. These instances
can be plotted in separate panels (if panel.group is `TRUE`) or used to create animates (if animate.group is `TRUE`).}

\item{panel.group, animate.group}{Determines whether the grouping variable is for paneling or animation. (both defaults: `FALSE`)}

\item{levels}{Levels of the confidence ellipses. (default: .5, .66, .8, .9., and .95)}

\item{category.ids}{Vector of category IDs to be plotted or leave `NULL` to plot all groups. (default: `NULL`) It is possible
to use \code{\link[tidybayes]{recover_types}} on the stanfit object prior to handing it to this plotting function.}

\item{category.labels}{Vector of group labels of same length as `category.ids` or `NULL` to use defaults. (default: `NULL`)}

\item{category.colors}{Vector of colors of same length as category.ids or `NULL` to use defaults. (default: `NULL`)}

\item{category.linetypes}{Vector of linetypes of same length as category.ids or `NULL` to use defaults. (default: `NULL`)
Currently being ignored.}
}
\value{
ggplot object.
}
\description{
Plot bivariate Gaussian categories expected given NIW belief(s). One NIW belief describes the uncertainty about the
category statistics of all categories. This includes the M (the mean of category means \eqn{\mu}), S (the scattermatrix),
kappa (the strength of the belief in M) and nu (the strength of the belief in S).
}
\details{
It is possible to hand more than one NIW belief to this function, one for each level of a grouping variable. This allows
to plot different sets of beliefs, for example, across different panels or as an animation. For example, one can plot
different priors for different talkers (grouping by talker), or different posteriors for different exposure conditions
(grouping by exposure condition), or the incremental updating of NIW beliefs (grouping by observations).
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
