% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-stanfit-ibbu.R
\name{plot_expected_ibbu_stanfit_categories_2D}
\alias{plot_expected_ibbu_stanfit_categories_2D}
\alias{plot_expected_ibbu_stanfit_categories_contour2D}
\alias{plot_expected_ibbu_stanfit_categories_density2D}
\title{Plot expected bivariate (2D) categories from MV IBBU stanfit.}
\usage{
plot_expected_ibbu_stanfit_categories_2D(model, type, ...)

plot_expected_ibbu_stanfit_categories_contour2D(
  model,
  categories = get_category_levels_from_stanfit(model),
  groups = get_group_levels_from_stanfit(model, include_prior = T),
  plot.test = T,
  plot.exposure = F,
  annotate_inferred_category_means = c("rug", "text"),
  untransform_cues = TRUE,
  levels = plogis(seq(-15, qlogis(0.95), length.out = 20)),
  category.colors = get_default_colors("category", categories)
)

plot_expected_ibbu_stanfit_categories_density2D(
  model,
  categories = get_category_levels_from_stanfit(model),
  groups = get_group_levels_from_stanfit(model, include_prior = T),
  plot.test = T,
  plot.exposure = F,
  annotate_inferred_category_means = c("rug", "text"),
  untransform_cues = TRUE,
  category.colors = get_default_colors("category", categories),
  xlim,
  ylim,
  resolution = 25
)
}
\arguments{
\item{model}{An mv-ibbu-stanfit object.}

\item{type}{Either `"contour"` or `"density"`, specifying the type of plot. Note that the contour plot is *much*
faster. It simply gets the expected values of \code{mu} (based on the NIW parameter \code{m}) and \code{Sigma}
(based on the NIW parameters \code{S} and \code{nu}) at each MCMC draw, and then averages over
all MCMC draws. The plotted categories represent those means of the expected \code{mu} and \code{Sigma}. The
density plot instead calculates the posterior predictive for each MCMC draw (i.e, the multivariate Student-T
density based on the NIW parameters \code{m, S, kappa, nu}), and then averages those densities. Since this is
done for *all* points defined by the data.grid this can be rather computationally expensive and slow.}

\item{categories, groups}{Character vector of categories/groups to be plotted. Typically, the levels of these factors
are automatically added to the fit during the creation of the fit. If necessary, however, it is possible to use
\code{\link[tidybayes]{recover_types}} on the stanfit object to add or change these levels later.
(default: all categories/groups will be plotted)}

\item{plot.test, plot.exposure}{Should the test and/or exposure stimuli be plotted? (default: `TRUE` for `plot.test`,
`FALSE` for `plot.exposure`) The test items are plotted as black points. The exposure mean is plotted as point,
and the .95 interval of cue distributions during exposure are plotted as dashed ellipse in the same color as the
expected categories.}

\item{annotate_inferred_category_means}{Character vector indicating whether the location and value of the mean be
indicated through data rugs (`"rug"`) and/or text labels (`"text"`)? Set to NULL to ignore. (default: `c("rug", "text")`)}

\item{untransform_cues}{Should m_0 and S_0 be transformed back into the original cue space? (default: `TRUE`)}

\item{levels}{Used only if `type` is `"contour"`. levels The cumulative probability levels that should be plotted (using
`geom_polygon()`) around the mean. By default the most transparent ellipse still drawn corresponds to .95.}

\item{category.colors}{Vector of colors of same length as categories.}

\item{xlim, ylim}{For density plots. Limits for the x- and y-axis.}

\item{resolution}{For density plots. How many steps along x and y should be calculated? Note that computational
complexity increases quadratically with resolution. (default: 25)}

\item{data.grid.xlim, data.grid.ylim, data.grid.resolution}{Used only if `type` is `"density"`. Limits for x- and y-axis as
well as resolution of the data.grid, defining the range over which the posterior predictive (multivariate Student-T density)
is calculated. Note that the number of densities to calculate is a *quadratic* function of `data.grid.resolution`. The default
for `data.grid.resolution` is 10, corresponding to 100 densities to be calculated for each MCMC draw.}
}
\value{
ggplot object.
}
\description{
Plot bivariate Gaussian categories expected given the parameters inferred by incremental Bayesian belief-
updating (IBBU). Specifically, the categories are derived by marginalizing over the uncertainty represented
by the (post-warmup) MCMC samples. Two methods are available (specified by `type`), which differ in their
computational demands and speed.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
