% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-fit.R
\name{plot_ibbu_categories_2D}
\alias{plot_ibbu_categories_2D}
\title{Plot prior and/or posterior category means.}
\usage{
plot_ibbu_categories_2D(
  fit,
  fit.input,
  which = c("prior", "posterior", "both")[3],
  summarize = T,
  n.draws = NULL,
  confidence.interval = c(0.025, 0.25, 0.75, 0.975),
  group.ids = NULL,
  group.labels = NULL,
  group.colors = NULL,
  group.linetypes = NULL
)
}
\value{
ggplot object.
}
\description{
Plot prior and/or posterior category means.
XXXX ADJUST BELOW
XXXX
XXXX think about distribution of means and distribution of samples.
XXXX
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
