% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-stanfit-ibbu.R
\name{plot_ibbu_stanfit_parameter_correlations}
\alias{plot_ibbu_stanfit_parameter_correlations}
\title{Plot correlations between IBBU parameters.}
\usage{
plot_ibbu_stanfit_parameter_correlations(
  model,
  category = NULL,
  group = "prior",
  cue = NULL,
  ndraws = NULL,
  untransform_cues = TRUE,
  category.colors = NULL
)
}
\arguments{
\item{model}{mv-ibbu-stanfit object.}

\item{category, group, cue}{Character vectors of categories, groups, and cues that should be included or `NULL` to include all.
(default: `NULL` for category and cue; `"prior"` for group since those are the only free parameters)}

\item{ndraws}{Number of draws to plot (or use to calculate the CIs), or `NULL` if all draws are to be returned. (default: `NULL`)}

\item{untransform_cues}{Should m_0 and S_0 be transformed back into the original cue space? (default: `TRUE`)}

\item{category.colors}{Vector of fill colors of same length as category or `NULL` to use defaults. (default: `NULL`)}
}
\value{
ggplot object.
}
\description{
Plot correlations between post-warmup MCMC samples for all parameters representing the prior and/or posterior beliefs.
}
\examples{
TBD

}
\seealso{
TBD
}
\keyword{TBD}
