% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-stanfit-ibbu.R
\name{plot_ibbu_stanfit_parameters}
\alias{plot_ibbu_stanfit_parameters}
\title{Plot distribution of IBBU parameters.}
\usage{
plot_ibbu_stanfit_parameters(
  model,
  categories = get_category_levels_from_stanfit(model),
  groups = get_group_levels_from_stanfit(model, include_prior = T),
  cues = get_cue_levels_from_stanfit(model),
  ndraws = NULL,
  untransform_cues = TRUE,
  panel_scaling = F,
  group.colors = get_default_colors("group", groups)
)
}
\arguments{
\item{model}{mv-ibbu-stanfit object.}

\item{categories, groups, cues}{Character vector of categories/groups/cues to be plotted. Typically, the levels of these factors
are automatically added to the fit during the creation of the fit. If necessary, however, it is possible to use
\code{\link[tidybayes]{recover_types}} on the stanfit object to add or change these levels later.
(default: all categories/groups/cues will be plotted)}

\item{ndraws}{Number of draws to plot (or use to calculate the CIs), or `NULL` if all draws are to be returned. (default: `NULL`)}

\item{untransform_cues}{Should m_0 and S_0 be transformed back into the original cue space? (default: `TRUE`)}

\item{panel_scaling}{Should the relative scaling be calculated separately for each panel? If not the scaling is calculated
globally. (default: `FALSE`)}

\item{group.colors}{Vector of fill colors of same length as group.ids or `NULL` to use defaults. (default: `NULL`)}

\item{which}{Should parameters for the prior, posterior, or both be plotted? (default: `"both"`)}
}
\value{
ggplot object.
}
\description{
Plot distribution of post-warmup MCMC samples for all parameters representing the
prior and/or posterior beliefs.
}
\examples{
TBD

}
\seealso{
TBD
}
\keyword{TBD}
