% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-stanfit-ibbu.R
\name{plot_ibbu_stanfit_test_categorization}
\alias{plot_ibbu_stanfit_test_categorization}
\title{Plot prior and posterior categorization of test tokens.}
\usage{
plot_ibbu_stanfit_test_categorization(
  fit,
  data.test = NULL,
  which = "both",
  summarize = T,
  n.draws = NULL,
  confidence.intervals = c(0.66, 0.95),
  target_category = 1,
  panel.group = FALSE,
  group.ids = NULL,
  group.labels = NULL,
  group.colors = NULL,
  group.linetypes = NULL,
  sort.by = "prior"
)
}
\arguments{
\item{fit}{\code{\link{NIW_ideal_adaptor_stanfit}} object.}

\item{data.test}{Optionally, a \code{tibble} or \code{data.frame} with test data.
If `NULL` the input will be extracted from fit. (default: `NULL`).}

\item{which}{Should categorization for the prior, posterior, or both be plotted? (default: `"both"`)}

\item{summarize}{Should one categorization function (optionally with CIs) be plotted (`TRUE`) or should separate
unique categorization function be plotted for each MCMC draw (`FALSE`)? (default: `TRUE`)}

\item{n.draws}{Number of draws to plot (or use to calculate the CIs), or `NULL` if all draws are to be returned.
(default: `NULL`)}

\item{confidence.intervals}{The two confidence intervals that should be plotted (using `geom_ribbon`) around the mean.
(default: `c(.66, .95)`)}

\item{target_category}{The index of the category for which categorization should be shown. (default: `1`)}

\item{panel.group}{Should the groups be plotted in separate panels? (default: `FALSE`)}

\item{group.ids}{Vector of group IDs to be plotted or leave `NULL` to plot all groups. (default: `NULL`)}

\item{group.labels}{Vector of group labels of same length as `group.ids` or `NULL` to use defaults. (default: `NULL`)
The default labels each categorization function based on whether it is showing prior or posterior categorization,
and by its group ID.}

\item{group.colors}{Vector of colors of same length as group.ids or `NULL` to use defaults. (default: `NULL`)}

\item{group.linetypes}{Vector of linetypes of same length as group.ids or `NULL` to use defaults. (default: `NULL`)}

\item{sort_by}{Which group, if any, should the x-axis be sorted by (in increasing order of posterior probability
from left to right). Set to 0 for sorting by prior (default). Set to `NULL` if no sorting is desired.}
}
\value{
ggplot object.
}
\description{
Categorize test tokens by prior and/or posterior beliefs, and plot the resulting categorization function along
a one-dimensional continuum (regardless of the dimensionality of the cue space in which categorization takes
place). This provides the type of categorization plot typical for, for example, perceptual recalibration or
phonetic tuning studies.
}
\details{
Tokens are sorted based on the increasing probability of a \code{target_category} response for the condition
(\code{group}, e.g., prior or a specific exposure group) specified in \code{sort.by}. By default both the mean
categorization and confidence intervals are plotted.
If `summarize=TRUE`, the function marginalizes over all posterior samples. The number of samples
is determined by n.draws. If n.draws is NULL, all samples are used. Otherwise n.draws random
samples will be used. If `summarize=FALSE`, separate categorization plots for all n.draws
individual samples will be plotted in separate panels.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
