% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-NIW-IA-simulation.R
\name{plot_NIW_belief_parameters}
\alias{plot_NIW_belief_parameters}
\title{Plot NIW belief or NIW beliefs object.}
\usage{
plot_NIW_belief_parameters(
  x,
  group.colors = NULL,
  facet_rows_by = NULL,
  facet_cols_by = NULL,
  animate_by = NULL,
  animation_follow = F
)
}
\arguments{
\item{x}{An \code{\link{NIW_belief}} or \code{\link{NIW_beliefs}} object.}

\item{group.colors}{Vector of fill colors of same length as the number of unique groups in the NIW_belief(s) object, or
`NULL` to use defaults. (default: `NULL`)}

\item{facet_rows_by, facet_cols_by, facet_wrap_by, animate_by}{Which group variables, if any, should be used for faceting and/or
animation? (defaults: `NULL`)}

\item{animation_follow}{Should the animation follow the data (zoom in and out)? (default: `FALSE`)}
}
\value{
ggplot object.
}
\description{
Plot the parameters of an NIW_belief or NIW_beliefs object.
}
\seealso{
TBD
}
\keyword{TBD}
