% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-visualize-basics.R
\name{plot_pairwise_cue_correlation_matrix}
\alias{plot_pairwise_cue_correlation_matrix}
\title{Deprecated: plot_pairwise_cue_correlation_matrix}
\usage{
plot_pairwise_cue_correlation_matrix(
  data,
  cues,
  category = category,
  category.colors = seq_along(unique(data$category))
)
}
\arguments{
\item{data}{Data frame.}

\item{cues}{Cue columns.}

\item{category}{Category column.}

\item{category.colors}{Category colors.}
}
\value{
A \code{ggplot} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{plot_pairwise_cue_correlation_matrix()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0. Please use \code{\link[=plot_parameters]{plot_parameters()}} instead.
}
\seealso{
\code{\link[=plot_parameters]{plot_parameters()}}
}
\keyword{internal}
