% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-visualize-NIW-IA-stanfit.R
\name{plot_parameter_correlations.ideal_adaptor_stanfit}
\alias{plot_parameter_correlations.ideal_adaptor_stanfit}
\title{Deprecated: plot_parameter_correlations.ideal_adaptor_stanfit}
\usage{
plot_parameter_correlations.ideal_adaptor_stanfit(model, ...)
}
\arguments{
\item{model}{Model object.}

\item{...}{Arguments passed to \code{\link[=plot_parameter_correlations]{plot_parameter_correlations()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{plot_parameter_correlations.ideal_adaptor_stanfit()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0. Please use \code{\link[=plot_parameter_correlations]{plot_parameter_correlations()}} instead.
}
\seealso{
\code{\link[=plot_parameter_correlations]{plot_parameter_correlations()}}
}
\keyword{internal}
