% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-class.R
\name{print.MVBU_model}
\alias{print.MVBU_model}
\title{Print MVBeliefUpdatr model}
\usage{
\method{print}{MVBU_model}(x, ...)
}
\arguments{
\item{x}{An \code{\link{MVBeliefUpdatr_model}} object.}
}
\description{
Deprecated. Use the S7-based predicates and constructors instead.
}
\details{
Specifies reasonable defaults for the parameters to be summarized for the MVBeliefUpdatr_model object.
}
\keyword{internal}
