% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-model-list.R
\name{print.MVBU_ModelList}
\alias{print.MVBU_ModelList}
\title{S3 Print method for MVBU_ModelList}
\usage{
\method{print}{MVBU_ModelList}(x, ...)
}
\arguments{
\item{x}{An \link{MVBU_ModelList} object.}

\item{...}{Additional arguments.}
}
\description{
S3 Print method for MVBU_ModelList
}
