% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mvg-ibbu-stanfit.R
\name{print.NIW_ideal_adaptor_stanfit}
\alias{print.NIW_ideal_adaptor_stanfit}
\title{Print NIW ideal adaptor stanfit}
\usage{
\method{print}{NIW_ideal_adaptor_stanfit}(x, pars = NULL, ...)
}
\arguments{
\item{x}{An \code{\link{NIW_ideal_adaptor_stanfit}} object.}
}
\description{
Specifies reasonable defaults for the parameters to be summarized for the stanfit object.
}
