% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit-postprocess.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{pp_check}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bayesplot}{\code{\link[bayesplot:pp_check]{pp_check()}}}
}}

