% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-class.R
\name{register_model_family}
\alias{register_model_family}
\title{Register a model family in the MVBU family registry}
\usage{
register_model_family(
  family,
  category_representation_class,
  cognitive_model_class
)
}
\arguments{
\item{family}{Model-family name.}

\item{category_representation_class}{Name of the category-representation
class.}

\item{cognitive_model_class}{Name of the cognitive-model class.}
}
\value{
Invisibly TRUE.
}
\description{
Register a model family in the MVBU family registry
}
