% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-objects.R
\name{sample_data_from_model}
\alias{sample_data_from_model}
\alias{sample_MVG_data}
\alias{sample_NIW_data}
\alias{sample_exemplar_data}
\title{Sample data from model.}
\usage{
sample_data_from_model(model, Ns, randomize.order = T)

sample_MVG_data(
  Ns,
  mus,
  Sigmas,
  category.labels = NULL,
  cue.labels = NULL,
  randomize.order = T
)

sample_NIW_data(
  Ns,
  ms,
  Ss,
  kappas,
  nus,
  category.labels = NULL,
  cue.labels = NULL,
  randomize.order = T
)

sample_exemplar_data(
  Ns,
  category.labels = NULL,
  cue.labels = NULL,
  randomize.order = T
)
}
\arguments{
\item{model}{\code{\link{MVG}}, \code{\link{MVG_ideal_observer}}, \code{\link{NIW_belief}}, or \code{\link{NIW_ideal_adaptor}} object.}

\item{Ns}{Integer vector, with each number specifying the number of observations to be drawn from the corresponding
Gaussian.}

\item{randomize.order}{Should the order of the data be randomized? If not, samples are sorted by category (default: TRUE)}

\item{mus, sigmas, nus, }{etc.  Lists of parameters with the n-th element of each list specifying relevant parameter necessary to characterize
the model's representations of the n-th category.}

\item{category.labels}{Character vector of category names, each specifying the category label of a multivariate Gaussian. (default: extracted
from \code{model} if available, otherwise 1:N)}

\item{cue.labels}{Character vector of cue names. If \code{NULL} (default) then the cues will be numbered cue1, cue2, ... (default: extracted
from \code{model} if available, otherwise cue1, cue2, ...)}
}
\value{
A tibble.
}
\description{
Returns a tibble of observations drawn from one of the models defined in \code{MVBeliefUpdatr}, with one observation per row. Each row
provides the category label and cue values. Lapse rate and lapse bias are currently not considered while sampling.
}
\seealso{
TBD
}
\keyword{TBD}
