% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get-info-from-model.R
\name{sample_observation}
\alias{sample_observation}
\title{Deprecated: sample_observation}
\usage{
sample_observation(x, ...)
}
\arguments{
\item{x}{A representation, template, or cognitive model object.}

\item{...}{Arguments passed to \code{\link{sample_observations}}.}
}
\value{
A tibble of sampled observations.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{sample_observation()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0. Please use \code{\link[=sample_observations]{sample_observations()}} instead.
}
\seealso{
\code{\link[=sample_observations]{sample_observations()}}
}
\keyword{internal}
