% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R
\name{set_staninput}
\alias{set_staninput}
\title{Set the stored Stan input on an S7 fit or fit-input object}
\usage{
set_staninput(x, staninput, ...)
}
\arguments{
\item{x}{An S7 fit or fit-input object.}

\item{staninput}{An S7 Stan input object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
The updated S7 object.
}
\description{
Set the stored Stan input on an S7 fit or fit-input object
}
