% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-fit.R
\name{signed_log}
\alias{signed_log}
\alias{signed_log_trans}
\title{Signed log transform and function.}
\usage{
signed_log(x)

signed_log_trans()
}
\description{
Makes it possible to use log-stepped scales or coordinate systems even when negative values
are included in the data. E.g., in `coord_trans(x = "signed_log")`.
}
\examples{
TBD
}
