% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{ss}
\alias{ss}
\title{Get sum-of-squares matrix}
\usage{
ss(x, center = TRUE)
}
\arguments{
\item{x}{A matrix of observations, with each row being a vector observation.}

\item{center}{Should the centered sum-of-squares be returned (\code{TRUE}) or the uncentered (\code{FALSE})? (default: \code{TRUE})}
}
\value{
A square matrix.
}
\description{
Get sum-of-square matrix.
}
\seealso{
\code{\link{css2cov}}, \code{\link{cov2css}}, \code{\link{uss2css}}
}
\keyword{TBD}
