% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{ss2cov}
\alias{ss2cov}
\title{Get covariance matrix from centered sum-of-squares matrix}
\usage{
ss2cov(SS, n)
}
\arguments{
\item{SS}{A sum-of-square matrix.}

\item{n}{Number of observations that have gone into the sum-of-square matrix.}
}
\value{
A square matrix.
}
\description{
Get covariance matrix from centered sum-of-square matrix.
}
\examples{
TBD
}
\seealso{
\code{\link{cov2ss}}
}
\keyword{TBD}
