% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-NIW-IA-stanfit.R
\name{stanfit_needs_refit}
\alias{stanfit_needs_refit}
\title{Check if cached \code{NIW_ideal_adaptor_stanfit} can be used.}
\usage{
stanfit_needs_refit(
  fit,
  current_version = get_current_versions(),
  data = NULL,
  staninput = NULL,
  silent = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{fit}{Old \code{NIW_ideal_adaptor_stanfit} object (e.g., loaded from file).}

\item{current_version}{Current version of relevant packages.}

\item{data}{New data to check consistency of factor level names.
Pass \code{NULL} to avoid this data check.}

\item{staninput}{New Stan data (result of a call to \code{\link[standata.default]{standata}}).
Pass \code{NULL} to avoid this data check.}

\item{silent}{Logical. If \code{TRUE}, no messages will be given.}

\item{verbose}{Logical. If \code{TRUE} detailed report of the differences
is printed to the console.}
}
\value{
A boolean indicating whether a refit is needed.
}
\description{
Checks whether a given cached fit can be used without refitting when
\code{file_refit = "on_change"} is used.
}
\details{
Use with \code{verbose = TRUE} to get additional info on how the stored
fit differs from the given data and code.
}
\keyword{internal}
