% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-model-list.R
\name{[.MVBU_ModelList}
\alias{[.MVBU_ModelList}
\title{Subset an MVBU_ModelList object}
\usage{
\method{[}{MVBU_ModelList}(x, i)
}
\arguments{
\item{x}{An \link{MVBU_ModelList} object.}

\item{i}{Index or logical vector.}
}
\value{
A subsetted \link{MVBU_ModelList} object.
}
\description{
Subset an MVBU_ModelList object
}
