% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-model-list.R
\name{[[.MVBU_ModelList}
\alias{[[.MVBU_ModelList}
\title{Extract a single model element from an MVBU_ModelList}
\usage{
\method{[[}{MVBU_ModelList}(x, i)
}
\arguments{
\item{x}{An \link{MVBU_ModelList} object.}

\item{i}{Index or scalar element name.}
}
\value{
The extracted model object.
}
\description{
Extract a single model element from an MVBU_ModelList
}
