% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit-input-methods.R
\name{summary,IdealAdaptorStanfitInput-method}
\alias{summary,IdealAdaptorStanfitInput-method}
\title{Summarize an IdealAdaptorStanfitInput object}
\usage{
## S7 method for class <IdealAdaptorStanfitInput>
summary(object, ...)
}
\arguments{
\item{object}{An \code{\link{IdealAdaptorStanfitInput}} object.}

\item{...}{Additional arguments.}
}
\value{
An object of class \code{Summary_IdealAdaptorStanfitInput} containing
exposure statistics and test data summaries.
}
\description{
Provides a detailed overview of the exposure sufficient statistics across
group and category combinations as well as key properties of the test dataset.
}
