% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit-methods.R
\name{summary,MVBU_Stanfit-method}
\alias{summary,MVBU_Stanfit-method}
\title{Summarize an MVBU Stanfit object}
\usage{
## S7 method for class <MVBU_Stanfit>
summary(
  object,
  pars = NULL,
  sufficient_pars_only = TRUE,
  indices_as_names = TRUE,
  include_transformed_pars = FALSE,
  ...
)
}
\arguments{
\item{object}{An \code{\link{MVBU_Stanfit}} object.}

\item{pars}{A character vector of parameter names to be summarized. If \code{NULL}, all model
parameters are summarized. (default: \code{NULL})}

\item{sufficient_pars_only}{Should only the sufficient parameters be summarized? (default: \code{TRUE})}

\item{indices_as_names}{Should the indices of the parameters be translated into level names in the summary?
This output is substantially more readable, with names shown in separate columns (distribution, group,
category, cue). (default: \code{TRUE})}

\item{include_transformed_pars}{Should transformed parameters be included in the summary? (default: \code{FALSE})}

\item{...}{Additional arguments passed to \code{rstan::summary}.}
}
\value{
An object of class \code{Summary_MVBU_Stanfit} containing parameter summaries and diagnostic statistics.
}
\description{
Specifies reasonable defaults for the parameters to be summarized for the stanfit object.
}
