% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-model-list.R
\name{summary.MVBU_ModelList}
\alias{summary.MVBU_ModelList}
\title{S3 Summary method for MVBU_ModelList}
\usage{
\method{summary}{MVBU_ModelList}(object, ...)
}
\arguments{
\item{object}{An \link{MVBU_ModelList} object.}

\item{...}{Additional arguments.}
}
\description{
S3 Summary method for MVBU_ModelList
}
