% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit-input-methods.R
\name{Summary_IdealAdaptorStanfitInput}
\alias{Summary_IdealAdaptorStanfitInput}
\title{An S7 class for IdealAdaptorStanfitInput summaries}
\usage{
Summary_IdealAdaptorStanfitInput(
  object_class = character(0),
  categories = character(0),
  cues = character(0),
  groups = character(0),
  exposure_statistics = NULL,
  test_summary = list()
)
}
\description{
An S7 class for IdealAdaptorStanfitInput summaries
}
\keyword{internal}
