% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit-methods.R
\name{Summary_MVBU_Stanfit}
\alias{Summary_MVBU_Stanfit}
\title{An S7 class for MVBU_Stanfit summaries}
\usage{
Summary_MVBU_Stanfit(
  fitted = NULL,
  fixed = NULL,
  expected_moments = NULL,
  marginal_moments = NULL,
  high_rhats = NULL
)
}
\description{
An S7 class for MVBU_Stanfit summaries
}
\keyword{internal}
