% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-visualize-basics.R
\name{symlog}
\alias{symlog}
\title{Deprecated: symlog}
\usage{
symlog(x, C = 0)
}
\arguments{
\item{x}{Numeric vector to transform.}

\item{C}{Scaling constant determining resolution around zero. Defaults to 0.}
}
\value{
Transformed numeric vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{symlog()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0. Please use \code{\link[scales:pseudo_log_trans]{scales::pseudo_log_trans()}} instead.
}
\seealso{
\code{\link[scales:pseudo_log_trans]{scales::pseudo_log_trans()}}
}
\keyword{internal}
