% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{transform_cues}
\alias{transform_cues}
\alias{untransform_cues}
\title{Transform and untransform cues by applying or undoing PCA, centering, and/or scaling}
\usage{
transform_cues(
  data,
  cues,
  center = if (pca) T else F,
  scale = F,
  pca = F,
  attach = T,
  transform.parameters = NULL,
  return.transformed.data = T,
  return.transform.parameters = F,
  return.transform.function = F,
  return.untransform.function = F
)

untransform_cues(
  data,
  cues,
  uncenter = NULL,
  unscale = NULL,
  unpca = NULL,
  attach = T,
  transform.parameters = NULL,
  return.untransformed.data = T,
  return.untransform.function = F
)
}
\arguments{
\item{data}{\code{tibble} or \code{data.frame}.}

\item{cues}{Vector of characters with names of cue variables.}

\item{center, uncenter}{Should the data be (un)centered? (default: \code{TRUE} unless \code{pca = TRUE})}

\item{scale, unscale}{Should the data be (un)standardized? (default: \code{TRUE} unless \code{pca = TRUE})}

\item{pca, unpca}{Should the data be transformed into/back from orthogonal principal components? If \code{TRUE}
then \code{center} and \code{scale} are default to \code{FALSE}. If PCA is applied, it is applied after
centering and/or scaling. (default: \code{FALSE})}

\item{attach}{Should the transformed cues be attached to \code{data} or should just the transformed cues be
returned? (default: \code{TRUE})}

\item{transform.parameters}{List of transforms (default: \code{NULL})}

\item{return.transformed.data, return.untransformed.data}{Should the (un)transformed data be returned? (default: \code{TRUE})}

\item{return.transform.parameters}{Should the list of transforms be returned? (default: \code{FALSE})}

\item{return.transform.function, return.untransform.function}{Should a function that applies the (un)transform be
returned? (default: \code{FALSE})}
}
\value{
By default a \code{data.frame}. If \code{return.transform.parameters = TRUE}, a list of parameters. If
\code{return.transform.function = TRUE} or \code{return.untransform.function = TRUE} a function. If more than one of
these flags is \code{TRUE} then a list in which
the data element has name "data", the transform parameters have name "transform.parameters" and the (un)transform
function(s) have the name "(un)transform.function".
}
\description{
If the \code{transform.parameters} argument
is specified, the transforms in that object will be applied. This can be useful when the goal is to transform one
data set (e.g., test data) based on the statistics of the another data set (e.g., exposure data). If no
\code{transform.parameters} are specified, then the transformations specified by the \code{center}, \code{scale}, and \code{pca}
flags will be applied. The transform and untransform functions can also return \emph{functions} that perform their
actions for the specific cues. This can be helpful if one wants to store those functions. For example, the
transform function can return both the transform and untransform functions necessary to perform the specified
centering, scaling, and/or pca \emph{and} to undo those transformations.
}
\keyword{TBD}
