% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-NIW-IA-stanfit.R
\docType{class}
\name{transform_information-class}
\alias{transform_information-class}
\alias{transform_information}
\title{S7 class for transformation information for Stan models}
\usage{
transform_information(
  transform.parameters = list(),
  transform.function = function() NULL,
  untransform.function = function() NULL
)
}
\description{
S7 class for transformation information for Stan models
}
