% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R, R/S7-update-model.R
\name{update_category_representation}
\alias{update_category_representation}
\alias{update_category_representation,MVBeliefUpdatr::NIW_CategoryRepresentation,any,any,any-method}
\alias{update_category_representation,MVBeliefUpdatr::NIX_CategoryRepresentation,any,any,any-method}
\alias{update_category_representation,MVBeliefUpdatr::MNIX_CategoryRepresentation,any,any,any-method}
\title{Update a category representation from observations}
\usage{
update_category_representation(x, x_N, x_mean, x_SS, ...)

## S7 method for classes <MVBeliefUpdatr::NIW_CategoryRepresentation>, <any>, <any>, <any>
update_category_representation(x, x_N, x_mean, x_SS, ...)

## S7 method for classes <MVBeliefUpdatr::NIX_CategoryRepresentation>, <any>, <any>, <any>
update_category_representation(x, x_N, x_mean, x_SS, ...)

## S7 method for classes <MVBeliefUpdatr::MNIX_CategoryRepresentation>, <any>, <any>, <any>
update_category_representation(x, x_N, x_mean, x_SS, ...)
}
\arguments{
\item{x}{A category representation object.}

\item{x_N}{Number of observations represented by the sufficient statistics.}

\item{x_mean}{Mean vector of the represented observations.}

\item{x_SS}{Centered sum-of-squares matrix of the represented observations.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An updated category representation.
}
\description{
Update a category representation from observations
}
