% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-decision-model.R
\name{update_model_decision_bias_incrementally}
\alias{update_model_decision_bias_incrementally}
\title{Update model's decision biases based on exposure data}
\usage{
update_model_decision_bias_incrementally(
  model,
  beta,
  exposure,
  exposure.category = "category",
  exposure.cues = get_cue_labels_from_model(model),
  exposure.order = NULL,
  noise_treatment = .infer_noise_treatment(dplyr::first(model$Sigma_noise)),
  lapse_treatment = "no_lapses",
  keep.update_history = TRUE,
  keep.exposure_data = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{model}{A \code{\link[=is.MVBU_model]{model}} object with decision biases.}

\item{beta}{The learning rate with which decision biases change.}

\item{exposure}{\code{data.frame} or \code{tibble} with exposure data. Each row is assumed to contain one observation.}

\item{exposure.category}{Name of variable in \code{data} that contains the category information. (default: "category")}

\item{exposure.cues}{Name(s) of variables in \code{data} that contain the cue information. By default these cue names are
extracted from the prior object.}

\item{exposure.order}{Name of variable in \code{data} that contains the order of the exposure data. If \code{NULL} the
exposure data is assumed to be in the order in which it should be presented.}

\item{noise_treatment}{Determines whether and how multivariate Gaussian noise is considered during categorization.
See \code{\link{update_model_decision_bias_by_one_observation}}.}

\item{lapse_treatment}{Determines whether attentional lapses can occur during which no updating occurs.
See \code{\link{update_model_decision_bias_by_one_observation}}.}

\item{keep.update_history}{Should the history of the updating be stored and returned? If so, the output is
tibble with the one model for each exposure observation. This is useful, for example, if one wants to
visualize the changes in the category parameters, posterior predictive, categorization function, or alike across time.
(default: \code{TRUE})}

\item{keep.exposure_data}{Should the input data be included in the output? If \code{FALSE} then only the category and cue
columns will be kept. If \code{TRUE} then all columns will be kept. (default: \code{FALSE})}

\item{verbose}{Should more informative output be provided?}
}
\value{
An model object.
}
\description{
Returns the model with updated decision biases.
}
\seealso{
\code{\link{update_model_decision_bias_by_one_observation}}, which is called by \code{update_model_decision_bias_incrementally}
}
