% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{update_NIW_belief_by_sufficient_statistics}
\alias{update_NIW_belief_by_sufficient_statistics}
\alias{update_NIW_belief_by_one_observation}
\title{Update NIW prior beliefs about multivariate Gaussian category based on sufficient statistics of observations.}
\usage{
update_NIW_belief_by_sufficient_statistics(
  prior,
  x_category,
  x_mean,
  x_S,
  x_N,
  add_noise = NULL
)

update_NIW_belief_by_one_observation(prior, x_category, x, add_noise = NULL)
}
\arguments{
\item{prior}{An \code{\link[=is.NIW_belief]{NIW_belief}} object, specifying the prior beliefs.}

\item{x_category}{Character. The label of the category that is to be updated.}

\item{x_mean}{Mean of the observations.}

\item{x_S}{Centered sum of squares matrix of the observations.}

\item{x_N}{Number of observations that went into the mean and sum of squares matrix.}

\item{add_noise}{Determines whether multivariate Gaussian noise is added to the input.
If `NULL`, no noise is added during the updating. If "sample" then a sample of
noise is added to the input. If "marginalize" then each observation is transformed into the marginal distribution
that results from convolving the input with noise. This latter option might be helpful, for example, if one is
interested in estimating the consequences of noise across individuals. If add_noise is not `NULL` a Sigma_noise
column must be present in the NIW_belief object specified as the priors argument. (default: `NULL`)}

\item{x}{A single observation.}
}
\value{
A tibble.
}
\description{
Returns updated/posterior beliefs about the Gaussian categories based on conjugate NIW prior.
}
\details{
The priors for the categories are specified through the \code{priors} argument, an
\code{\link[=is.NIW_belief]{NIW_belief}} object. Which category is updated is determined by x_category.
Updating proceeds as in \insertCite{@see @murphy2012 p. 134;textual}{MVBeliefUpdatr}. The prior kappa
and nu will be incremented by the number of observations (x_N). The prior M and S will be updated based on the
prior kappa, prior nu, x_N and, of course, the sample mean (x_mean) and sum of squares (x_S) of the observations.
}
\examples{
TBD
}
\references{
\insertRef{murphy2012}{MVBeliefUpdatr}
}
\seealso{
TBD
}
\keyword{TBD}
