% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-update-NIW-IA.R
\name{update_NIW_belief_kappa}
\alias{update_NIW_belief_kappa}
\title{Deprecated: update_NIW_belief_kappa}
\usage{
update_NIW_belief_kappa(kappa_0, x_N)
}
\arguments{
\item{kappa_0}{Prior kappa value.}

\item{x_N}{Number of observations.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{update_NIW_belief_kappa()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0. Please use \code{\link[=update_category_representation]{update_category_representation()}} instead.
}
\seealso{
\code{\link[=update_category_representation]{update_category_representation()}}
}
\keyword{internal}
