% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-update-NIW-IA.R
\name{update_NIW_belief_S}
\alias{update_NIW_belief_S}
\title{Deprecated: update_NIW_belief_S}
\usage{
update_NIW_belief_S(kappa_0, m_0, S_0, x_N, x_mean, x_SS)
}
\arguments{
\item{kappa_0}{Prior kappa value.}

\item{m_0}{Prior mean vector.}

\item{S_0}{Prior scatter matrix.}

\item{x_N}{Number of observations.}

\item{x_mean}{Observation mean.}

\item{x_SS}{Centered observation sum-of-squares matrix.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{update_NIW_belief_S()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0. Please use \code{\link[=update_category_representation]{update_category_representation()}} instead.
}
\seealso{
\code{\link[=update_category_representation]{update_category_representation()}}
}
\keyword{internal}
