% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-update-NIW-IA.R
\name{update_NIW_parameters}
\alias{update_NIW_parameters}
\alias{update_NIW_belief_kappa}
\alias{update_NIW_belief_nu}
\alias{update_NIW_belief_m}
\alias{update_NIW_belief_S}
\title{Deprecated NIW parameter helpers}
\usage{
update_NIW_belief_kappa(kappa_0, x_N)

update_NIW_belief_nu(nu_0, x_N)

update_NIW_belief_m(kappa_0, m_0, x_N, x_mean)

update_NIW_belief_S(kappa_0, m_0, S_0, x_N, x_mean, x_SS)
}
\arguments{
\item{kappa_0}{Prior kappa value.}

\item{x_N}{Number of observations.}

\item{nu_0}{Prior nu value.}

\item{m_0}{Prior mean vector.}

\item{x_mean}{Observation mean.}

\item{S_0}{Prior scatter matrix.}

\item{x_SS}{Centered observation sum-of-squares matrix.}
}
\description{
Deprecated NIW parameter helpers

Deprecated: update_NIW_belief_kappa

Deprecated: update_NIW_belief_nu

Deprecated: update_NIW_belief_m

Deprecated: update_NIW_belief_S
}
\keyword{internal}
