% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{uss2css}
\alias{uss2css}
\alias{css2uss}
\alias{css2cov}
\alias{cov2css}
\title{Convert sum-of-square and covariance matrices}
\usage{
uss2css(uss, n, mean)

css2uss(css, n, mean)

css2cov(css, n)

cov2css(cov, n)
}
\arguments{
\item{uss, css, cov}{Uncentered sum-of-square, centered sum-of-square, and covariance matrix.}

\item{n}{Number of observations that have gone into the sum-of-square matrix.}

\item{mean}{Mean of the observations that have gone into the sum-of-square matrix.}
}
\value{
A square matrix.
}
\description{
Convert centered sum-of-square matrices (css), uncentered sum-of-square matrices (uss),
and covariance matrices (cov) into each other.
}
\seealso{
\code{\link{ss}}, \code{\link[stats]{cov2cor}}, \code{\link{cor2cov}}, \code{\link{cov2tau}}
}
\keyword{TBD}
