% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-classes.R
\name{UVG-IdealObserver-class}
\alias{UVG-IdealObserver-class}
\alias{UVG_IdealObserver}
\title{UVG Ideal Observer Model Class}
\usage{
UVG_IdealObserver(
  category_template = (structure(function (representations = list(), metadata = list()) 

    S7::new_object(MVBU_Object(), representations = representations, metadata =
    metadata), name = "MVBU_CategoryRepresentationTemplate", parent = structure(function
    () 
 {
     S7::new_object(S7::S7_object())
 }, name = "MVBU_Object", parent =
    structure(function () 
 {
     .Call(S7_object_)
 }, name = "S7_object", properties =
    list(), abstract = FALSE, constructor = function () 
 {
     .Call(S7_object_)
 },
    validator = function (self) 
 {
     if (!is_S7_type(self)) {
        
    "Underlying data is corrupt"
     }
 }, class = c("S7_class", "S7_object")),
    properties = list(), abstract = FALSE, constructor = function () 
 {
    
    S7::new_object(S7::S7_object())
 }, class = c("S7_class", "S7_object")), properties =
    list(representations = structure(list(name = "representations", class =
    structure(list(class = "list", constructor_name = "list", constructor = function
    (.data = list()) 
 .data, validator = function (object) 
 {
     if
    (base_class(object) != name) {
        
    sprintf("Underlying data must be <\%s> not <\%s>", name, base_class(object))
     }

    }), class = "S7_base_class"), getter = NULL, setter = NULL, validator = NULL, default
    = NULL), class = "S7_property"), metadata = structure(list(name = "metadata", class =
    structure(list(class = "list", constructor_name = "list", constructor = function
    (.data = list()) 
 .data, validator = function (object) 
 {
     if
    (base_class(object) != name) {
        
    sprintf("Underlying data must be <\%s> not <\%s>", name, base_class(object))
     }

    }), class = "S7_base_class"), getter = NULL, setter = NULL, validator = NULL, default
    = NULL), class = "S7_property")), abstract = FALSE, constructor = function
    (representations = list(), metadata = list()) 
 S7::new_object(MVBU_Object(),
    representations = representations, metadata = metadata), validator = function (self)
    
 {
     if (length(self@representations) < 1) {
        
    return("representations must contain at least one category representation object.")
 
       }
     if (!all(vapply(self@representations, function(r) S7::S7_inherits(r,
    MVBU_CategoryRepresentation), logical(1)))) {
        
    return("all representations entries must inherit from MVBU_CategoryRepresentation.")

        }
     repr_names <- names(self@representations)
     if (!is.null(repr_names))
    {
         if (length(repr_names) != length(self@representations) || any(repr_names
    == "") || anyDuplicated(repr_names) > 0) {
            
    return("if representations are named, names must be non-empty and unique.")
        
    }
     }
     NULL
 }, class = c("S7_class", "S7_object")))(),
  category_posterior_functions = list(),
  decision_rule = character(0),
  category_prior = integer(0),
  lapse_behavior = list(),
  noise_behavior = list(),
  metadata = list()
)
}
\description{
Cognitive model class for univariate Gaussian observation.
Inherits all base cognitive-model machinery and constrains
category representations to \link{UVG_CategoryRepresentation} at constructor level.
}
\seealso{
\link{NIX_IdealAdaptor}, \link{UVG_IdealObserverDistribution}
}
\keyword{internal}
