% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-classes.R, R/s7-foundation.R
\name{validate_object}
\alias{validate_object}
\title{Validate a model object}
\usage{
validate_object(x)

validate_object(x)
}
\description{
Validate a model object

Validate a model object
}
\keyword{internal}
