% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-information-from-fit.R
\name{add_ibbu_draws}
\alias{add_ibbu_draws}
\title{Add MCMC draws of all parameters from incremental Bayesian belief-updating (IBBU) to a data.frame.}
\usage{
add_ibbu_draws(
  fit,
  which = c("prior", "posterior", "both")[2],
  draws = NULL,
  summarize = FALSE,
  spread = TRUE
)
}
\arguments{
\item{fit}{mv-ibbu-stanfit object.}

\item{which}{Should parameters for the prior, posterior, or both be added? (default: posterior)}

\item{draws}{Vector with specific draw(s) to be returned, or NULL if all draws are to be returned. (default: NULL)}

\item{summarize}{Should the mean of the draws be returned instead of all of the draws? (default: FALSE)}

\item{spread}{Should all parameters be returned in one row? (default: FALSE)}
}
\value{
tibble with post-warmup (posterior) MCMC draws of the prior/posterior parameters of the IBBU model (kappa, nu, mu, sigma, lapse_rate)
}
\description{
DESCRIBE HERE
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
