% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-2025.R
\name{add_ibbu_stanfit_draw}
\alias{add_ibbu_stanfit_draw}
\title{Legacy wrapper for get_draws.}
\usage{
add_ibbu_stanfit_draw(...)
}
\description{
Deprecated. Use \code{\link{get_draws}} instead.

Deprecated. Use get_draws() instead.
}
\keyword{internal}
