% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-2025.R
\name{add_ibbu_stanfit_draw}
\alias{add_ibbu_stanfit_draw}
\title{Deprecated: add_ibbu_stanfit_draw}
\usage{
add_ibbu_stanfit_draw(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{add_ibbu_stanfit_draw()} was deprecated in MVBeliefUpdatr 0.1.0
and will be removed in 0.2.0. Please use \code{\link[=get_draws]{get_draws()}} instead.
}
\seealso{
\code{\link[=get_draws]{get_draws()}}
}
\keyword{internal}
