% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-2025.R, R/deprecated.R
\name{add_ibbu_stanfit_draw}
\alias{add_ibbu_stanfit_draw}
\title{DEPRECATED: add_ibbu_stanfit_draw}
\usage{
add_ibbu_stanfit_draw(...)

add_ibbu_stanfit_draw(...)
}
\description{
DEPRECATED: add_ibbu_stanfit_draw

DEPRECATED: add_ibbu_stanfit_draw
}
