% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_noise.R
\name{add_noise}
\alias{add_noise}
\title{Add Gaussian noise to cues in a data set.}
\usage{
add_noise(data, cues, method = c("sample", "marginalize"), Sigma)
}
\arguments{
\item{data}{\code{data.frame} or \code{tibble}.}

\item{cues}{Column names of cues in \code{data}.}

\item{method}{If "sample" then a sample of
noise is added to the input. If "marginalize" then each observation is transformed into the marginal distribution
that results from convolving the input with noise. Currently, only "sample" is implemented (default: "sample")}

\item{Sigma}{Covariance matrix of noise.}
}
\value{
Same as \code{data}.
}
\description{
Adds Gaussian noise to cues in a data set.
}
