% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-make-objects.R
\name{aggregate_models_by_group_structure}
\alias{aggregate_models_by_group_structure}
\title{Aggregate models by group structure (Deprecated)}
\usage{
aggregate_models_by_group_structure(x, group_structure = NULL)
}
\arguments{
\item{x}{An MVG, MVG_ideal_observer, NIW_belief, or NIW_ideal_adaptor object.}

\item{group_structure}{The group structure that will be used for aggregation.}
}
\value{
The aggregated object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Deprecated in MVBeliefUpdatr 0.0.9.
Use \code{purrr::map()} to create lists of models and \code{\link{aggregate_models}} instead.
}
\keyword{internal}
