% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-make-objects.R
\name{aggregate_models_by_group_structure}
\alias{aggregate_models_by_group_structure}
\title{Deprecated: aggregate_models_by_group_structure}
\usage{
aggregate_models_by_group_structure(x, group_structure = NULL)
}
\arguments{
\item{x}{An MVG, MVG_ideal_observer, NIW_belief, or NIW_ideal_adaptor object.}

\item{group_structure}{The group structure that will be used for aggregation.}
}
\value{
The aggregated object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{aggregate_models_by_group_structure()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0. Please use \code{\link[=aggregate_models]{aggregate_models()}} instead.
}
\seealso{
\code{\link[=aggregate_models]{aggregate_models()}}
}
\keyword{internal}
