% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mvg-ibbu-stanfit.R
\name{as.NIW_ideal_adaptor_stanfit}
\alias{as.NIW_ideal_adaptor_stanfit}
\title{Coerce stanfit into NIW_ideal_adaptor_stanfit}
\usage{
as.NIW_ideal_adaptor_stanfit(stanfit, input_data, transform_information = NULL)
}
\arguments{
\item{stanfit}{stanfit object}

\item{transform_information}{Optionally, a list of transform parameters, transform function, and corresponding untransform
function of the type returned by \code{\link[transform_cues]{transform_cues}}.}

\item{input}{Input for a call to rstan prepared for one of the acceptable
stan programs. See \code{\link{compose_data}}.}
}
\value{
NIW_ideal_adaptor_stanfit object
}
\description{
Combines a \code{\link[rstan]{stanfit}} object and its input into an \code{\link{NIW_ideal_adaptor_stanfit}} object.
}
\examples{
TBD
}
\seealso{
\code{\link[rstan]{stanfit}}
}
