% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-asserts.R
\name{assert_MVG_ideal_observer}
\alias{assert_MVG_ideal_observer}
\title{Deprecated: assert_MVG_ideal_observer}
\usage{
assert_MVG_ideal_observer(x, category = "category", verbose = F)
}
\arguments{
\item{x}{Object to test.}

\item{category}{Name of the category field to inspect.}

\item{verbose}{Logical. Whether to emit additional diagnostics while testing the object.}
}
\value{
Invisibly TRUE if the assertion passes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{assert_MVG_ideal_observer()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0. Please use S7 assertions or \link{MVG_IdealObserver} instead.
}
\seealso{
\link{MVG_IdealObserver}
}
\keyword{internal}
