% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-forward-ibbu.R
\name{check_compatibility_between_NIW_belief_and_data}
\alias{check_compatibility_between_NIW_belief_and_data}
\title{Checking that input to plotting function is compatible}
\usage{
check_compatibility_between_NIW_belief_and_data(
  x,
  data.exposure,
  data.test,
  grouping.var,
  panel.group,
  animate.group,
  return.cues = T
)
}
\description{
Checking that input to plotting function is compatible
}
