% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChodroffWilson2018}
\alias{ChodroffWilson2018}
\title{Chodroff & Wilson (2018) data on word-initial stop-voicing in native US English}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 65800 rows and 12 columns.
}
\usage{
data(ChodroffWilson2018)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\code{ChodroffWilson2018} is deprecated. Please use \link{mixer6} instead, which provides
the full Mixer 6 connected speech dataset with standardized column names and
Unicode IPA notation.
}
\references{
\insertRef{chodroff-wilson2018}{MVBeliefUpdatr}
}
\seealso{
\code{\link{mixer6}}
}
\keyword{data}
