% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose-input-for-stan.R
\name{compose_data_to_infer_prior_via_conjugate_ibbu_w_sufficient_stats}
\alias{compose_data_to_infer_prior_via_conjugate_ibbu_w_sufficient_stats}
\alias{compose_data_to_infer_prior_kappanu_via_conjugate_ibbu_w_sufficient_stats}
\title{Compose data for input to RStan}
\usage{
compose_data_to_infer_prior_via_conjugate_ibbu_w_sufficient_stats(
  exposure,
  test,
  cues,
  category = "category",
  response = "response",
  group = NULL,
  center.observations = T,
  scale.observations = T,
  pca.observations = F,
  pca.cutoff = 1,
  tau_scale,
  L_omega_scale,
  verbose = F
)

compose_data_to_infer_prior_kappanu_via_conjugate_ibbu_w_sufficient_stats()
}
\arguments{
\item{exposure}{`tibble` or `data.frame` with the exposure data. Each row should be an observation of a category,
and contain information about the category label, the cue values of the observation, and optionally grouping variables.}

\item{test}{`tibble` or `data.frame` with the test data. Each row should be an observation, and contain information
about the cue values of the test stimulus and the participant's response.}

\item{cues}{Names of columns with cue values.}

\item{category}{Name of column that contains the category label for the exposure data. Can be `NULL` for unsupervised updating
(not yet implemented). (default: "category")}

\item{response}{Name of column that contains participants' responses for the test data. (default: "response")}

\item{group}{Name of column that contains information about which observations form a group. This could be individual
subjects or conditions in an experiment. The latter is more efficient, but should only be used if exposure is
identical for every individual within the group. Test does not have to be identical for every individual within
the same group. For example, one can group multiple groups of subjects that have received the same exposure
but were tested on different test tokens.}

\item{center.observations}{Should the data be centered? (default: `TRUE`)}

\item{scale.observations}{Should the data be standardized? (default: `TRUE`)}

\item{pca.observations}{Should the data be transformed into orthogonal principal components? (default: `FALSE`)}

\item{pca.cutoff}{Determines which principal components are handed to the MVBeliefUpdatr Stan program: all
components necessary to explain at least the pca.cutoff of the total variance. (default: .95) Ignored if
`pca.observation = FALSE`. (default: 1)}
}
\value{
A list that is an \code{mvg_ibbu_input}.
}
\description{
Take exposure and test data as input, and prepare the data for input into an MVBeliefUpdatr Stan program.
}
\examples{
TBD
}
\seealso{
\code{\link{is.mvg_ibbu_input}}
}
\keyword{TBD}
