% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose-input-for-stan.R
\name{compose_data_to_infer_prior_via_conjugate_ibbu_w_sufficient_stats}
\alias{compose_data_to_infer_prior_via_conjugate_ibbu_w_sufficient_stats}
\title{Compose data for input to RStan}
\usage{
compose_data_to_infer_prior_via_conjugate_ibbu_w_sufficient_stats(
  exposure,
  test,
  cues,
  category = "category",
  response = "response",
  group = "group",
  group.unique,
  center.observations = T,
  scale.observations = T,
  pca.observations = F,
  pca.cutoff = 1,
  m_0 = NULL,
  S_0 = NULL,
  tau_scale = 0,
  L_omega_scale = 0,
  verbose = F
)
}
\arguments{
\item{exposure}{`tibble` or `data.frame` with the exposure data. Each row should be an observation of a category,
and contain information about the category label, the cue values of the observation, and optionally grouping variables.}

\item{test}{`tibble` or `data.frame` with the test data. Each row should be an observation, and contain information
about the cue values of the test stimulus and the participant's response.}

\item{cues}{Names of columns with cue values. Must exist in both exposure and test data.}

\item{category}{Name of column in exposure data that contains the category label. Can be `NULL` for unsupervised updating
(not yet implemented). (default: "category")}

\item{response}{Name of column in test data that contains participants' responses. (default: "response")}

\item{group}{Name of column that contains information about which observations form a group. Typically, this is
a variable identifying subjects/participants. Must exist in both exposure and test data. (default: "group")}

\item{group.unique}{Name of column that uniquely identifies each group with identical exposure. This could be a
variable indicating the different conditions in an experiment. Using group.unique is optional, but can be
substantially more efficient if many groups share the same exposure.}

\item{center.observations}{Should the data be centered? (default: `TRUE`)}

\item{scale.observations}{Should the data be standardized? (default: `TRUE`)}

\item{pca.observations}{Should the data be transformed into orthogonal principal components? (default: `FALSE`)}

\item{pca.cutoff}{Determines which principal components are handed to the MVBeliefUpdatr Stan program: all
components necessary to explain at least the pca.cutoff of the total variance. (default: .95) Ignored if
`pca.observation = FALSE`. (default: 1)}

\item{m_0, S_0}{Optionally, prior mean of means (m_0) and prior scatter matrix (S_0). If provided, these parameters
won't be fitted.}

\item{tau_scale, L_omega_scale}{Optionally, scale for the Cauchy prior for standard deviations of the covariance
matrix of mu_0 and scale for the LKJ prior for the correlations of the covariance matrix of mu_0. Set to 0 to
ignore. (default: 0)}
}
\value{
A list that is an \code{NIW_ibbu_input}.
}
\description{
Take exposure and test data as input, and prepare the data for input into an MVBeliefUpdatr Stan program.
}
\details{
It is important to use \code{group} to identify individuals that had a specific exposure (or no exposure at all)
and specific test trials. You should \emph{not} use \code{group} to identify exposure conditions. Setting \code{group} to an exposure condition
results in an exposure that concatenates the exposure observations from all subjects in that condition. Instead, use
\code{group.unique} to identify groups with identical exposure. This will correctly use only one unique instance of the
observations that any level of \code{group} receives during exposure.
}
\examples{
TBD
}
\seealso{
\code{\link{is.NIW_ibbu_input}}
}
\keyword{TBD}
