% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose-input-for-stan.R
\name{compose_data_to_infer_prior_via_conjugate_ibbu_w_sufficient_stats}
\alias{compose_data_to_infer_prior_via_conjugate_ibbu_w_sufficient_stats}
\title{Compose data for input to RStan}
\usage{
compose_data_to_infer_prior_via_conjugate_ibbu_w_sufficient_stats(
  exposure,
  test,
  cues,
  category = "category",
  response = "response",
  group = "group",
  group.unique = NULL,
  center.observations = T,
  scale.observations = T,
  pca.observations = F,
  pca.cutoff = 1,
  m_0 = NULL,
  S_0 = NULL,
  tau_scale = 0,
  L_omega_scale = 0,
  verbose = F
)
}
\arguments{
\item{exposure}{`tibble` or `data.frame` with the exposure data. Each row should be an observation of a category,
and contain information about the category label, the cue values of the observation, and optionally grouping variables.}

\item{test}{`tibble` or `data.frame` with the test data. Each row should be an observation, and contain information
about the cue values of the test stimulus and the participant's response.}

\item{cues}{Names of columns with cue values. Must exist in both exposure and test data.}

\item{category}{Name of column in exposure data that contains the category label. Can be `NULL` for unsupervised updating
(not yet implemented). (default: "category")}

\item{response}{Name of column in test data that contains participants' responses. (default: "response")}

\item{group}{Name of column that contains information about which observations form a group. Typically, this is
a variable identifying subjects/participants. Must exist in both exposure and test data. (default: "group")}

\item{group.unique}{Name of column that uniquely identifies each group with identical exposure. This could be a
variable indicating the different conditions in an experiment. Using group.unique is optional, but can be
substantially more efficient if many groups share the same exposure. To ignore, set to `NULL`. (default: `NULL`)}

\item{center.observations}{Should the data be centered? Centering will not affect the inferred correlation or
covariance matrices but it will affect the absolute position of the inferred means. The relative position of
the inferred means remains unaffected. If `TRUE` and `m_0` is specified, `m_0` will also be centered (`S_0` is not
affected by centering and thus not changed). (default: `TRUE`)}

\item{scale.observations}{Should the data be standardized? Scaling will not affect the inferred correlation matrix,
but it will affect the inferred covariance matrix because it affects the inferred standard deviations. It will also
affect the absolute position of the inferred means. The relative position of the inferred means remains unaffected.
If `TRUE` and `m_0` and `S_0` are specified, `m_0` and `S_0` will also be scaled.
(default: `TRUE`)}

\item{pca.observations}{Should the data be transformed into orthogonal principal components? (default: `FALSE`)}

\item{pca.cutoff}{Determines which principal components are handed to the MVBeliefUpdatr Stan program: all
components necessary to explain at least the pca.cutoff of the total variance. (default: .95) Ignored if
`pca.observation = FALSE`. (default: 1)}

\item{m_0, S_0}{Optionally, prior means (m_0) and/or prior scatter matrices (S_0) for all categories. Each should be
a list, with each element being the mean/scatter matrix for a specific category. Elements should be ordered in the same order as
the levels of the category variable in \code{exposure} and \code{test}. The means and scatter matrices could be
estimated, for example, from phonetically annotated speech recordings (see \code{\link{make_NIW_prior_from_data}}
for a convenient way to do so). To aspects should be kept in mind, however. First, an \emph{inferred} scatter
matrix includes variability from internal perceptual and/or external environmental noise, \emph{in addition} to the motor noise that
is reflected in production data. That means that any prior for the scatter matrix (S_0) should take into account
not only the variability in the production data but also any additional variability added by internal (perceptual)
or external (environmental) noise. Second, the prior scatter matrix has an implicit nu associated with it, so that
the nu inferred by \code{\link{infer_prior_beliefs}} is best thought of as a multiple of the implicit nu used
during the creation of the scatter matrix S_0. For that reason, we recommend the use of nu = D + 2 in the call to
\code{\link{make_NIW_prior_from_data}} (the default), since the S_0 obtained that way is identical to the category
covariance matrix Sigma.}

\item{tau_0_scales}{Optionally, a vector of scales for the Cauchy priors for each cue's standard deviations. Used in
both the prior for m_0 and the prior for S_0. (default: vector of 5s of length of cues, assumes scaled input)}

\item{omega_0_eta}{Optionally, etas the LKJ prior for the correlations of the covariance matrix of mu_0. Set to 0 to
ignore. (default: 0)}
}
\value{
A list that is an \code{NIW_ideal_adaptor_input}.
}
\description{
Take exposure and test data as input, and prepare the data for input into an MVBeliefUpdatr Stan program.
}
\details{
It is important to use \code{group} to identify individuals that had a specific exposure (or no exposure at all)
and specific test trials. You should \emph{not} use \code{group} to identify exposure conditions. Setting \code{group} to an exposure condition
results in an exposure that concatenates the exposure observations from all subjects in that condition. Typically, this
is not what users intend, as it models exposure to the combination of exposure tokens across all subjects, rather than
exposure to one set of those exposure tokens. To achieve this intended outcome, use
\code{group.unique} to identify groups with identical exposure. This will correctly use only one unique instance of the
observations that any level of \code{group} receives during exposure.
}
\examples{
TBD
}
\seealso{
\code{\link{is.NIW_ideal_adaptor_input}}
}
\keyword{TBD}
