% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfit-diagnostics.R
\name{control_params}
\alias{control_params}
\alias{control_params.ideal_adaptor_stanfit}
\title{Extract Control Parameters of the NUTS Sampler}
\usage{
control_params(x, ...)

\method{control_params}{ideal_adaptor_stanfit}(x, pars = NULL, ...)
}
\arguments{
\item{x}{An \R object}

\item{...}{Currently ignored.}

\item{pars}{Optional names of the control parameters to be returned.
If \code{NULL} (the default) all control parameters are returned.
See \link[rstan:stan]{rstan::stan} for more details.}
}
\value{
A named \code{list} with control parameter values.
}
\description{
Extract control parameters of the NUTS sampler such as
\code{adapt_delta} or \code{max_treedepth}.
}
