% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{cor2cov}
\alias{cor2cov}
\title{Get covariance matrix from correlation matrix and standard deviations}
\usage{
cor2cov(omega, tau)
}
\arguments{
\item{omega}{A correlation matrix.}

\item{tau}{A vector of standard deviations.}
}
\value{
A square matrix.
}
\description{
Get covariance matrix from correlation matrix and vector of standard deviations. Based on code recommended
by Ben Bolder for efficiency.
}
\references{
\url{https://stats.stackexchange.com/questions/62850/obtaining-covariance-matrix-from-correlation-matrix}
}
\seealso{
\code{\link[stats]{cov2cor}}, \code{\link{cov2tau}}
}
\keyword{TBD}
