% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{cov2tau}
\alias{cov2tau}
\title{Get vector of standard deviations from covariance matrix}
\usage{
cov2tau(v)
}
\arguments{
\item{v}{A covariance matrix.}
}
\value{
A numeric vector.
}
\description{
Get vector of standard deviations from covariance matrix.
}
\seealso{
\link[stats:cor]{stats::cov2cor}, \code{\link{cor2cov}}
}
\keyword{TBD}
