% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit.R
\name{is.ideal_adaptor_stanfit}
\alias{is.ideal_adaptor_stanfit}
\title{Deprecated: is.ideal_adaptor_stanfit}
\usage{
is.ideal_adaptor_stanfit(x, verbose = FALSE)
}
\arguments{
\item{x}{Object to be checked.}

\item{verbose}{Currently being ignored.}
}
\value{
A logical.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{is.ideal_adaptor_stanfit()} is deprecated; use
\code{S7::S7_inherits(x, IdealAdaptorStanfit)} instead.
}
