% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-plot-expected-categories.R,
%   R/deprecated-visualize-NIW-IA-stanfit.R, R/deprecated-visualize-NIW-IA.R
\name{deprecated-expected-categories-plots}
\alias{deprecated-expected-categories-plots}
\alias{plot_expected_categories}
\alias{plot_expected_categories_contour}
\alias{plot_expected_categories_density}
\alias{plot_expected_categories_contour2D}
\alias{plot_expected_categories_density1D}
\alias{plot_expected_categories_density2D}
\alias{plot_expected_categories.ideal_adaptor_stanfit}
\alias{deprecated-stanfit-plots}
\alias{plot_parameters.ideal_adaptor_stanfit}
\alias{plot_parameter_correlations.ideal_adaptor_stanfit}
\alias{plot_expected_categorization_function_from_stanfit}
\alias{deprecated-niw-ia-plots}
\alias{plot_expected_categorization_function_1D}
\alias{plot_expected_categorization_function_2D}
\title{Deprecated Expected Categories Functions}
\usage{
plot_expected_categories(model, ...)

plot_expected_categories_contour(model, ...)

plot_expected_categories_density(model, ...)

plot_expected_categories_contour2D(model, ...)

plot_expected_categories_density1D(model, ...)

plot_expected_categories_density2D(model, ...)

plot_expected_categories.ideal_adaptor_stanfit(
  model,
  type = "density",
  cues = get_cue_labels(model),
  ...
)

plot_parameters.ideal_adaptor_stanfit(model, ...)

plot_parameter_correlations.ideal_adaptor_stanfit(model, ...)

plot_expected_categorization_function_from_stanfit(model, ...)

plot_expected_categorization_function_1D(x, ...)

plot_expected_categorization_function_2D(x, ...)
}
\description{
Deprecated Expected Categories Functions

Deprecated Stanfit Plotting Functions

Deprecated NIW-IA Visualization Functions
}
\keyword{internal}
