% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R, R/S7-stanfit-diagnostics.R
\name{log_posterior}
\alias{log_posterior}
\alias{nuts_params}
\alias{rhat}
\alias{neff_ratio}
\alias{diagnostic-quantities}
\alias{log_posterior,MVBU_Stanfit-method}
\alias{nuts_params,MVBU_Stanfit-method}
\alias{rhat,MVBU_Stanfit-method}
\alias{neff_ratio,MVBU_Stanfit-method}
\alias{as_draws.MVBU_Stanfit}
\alias{as_draws_df.MVBU_Stanfit}
\alias{as_draws_array.MVBU_Stanfit}
\alias{as_draws_matrix.MVBU_Stanfit}
\alias{as_draws_list.MVBU_Stanfit}
\alias{as_draws_rvars.MVBU_Stanfit}
\title{Extract log posterior draws}
\usage{
log_posterior(x, ...)

nuts_params(x, ...)

rhat(x, ...)

neff_ratio(x, ...)

## S7 method for class <MVBU_Stanfit>
log_posterior(x, ...)

## S7 method for class <MVBU_Stanfit>
nuts_params(x, pars = NULL, ...)

## S7 method for class <MVBU_Stanfit>
rhat(x, pars = NULL, ...)

## S7 method for class <MVBU_Stanfit>
neff_ratio(x, pars = NULL, ...)

\method{as_draws}{MVBU_Stanfit}(x, ...)

\method{as_draws_df}{MVBU_Stanfit}(x, ...)

\method{as_draws_array}{MVBU_Stanfit}(x, ...)

\method{as_draws_matrix}{MVBU_Stanfit}(x, ...)

\method{as_draws_list}{MVBU_Stanfit}(x, ...)

\method{as_draws_rvars}{MVBU_Stanfit}(x, ...)
}
\arguments{
\item{x}{An \code{MVBeliefUpdatr} object.}

\item{...}{Arguments passed to individual methods.}

\item{pars}{An optional character vector of parameter names.
For \code{nuts_params} these will be NUTS sampler parameter
names rather than model parameters. If pars is omitted
all parameters are included.}
}
\value{
Log posterior draws.

NUTS sampler parameters.

Named numeric vector of Rhat values.

Named numeric vector of Neff ratios.

The exact form of the output depends on the method.
}
\description{
Extract quantities that can be used to diagnose sampling behavior
of the algorithms applied by \pkg{Stan} at the back-end of \pkg{MVBeliefUpdatr}.
These diagnostic functions are all copied and modified from \pkg{brms}.
}
\details{
For more details see
\link[bayesplot:bayesplot-extractors]{bayesplot::bayesplot-extractors}.
}
