% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostic-quantities}
\alias{diagnostic-quantities}
\alias{log_posterior}
\alias{nuts_params}
\alias{rhat}
\alias{neff_ratio}
\alias{log_posterior.NIW_ideal_adaptor_stanfit}
\alias{nuts_params.NIW_ideal_adaptor_stanfit}
\alias{rhat.NIW_ideal_adaptor_stanfit}
\alias{neff_ratio.NIW_ideal_adaptor_stanfit}
\title{Extract Diagnostic Quantities of \pkg{MVBeliefUpdatr} Models}
\usage{
\method{log_posterior}{NIW_ideal_adaptor_stanfit}(x, ...)

\method{nuts_params}{NIW_ideal_adaptor_stanfit}(x, pars = NULL, ...)

\method{rhat}{NIW_ideal_adaptor_stanfit}(x, pars = NULL, ...)

\method{neff_ratio}{NIW_ideal_adaptor_stanfit}(x, pars = NULL, ...)
}
\arguments{
\item{x}{An \code{MVBeliefUpdatr} object.}

\item{...}{Arguments passed to individual methods.}

\item{pars}{An optional character vector of parameter names.
For \code{nuts_params} these will be NUTS sampler parameter
names rather than model parameters. If pars is omitted
all parameters are included.}
}
\value{
The exact form of the output depends on the method.
}
\description{
Extract quantities that can be used to diagnose sampling behavior
of the algorithms applied by \pkg{Stan} at the back-end of \pkg{MVBeliefUpdatr}.
These diagnostic functions are all copied and modified from \pkg{brms}.
}
\details{
For more details see
  \code{\link[bayesplot:bayesplot-extractors]{bayesplot-extractors}}.
}
