% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-core-classes.R
\name{.mvbu_label_information}
\alias{.mvbu_label_information}
\title{Construct a category representation set}
\usage{
.mvbu_label_information(metadata = list())
}
\description{
Construct a category representation set
}
\keyword{internal}
