% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate the fit of a model}
\usage{
evaluate_model(
  model,
  x,
  response_category,
  method = "likelihood",
  ...,
  return_by_x = F
)
}
\arguments{
\item{model}{A model object.}

\item{x}{A vector of inputs (cue values).}

\item{response_category}{A vector of category responses corresponding to each input. The model is evaluated against
this ground truth. Must be of the same length as the list of inputs, and each element of the `response_category` must
be one of the `category` levels of `model`.}

\item{method}{Method for evaluating the model. Can be "accuracy" or "likelihood". The latter returns the log-likelihood.}

\item{return_by_x}{Should results be returned separately for each unique `x`? (default: `FALSE`)}
}
\value{
If `return_by_x`, the accuracy and/or log-likelihood of each unique input `x`. Otherwise, the overall accuracy
and/or log-likelihood of the all observations. Note that the overall log-likelihood is *not* simply the sum of the
log-likelihoods of all unique inputs `x`.
}
\description{
Evaluate the fit of a categorization model against a ground truth (e.g., human responses or the category intended
by a talker).
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
