% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate the fit of a model}
\usage{
evaluate_model(
  model,
  x,
  correct_category,
  method = "likelihood",
  ...,
  return_by_x = F
)
}
\arguments{
\item{model}{A model object.}

\item{x}{A list of inputs values (vectors of cue values).}

\item{correct_category}{A list of category labels that is taken to be the ground truth. Must be of the same length as the
list of cue values.}

\item{method}{Method for evaluating the model. Can be "accuracy" or "likelihood".}

\item{return_by_x}{Should results be returned separately for each unique x? (default: `FALSE`)}
}
\value{
Either a tibble of observations with posterior probabilities for each category (in long format), or a
character vector indicating the chosen category in the same order as the observations in x (if simplify = `TRUE`).
}
\description{
Evaluate the fit of a categorization model against a ground truth (e.g., human responses or the category intended
by a talker).
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
