% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate the fit of a model}
\usage{
evaluate_model(model, cues, correct_category, method = "likelihood", ...)
}
\arguments{
\item{model}{A model object.}

\item{cues}{A list of cue values.}

\item{correct_category}{A list of category labels that is taken to be the ground truth. Must be of the same length as the
list of cue values.}

\item{method}{Method for evaluating the model. Can be "accuracy" or "likelihood".}

\item{x}{A vector of observations.}

\item{decision_rule}{Must be one of "criterion", "proportional", or "sampling".}

\item{noise_treatment}{Determines whether and how multivariate Gaussian noise is added to the input.
See \code{\link[=get_MVG_likelihood]{get_MVG_likelihood}}. (default: "sample" if decision_rule is
"sample"; "marginalize" otherwise).}

\item{lapse_treatment}{Determines whether and how lapses will be treated. Can be "no_lapses", "sample" or "marginalize".
If "sample", whether a trial is lapsing or not will be sampled for each observations. If a trial is sampled to be
a lapsing trial the lapse biases are used as the posterior for that trial. If "marginalize", the posterior probability
will be adjusted based on the lapse formula lapse_rate * lapse_bias + (1 - lapse_rate) * posterior probability from
perceptual model. (default: "sample" if decision_rule is "sample"; "marginalize" otherwise).}

\item{simplify}{Should the output be simplified, and just the label of the selected category be returned? This
option is only available for the criterion and sampling decision rules. (default: `FALSE`)}
}
\value{
Either a tibble of observations with posterior probabilities for each category (in long format), or a
character vector indicating the chosen category in the same order as the observations in x (if simplify = `TRUE`).
}
\description{
Evaluate the fit of a categorization model against a ground truth (e.g., human responses or the category intended
by a talker).
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
