% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-MVG-IO.R
\name{example_MVG_ideal_observer}
\alias{example_MVG_ideal_observer}
\title{Example MVG ideal observers.}
\usage{
example_MVG_ideal_observer(example = 1)
}
\description{
Example MVG ideal observers.
}
