% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-MVG-IO.R,
%   R/get-info-from-NIW-IA.R, R/get-info-from-exemplar-model.R,
%   R/get-info-from-model.R
\name{get_categorization_from_MVG_ideal_observer}
\alias{get_categorization_from_MVG_ideal_observer}
\alias{get_categorization_from_NIW_ideal_adaptor}
\alias{get_categorization_from_exemplar_model}
\alias{get_categorization_from_model}
\title{Legacy wrapper for categorize.}
\usage{
get_categorization_from_MVG_ideal_observer(
  x,
  model,
  decision_rule = "sampling",
  noise_treatment = if (decision_rule == "sampling") "sample" else
    infer_default_noise_treatment(model$Sigma_noise),
  lapse_treatment = if (decision_rule == "sampling") "sample" else "marginalize",
  simplify = F
)

get_categorization_from_NIW_ideal_adaptor(
  x,
  model,
  decision_rule = "sampling",
  noise_treatment = if (decision_rule == "sampling") "sample" else
    infer_default_noise_treatment(model$Sigma_noise),
  lapse_treatment = if (decision_rule == "sampling") "sample" else "marginalize",
  simplify = F,
  verbose = F
)

get_categorization_from_exemplar_model(
  x,
  model,
  decision_rule,
  noise_treatment = if (decision_rule == "sampling") "sample" else
    infer_default_noise_treatment(model$Sigma_noise),
  lapse_treatment = if (decision_rule == "sampling") "sample" else "marginalize",
  simplify = F
)

get_categorization_from_model(model, decision_rule = "sampling", ...)
}
\description{
Deprecated. Use \code{\link{categorize}} instead.

Deprecated. Use categorize() instead.

Deprecated. Use \code{\link{categorize}} instead.

Deprecated. Use categorize() instead.

Deprecated. Use \code{\link{categorize}} instead.

Deprecated. Use categorize() instead.

Deprecated. Use \code{\link{categorize}} instead.

Deprecated. Use categorize() instead.
}
\keyword{internal}
