% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-fit.R
\name{get_categorization_function}
\alias{get_categorization_function}
\title{Get categorization function}
\usage{
get_categorization_function(
  Ms,
  Ss,
  kappas,
  nus,
  lapse_rate,
  priors = 1/n.cat,
  n.cat = length(Ms),
  logit = FALSE
)
}
\arguments{
\item{Ms}{List of IBBU-inferred means describing the multivariate normal distribution over category means.}

\item{Ss}{List of IBBU-inferred scatter matrices describing the inverse Wishart distribution over category
covariance matrices.}

\item{kappas}{List of IBBU-inferred kappas describing the strength of the beliefs into the distribution over catgory means.}

\item{nus}{List of IBBU-inferred nus describing the strength of the beliefs into the distribution over catgory covariance matrices.}

\item{lapse_rate}{An IBBU-inferred lapse rate for the categorization responses.}

\item{priors}{Vector of categories' prior probabilities. (default: uniform prior over categories)}

\item{n.cat}{Number of categories. Is inferred from the input, but can be set manually.}

\item{logit}{Should the function that is returned return log-odds (TRUE) or probabilities (FALSE)? (default: TRUE)}
}
\value{
A function that takes as input cue values and returns posterior probabilities of the first category,
based on the posterior predictive of the cues given the (IBBU-derived parameters for the) categories' M, S,
kappa, nu, and prior, as well as the lapse rate.
}
\description{
Returns a categorization function for the first category, based on a set of parameters for the Normal-inverse-wishart (NIW)
distribtuion. Ms, Ss, kappas, nus, and priors are assumed to be of the same length and sorted the same way, so that the first
element of Ms is corresponding to the same category as the first element of Ss, kappas, nus, and priors, etc.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
