% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{get_categorization_function}
\alias{get_categorization_function}
\alias{get_categorization_function.ideal_adaptor_stanfit}
\title{Get categorization function from a model object}
\usage{
get_categorization_function(x, ...)

\method{get_categorization_function}{ideal_adaptor_stanfit}(
  x,
  lapse_treatment = c("no_lapses", "sample", "marginalize")[3],
  groups = get_group_levels(x, include_prior = TRUE),
  ...
)
}
\arguments{
\item{x}{A model object. Currently, only \code{\link{ideal_adaptor_stanfit}} objects are supported.}

\item{...}{Optionally, additional arguments handed to \code{\link{get_draws}}.}

\item{lapse_treatment}{Should the consequences of attentional lapses be included in the categorization function
("marginalize") or not ("no_lapses")? (default: "marginalize")}
}
\value{
A function that takes as input cue values and returns posterior probabilities of the first category,
  based on the posterior predictive of the cues given the ideal adaptor categories' parameters. The function
  will accept the following arguments:
  \itemize{
    \item{`target_category`:}{The index of the category for which categorization should be shown. (default: `1`)}
    \item{`logit`:}{Should the function return log-odds (TRUE) or probabilities (FALSE)? (default: FALSE)}
  }
}
\description{
Returns a categorization function that can be used to categorize a set of cues into categories.
}
