% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get-info-from-model.R
\name{get_category_labels_from_model}
\alias{get_category_labels_from_model}
\title{Deprecated: get_category_labels_from_model}
\usage{
get_category_labels_from_model(x, indices = NULL)
}
\arguments{
\item{x}{A model, template, or representation.}

\item{indices}{Optional integer indices of category labels to return.}
}
\value{
Character vector of category labels.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_category_labels_from_model()} is deprecated. Use \code{get_category_labels()} instead.
}
\keyword{internal}
