% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-model.R
\name{get_category_labels_from_model}
\alias{get_category_labels_from_model}
\title{Get category labels from likelihood or model}
\usage{
get_category_labels_from_model(x)
}
\arguments{
\item{x}{A likelihood or model object.}
}
\description{
Get the unique labels for all categories from a likelihood (e.g., MVG or NIW_belief) or model (e.g., an MVG ideal observer
or NIW ideal adaptor) object.
}
