% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{get_constructor}
\alias{get_constructor}
\alias{get_category_constructor}
\alias{get_group_constructor}
\alias{get_cue_constructor}
\alias{get_cue2_constructor}
\title{Get tidybayes constructor from an ideal adaptor stanfit.}
\usage{
get_constructor(x, variable = NULL)

get_category_constructor(x)

get_group_constructor(x)

get_cue_constructor(x)

get_cue2_constructor(x)
}
\arguments{
\item{x}{\code{\link{ideal_adaptor_stanfit}} object.}

\item{variable}{Either "category" or "group". If set to \code{NULL} then a list of all constructors is
returned. That list is \code{NULL} if not tidybayes constructors are found in fit. (default: c("category", "group"))}
}
\value{
A constructor function, a list of constructor functions, or \code{NULL}. If a specific constructor
function is requested but not found, a warning is shown.
}
\description{
Gets the tidybayes constructor function from the stanfit object. \code{get_category_constructor()} and
\code{get_group_constructor()} are convenience functions, calling \code{get_constructor()}. See \link[tidybayes:recover_types]{tidybayes::recover_types}.
}
\seealso{
\link[tidybayes:recover_types]{tidybayes::recover_types}, \code{\link{get_staninput_variable_levels}}
}
\keyword{TBD}
