% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-information-from-stanfit.R
\name{get_constructor}
\alias{get_constructor}
\alias{get_category_constructor}
\alias{get_group_constructor}
\title{Get tidybayes constructor from IBBU stanfit}
\usage{
get_constructor(fit, variable = c("category", "group"))

get_category_constructor(fit)

get_group_constructor(fit)
}
\arguments{
\item{fit}{mv_ibbu_stanfit object.}

\item{variable}{Either "category" or "group".}
}
\value{
A constructor function.
}
\description{
Gets the tidybayes constructor function from the stanfit object. `get_category_constructor()` and
`get_group_constructor()` are convenience functions, calling `get_constructor()`. See \code{
\link[tidybayes]{recover_types}}.
}
\examples{
TBD
}
\seealso{
\code{\link[tidybayes]{recover_types}}, \code{\link{get_original_levels}}
}
\keyword{TBD}
