% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{get_constructor}
\alias{get_constructor}
\alias{get_category_constructor}
\alias{get_group_constructor}
\alias{get_cue_constructor}
\alias{get_cue2_constructor}
\title{Get tidybayes constructor from an NIW ideal adaptor stanfit.}
\usage{
get_constructor(fit, variable = NULL)

get_category_constructor(fit)

get_group_constructor(fit)

get_cue_constructor(fit)

get_cue2_constructor(fit)
}
\arguments{
\item{fit}{\code{\link{NIW_ideal_adaptor_stanfit}} object.}

\item{variable}{Either "category" or "group". If set to `NULL` then a list of all constructors is
returned. That list is `NULL` if not tidybayes constructors are found in fit. (default: c("category", "group"))}
}
\value{
A constructor function, a list of constructor functions, or `NULL`. If a specific constructor
function is requested but not found, a warning is shown.
}
\description{
Gets the tidybayes constructor function from the stanfit object. `get_category_constructor()` and
`get_group_constructor()` are convenience functions, calling `get_constructor()`. See \code{
\link[tidybayes]{recover_types}}. If variable is
}
\seealso{
\code{\link[tidybayes]{recover_types}} from tidybayes, \code{\link{get_input_variable_levels}}
}
\keyword{TBD}
