% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-information-from-stanfit.R
\name{get_constructor}
\alias{get_constructor}
\alias{get_category_constructor}
\alias{get_group_constructor}
\title{Get tidybayes constructor from IBBU stanfit}
\usage{
get_constructor(fit, variable = c("category", "group"))

get_category_constructor(fit)

get_group_constructor(fit)
}
\arguments{
\item{fit}{mv_ibbu_stanfit object.}

\item{variable}{Either "category" or "group". If set to `NULL` then a list of all constructors is
returned. That list is `NULL` if not tidybayes constructors are found in fit. (default: c("category", "group"))}
}
\value{
A constructor function, a list of constructor functions, or `NULL`. If a specific constructor
function is requested but not found, a warning is shown.
}
\description{
Gets the tidybayes constructor function from the stanfit object. `get_category_constructor()` and
`get_group_constructor()` are convenience functions, calling `get_constructor()`. See \code{
\link[tidybayes]{recover_types}}. If variable is
}
\examples{
TBD
}
\seealso{
\code{\link[tidybayes]{recover_types}} from tidybayes, \code{\link{get_original_levels}}
}
\keyword{TBD}
