% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-model.R
\name{get_cue_dimensionality_from_model}
\alias{get_cue_dimensionality_from_model}
\title{Get cue dimensionality from likelihood or model}
\usage{
get_cue_dimensionality_from_model(x, indices = NULL)
}
\arguments{
\item{x}{A likelihood or model object.}
}
\value{
A numeric.
}
\description{
Get the number of cues from a likelihood (e.g., MVG or NIW_belief) or model (e.g., an MVG ideal observer
or NIW ideal adaptor) object.
}
