% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R
\name{get_cue_labels}
\alias{get_cue_labels}
\title{Get cue labels from a representation, template, model, or model distribution.}
\usage{
get_cue_labels(x, indices, ...)
}
\arguments{
\item{x}{A representation, representation template, or cognitive model.}

\item{indices}{An optional integer vector of indices to subset the returned labels.}
}
\value{
A character vector of cue labels.
}
\description{
Get cue labels from a representation, template, model, or model distribution.
}
