% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-model.R
\name{get_cue_labels_from_model}
\alias{get_cue_labels_from_model}
\title{Get cue labels from likelihood or model}
\usage{
get_cue_labels_from_model(x, indices = NULL)
}
\arguments{
\item{x}{A likelihood or model object.}

\item{indeces}{A vector of indices that should be turned into the original cue labels corresponding to those
indices, or \code{NULL} if all cue labels should be returned. (default: \code{NULL})}
}
\value{
A character vector.

Legacy wrapper for get_cue_labels.
}
\description{
Deprecated. Use \code{\link{get_cue_labels}} instead.
}
\details{
Get the names for all cues from a likelihood (e.g., MVG or NIW_belief) or model (e.g., an MVG ideal observer
or NIW ideal adaptor) object.
}
