% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get-info-from-model.R
\name{get_cue_labels_from_model}
\alias{get_cue_labels_from_model}
\title{Deprecated: get_cue_labels_from_model}
\usage{
get_cue_labels_from_model(x, indices = NULL)
}
\arguments{
\item{x}{A model, template, or representation.}

\item{indices}{Optional integer indices of cue labels to return.}
}
\value{
Character vector of cue labels.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_cue_labels_from_model()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0. Please use \code{\link[=get_cue_labels]{get_cue_labels()}} instead.
}
\seealso{
\code{\link[=get_cue_labels]{get_cue_labels()}}
}
\keyword{internal}
