% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get-info-from-NIW-IA-stanfit.R
\name{get_cue_levels}
\alias{get_cue_levels}
\title{Deprecated: get_cue_levels}
\usage{
get_cue_levels(x, indices = NULL, ...)
}
\arguments{
\item{x}{A model object.}

\item{indices}{Optional indices.}

\item{...}{Additional arguments.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_cue_levels()} is deprecated in favor of
\code{\link{get_cue_labels}}.
}
