% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIX-NIW-basics.R
\name{get_D}
\alias{get_D}
\title{Get dimensionality of cue representation from model (deprecated)}
\usage{
get_D(x)
}
\arguments{
\item{x}{A model, template, representation, or object from which cue labels can be extracted.}
}
\value{
Number of cues (dimensionality) as an integer.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_D()} is deprecated. Use \code{length(get_cue_labels(x))} instead.
}
