% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-belief.R
\name{get_D}
\alias{get_D}
\title{Get dimensionality of mean or covariance matrix}
\usage{
get_D(x)
}
\arguments{
\item{x}{A mean or covariance matrix, or lists of means or covariance matrices. If x is a list, then the
dimensionality of the first list element is determined. This function does not currently check whether all
list elements have the same dimensionality.}
}
\description{
Get dimensionality of mean or covariance matrix
}
