% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-basics.R
\name{get_default_colors}
\alias{get_default_colors}
\title{Get default colors}
\usage{
get_default_colors(var, n)
}
\arguments{
\item{var}{Variable for which default color is requested. Can be either `"category"` or
`"group"`.}
}
\description{
Get default colors
}
\examples{
TBD
}
