% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R, R/S7-stanfit-methods.R
\name{get_draws}
\alias{get_draws}
\alias{get_draws,MVBU_Stanfit-method}
\title{Get MCMC prior or posterior draws from a stanfit object}
\usage{
get_draws(fit, ...)

## S7 method for class <MVBU_Stanfit>
get_draws(
  fit,
  categories = get_category_labels(fit),
  groups = get_group_labels(fit, include_prior = TRUE),
  which = if ("prior" \%in\% groups) {
     if (length(groups) > 1) 
         "both"
    
    else "prior"
 } else {
     "posterior"
 },
  ndraws = NULL,
  untransform_cues = FALSE,
  summarize = FALSE,
  wide = FALSE,
  nest = TRUE,
  seed = if (!is.null(ndraws)) runif(1, -1e+06, 1e+06) else NULL,
  ...
)
}
\arguments{
\item{fit}{An \code{\link{IdealAdaptorStanfit}} (or \code{\link{MVBU_Stanfit}}) object.}

\item{...}{Additional arguments passed to methods.}

\item{categories}{Character vector of category names for which draws should be returned. (default: all categories in model)}

\item{groups}{Character vector of group names for which draws should be returned. (default: all groups in model including prior)}

\item{which}{DEPRECATED. Use \code{groups} instead. Should parameters for the prior, posterior, or both be returned? (default: \code{"posterior"})}

\item{ndraws}{Number of random draws or \code{NULL} if all draws are to be returned. (default: \code{NULL})}

\item{untransform_cues}{Should m_0 and S_0 be transformed back into the original cue space? (default: \code{FALSE})}

\item{summarize}{Should the mean of the draws be returned instead of all of the draws? (default: \code{FALSE})}

\item{wide}{Should all parameters be returned in one row? (default: \code{FALSE})}

\item{nest}{Should the category mean vectors and scatter matrices be nested into one cell each, or should each element
be stored in a separate column? (default: \code{TRUE})}

\item{seed}{Optional seed for reproducibility when sampling draws. (default: \code{NULL})}
}
\value{
A tibble with MCMC draws of the model parameters.
}
\description{
Get MCMC draws of all parameters from incremental Bayesian belief-updating (IBBU) as a tibble. Both wide
(\code{wide = TRUE}) or long format (\code{wide = FALSE}) can be chosen as output. By default all post-warmup draws are
returned, but if \code{summarize = TRUE} then just the mean of each parameter is returned instead.
}
\details{
By default, the category means and scatter matrices are nested, rather than each of their elements being
stored separately (\code{nest = TRUE}).
}
\seealso{
\code{\link{get_parameters}}, \code{\link{get_stanfit}}
}
