% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIW.R
\name{get_expected_category_statistic}
\alias{get_expected_category_statistic}
\alias{get_expected_mu}
\alias{get_expected_sigma}
\title{Get expected category mean mu or covariance matrix sigma}
\usage{
get_expected_category_statistic(
  x,
  category = NULL,
  group = NULL,
  statistic = c("mu", "Sigma")
)

get_expected_mu(x, category, group)

get_expected_sigma(x, category, group)
}
\arguments{
\item{x}{An mv_ibbu_stanfit or mv_ibbu_MCMC object.}

\item{category}{Character vector with categories (or category) for which category statistics are to be
returned.  If `NULL` then all categories are included. (default: `NULL`)}

\item{group}{Character vector with groups (or group) for which category statistics are to be
returned. If `NULL` then all groups are included. (default: `NULL`)}

\item{statistic}{Which category statistic should be returned? `mu` for category mean or `Sigma` for category
covariance matrix, or `c("mu", "Sigma")` for both. (default: both)}
}
\description{
Returns the expected value of posterior marginal distribution over category means mu or
category covariance matrix Sigma, marginalized over all MCMC samples.
}
\details{
Each MCMC samples' expected value \code{E[mu] = M_n}
(i.e, the posterior/updated mean of the mulativariate Normal over category means \code{mu}).
Marginalizing across all MCMC samples (representing uncertainty in the true value of
\code{M_n}), we get \code{E\[E\[mu\]\] = mean(M_n)}.

Each MCMC samples' expected value
\code{E[Sigma] = S_n / (nu_n - D - 1)}, where \code{S_n} is the posterior/updated scatter matrix,
\code{nu_n} is the posterior/updated pseudocount representing the strength of the posterior/updated
beliefs over category covariance matrices sigma (i.e., the inverse-Wishart), and \code{D} is
the dimension of the multivariate Normal. Marginalizing across all MCMC samples
(representing uncertainty in the true value of \code{S_n}), we get
\code{E[E[Sigma]] = mean(S_n / (nu_n - D - 1))}.
}
\examples{
TBD
}
\references{
Murphy, K. P. (2012). Machine learning: a probabilistic perspective. MIT press.
}
\seealso{
TBD
}
\keyword{TBD}
