% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-class.R
\name{get_expected_columns_for_model}
\alias{get_expected_columns_for_model}
\title{Deprecated: get_expected_columns_for_model}
\usage{
get_expected_columns_for_model()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_expected_columns_for_model()} was deprecated in MVBeliefUpdatr 0.1.0 and will be removed in 0.2.0.
}
\keyword{internal}
