% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIX-NIW-basics.R
\name{get_expected_mu_from_m}
\alias{get_expected_mu_from_m}
\alias{get_m_from_expected_mu}
\title{Deprecated: get_expected_mu_from_m}
\usage{
get_expected_mu_from_m(m)

get_m_from_expected_mu(mu)
}
\arguments{
\item{m}{Mean of means m.}

\item{mu}{Expected category mean \eqn{\mu}.}
}
\value{
Expected category mean \eqn{\mu = m}.

Mean of means \eqn{m = \mu}.
}
\description{
See Murphy (2012, p. 134).
}
