% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIW.R
\name{get_expected_mu_from_m}
\alias{get_expected_mu_from_m}
\alias{get_m_from_expected_mu}
\title{Get expected category mu from m}
\usage{
get_expected_mu_from_m(m)

get_m_from_expected_mu(mu)
}
\arguments{
\item{m}{Mean of means m.}

\item{mu}{expected category mean mu.}
}
\description{
See Murphy (2012, p. 134).
}
