% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIW.R
\name{get_expected_Sigma_from_S}
\alias{get_expected_Sigma_from_S}
\title{Get expected category covariance from Scatter matrix S and pseudocount nu}
\usage{
get_expected_Sigma_from_S(S, nu)
}
\description{
Get expected category covariance from Scatter matrix S and pseudocount nu
}
\seealso{
\code{\link{get_S_from_Sigma}}
}
