% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R, R/S7-stanfit-methods.R
\name{get_exposure_category_statistic}
\alias{get_exposure_category_statistic}
\alias{get_exposure_category_mean}
\alias{get_exposure_category_css}
\alias{get_exposure_category_uss}
\alias{get_exposure_category_cov}
\alias{get_exposure_category_statistic,MVBU_Stanfit-method}
\alias{get_exposure_category_mean,MVBU_Stanfit-method}
\alias{get_exposure_category_css,MVBU_Stanfit-method}
\alias{get_exposure_category_uss,MVBU_Stanfit-method}
\alias{get_exposure_category_cov,MVBU_Stanfit-method}
\title{Get exposure category statistics from an S7 stanfit or stanfit-input object}
\usage{
get_exposure_category_statistic(x, ...)

get_exposure_category_mean(x, ...)

get_exposure_category_css(x, ...)

get_exposure_category_uss(x, ...)

get_exposure_category_cov(x, ...)

## S7 method for class <MVBU_Stanfit>
get_exposure_category_statistic(
  x,
  categories = get_category_labels(x),
  groups = get_group_labels(x, include_prior = FALSE),
  statistic = c("n", "mean", "css", "uss", "cov"),
  untransform_cues = FALSE,
  ...
)

## S7 method for class <MVBU_Stanfit>
get_exposure_category_mean(x, ...)

## S7 method for class <MVBU_Stanfit>
get_exposure_category_css(x, ...)

## S7 method for class <MVBU_Stanfit>
get_exposure_category_uss(x, ...)

## S7 method for class <MVBU_Stanfit>
get_exposure_category_cov(x, ...)
}
\arguments{
\item{x}{An S7 stanfit or stanfit-input object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A tibble, vector, or matrix of exposure category statistics.
}
\description{
Get exposure category statistics from an S7 stanfit or stanfit-input object
}
\seealso{
\code{\link{get_exposure_category_mean}}, \code{\link{get_exposure_category_cov}}
}
