% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{get_exposure_category_statistic}
\alias{get_exposure_category_statistic}
\alias{get_exposure_category_mean}
\alias{get_exposure_category_css}
\alias{get_exposure_category_uss}
\alias{get_exposure_category_cov}
\alias{get_exposure_category_statistic.NIW_ideal_adaptor_stanfit}
\alias{get_exposure_category_mean.NIW_ideal_adaptor_stanfit}
\alias{get_exposure_category_css.NIW_ideal_adaptor_stanfit}
\alias{get_exposure_category_uss.NIW_ideal_adaptor_stanfit}
\alias{get_exposure_category_cov.NIW_ideal_adaptor_stanfit}
\title{Get category sample mean or covariance matrix of exposure data from NIW ideal adaptor stanfit}
\usage{
get_exposure_category_statistic(x, ...)

get_exposure_category_mean(x, ...)

get_exposure_category_css(x, ...)

get_exposure_category_uss(x, ...)

get_exposure_category_cov(x, ...)

\method{get_exposure_category_statistic}{NIW_ideal_adaptor_stanfit}(
  x,
  categories = get_category_levels(x),
  groups = get_group_levels(x, include_prior = FALSE),
  statistic = c("n", "mean", "css", "uss", "cov"),
  untransform_cues = FALSE,
  ...
)

\method{get_exposure_category_mean}{NIW_ideal_adaptor_stanfit}(...)

\method{get_exposure_category_css}{NIW_ideal_adaptor_stanfit}(...)

\method{get_exposure_category_uss}{NIW_ideal_adaptor_stanfit}(...)

\method{get_exposure_category_cov}{NIW_ideal_adaptor_stanfit}(...)
}
\arguments{
\item{x}{An \code{\link{NIW_ideal_adaptor_stanfit}}.}

\item{...}{additional arguments to \code{\link{get_staninput}}.}

\item{categories}{Character vector with categories for which category statistics are to be
returned. (default: all categories)}

\item{groups}{Character vector with groups for which category statistics are to be
returned. (default: all groups except `"prior"`)}

\item{statistic}{Which exposure statistic should be returned? `n` for number of observations, `mean` for
category mean, `css` or `uss` for centered or uncentered category sum-of-square matrix, `cov` for the
category covariance matrix, or a character vector with any combination thereof. (default: all)}

\item{untransform_cues}{DEPRECATED. Should statistics be transformed back into the original cue space? (default: `FALSE`)}
}
\value{
If just one group and category was requested, a vector (for the mean) or matrix (for the covariance
matrix). If more than one group or category was requested, a tibble with one row for each unique combination
of group and category.
}
\description{
Returns the category means mu and/or category covariance matrix Sigma for the exposure data for an
\code{\link{NIW_ideal_adaptor_stanfit}}.
}
\seealso{
TBD
}
\keyword{TBD}
