% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get-info-from-NIW-IA-stanfit.R
\name{get_group_levels}
\alias{get_group_levels}
\title{Get group levels from model (deprecated)}
\usage{
get_group_levels(x, indices = NULL, include_prior = FALSE, ...)
}
\arguments{
\item{x}{A model object.}

\item{indices}{Optional indices.}

\item{include_prior}{Whether to include \code{"prior"}.}

\item{...}{Additional arguments.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_group_levels()} is deprecated in favor of
\code{\link{get_group_labels}}.
}
