% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-information-from-stanfit.R
\name{get_ibbu_input}
\alias{get_ibbu_input}
\title{Get the input data from an MV IBBU stanfit object.}
\usage{
get_ibbu_input(x)
}
\arguments{
\item{x}{An mv_ibbu_stanfit object.}
}
\value{
A list with element names and structure determined by the type of MV IBBU model.
}
\description{
Returns the inputs handed to \code{stan} or \code{sampling} during the creation of the \code{stanfit}
object.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
