% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-stanfit.R
\name{get_ibbu_stanfit_input}
\alias{get_ibbu_stanfit_input}
\title{Get the input data from an NIW IBBU stanfit.}
\usage{
get_ibbu_stanfit_input(fit)
}
\arguments{
\item{fit}{\code{\link{NIW_ibbu_stanfit}} object.}
}
\value{
A list with element names and structures determined by the type of stanfit model.
}
\description{
Returns the inputs handed to \code{stan} or \code{sampling} during the creation of the \code{stanfit}
object.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
