% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-model.R
\name{get_lapse_biases_from_model}
\alias{get_lapse_biases_from_model}
\title{Get lapse bias from model}
\usage{
get_lapse_biases_from_model(model, categories = model$category)
}
\arguments{
\item{model}{A model object.}

\item{categories}{A vector of category values.}
}
\value{
A vector of lapse bias values of the same length as \code{categories}.
}
\description{
Get lapse bias from model
}
