% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-model.R
\name{get_lapse_rate_from_model}
\alias{get_lapse_rate_from_model}
\title{Get lapse rate from model}
\usage{
get_lapse_rate_from_model(model)
}
\arguments{
\item{model}{A model object.

Legacy wrapper for get_lapse_rate.}
}
\description{
Deprecated. Use \code{\link{get_lapse_rate}} instead.

Deprecated. Use get_lapse_rate() instead.
}
\keyword{internal}
