% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get-info-from-model.R
\name{get_lapse_rate_from_model}
\alias{get_lapse_rate_from_model}
\title{Get lapse rate from model (deprecated)}
\usage{
get_lapse_rate_from_model(model)
}
\arguments{
\item{model}{A model object.}
}
\value{
Numeric lapse rate.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_lapse_rate_from_model()} is deprecated. Use \code{get_lapse_rate()} instead.
}
\keyword{internal}
