% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get-info-from-exemplar-model.R
\name{get_likelihood_from_exemplars}
\alias{get_likelihood_from_exemplars}
\title{Get likelihood from exemplar model (deprecated)}
\usage{
get_likelihood_from_exemplars(
  x,
  model,
  noise_treatment = "no_noise",
  log = TRUE,
  category = "category",
  category.label = NULL
)
}
\arguments{
\item{x}{Observations.}

\item{model}{Exemplar model.}

\item{noise_treatment}{Noise treatment.}

\item{log}{Logical; whether log likelihood is returned.}

\item{category}{Category column name.}

\item{category.label}{Category labels.}
}
\value{
Likelihood data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_likelihood_from_exemplars()} is deprecated. Use \code{\link{likelihood}} instead.
}
