% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-basics.R
\name{get_limits}
\alias{get_limits}
\title{Get suitable limits for coordinate system based on the MCMC samples of a variable.}
\usage{
get_limits(data, measure, hdi.prob = 0.99, min = NULL, max = NULL)
}
\arguments{
\item{data}{A `tibble`` or `data.frame` that contains `measure`.}

\item{measure}{Name of variable in `data` for which limits are sought.}

\item{hdi.prob}{Proportion of MCMC samples that are within the limits.}

\item{min, max}{If min or max are specified, then those limits are returned instead of the HDI-based limits.}
}
\value{
Vector with two values.
}
\description{
Useful for, for example, plotting of distribution.
}
\examples{
TBD

}
\seealso{
TBD
}
\keyword{TBD}
