% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-fit.R
\name{get_limits}
\alias{get_limits}
\title{Get suitable limits for coordinate system.}
\usage{
get_limits(data, measure, hdi.prob = 0.99)
}
\description{
Useful for, for example, plotting of distribution.
}
