% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-classes.R, R/s7-foundation.R
\name{get_model_family_registration}
\alias{get_model_family_registration}
\title{Get class registration for an MVBU model family}
\usage{
get_model_family_registration(family)

get_model_family_registration(family)
}
\description{
Get class registration for an MVBU model family

Get class registration for an MVBU model family
}
\keyword{internal}
