% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R, R/S7-stanfit-methods.R
\name{get_model_type}
\alias{get_model_type}
\alias{get_model_type,MVBU_Stanfit-method}
\title{Get the model type of a cognitive model or model distribution}
\usage{
get_model_type(x, ...)

## S7 method for class <MVBU_Stanfit>
get_model_type(x)
}
\arguments{
\item{x}{A model object.}
}
\value{
A character scalar identifying the model family/type.
}
\description{
Get the model type of a cognitive model or model distribution
}
