% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-MVG-IO.R
\name{get_MVG_likelihood}
\alias{get_MVG_likelihood}
\alias{get_likelihood_from_MVG}
\title{Get likelihood}
\usage{
get_MVG_likelihood(
  x,
  mu,
  Sigma,
  log = T,
  noise_treatment = "no_noise",
  Sigma_noise = NULL
)

get_likelihood_from_MVG(
  x,
  model,
  log = T,
  noise_treatment = if (is.MVG_ideal_adaptor(model)) "marginalize" else "no_noise",
  category = "category",
  category.label = NULL,
  wide = FALSE
)
}
\arguments{
\item{x}{Observations. Can be a vector with k elements for a single observation or a matrix with k
columns and n rows, in which case each row of the matrix is taken to be one observation. If x is a
tibble with k columns or a list of vectors of length k, it is reduced into a matrix with k columns.}

\item{mu}{The category mean mu. Should be a
matrix or vector of length k.}

\item{Sigma}{The category covariance matrix Sigma. Should be a square k x k matrix.}

\item{log}{Should the log-transformed density be returned (`TRUE`)? (default: `TRUE`)}

\item{noise_treatment}{Determines whether and how multivariate Gaussian noise is added to the input. Can be "no_noise", "sample"
or "marginalize". If "no_noise", no noise will be applied. If "sample" or "marginalize", `Sigma_noise` must be a covariance
matrix of appropriate dimensions. If "sample", observations are adjusted by samples drawn from the noise distribution before applying
categorization. If "marginalize" then each observation is transformed into the marginal distribution
that results from convolving the input with noise. This latter option might be helpful, for example, if one is
interested in estimating the consequences of noise across individuals. (default: "no_noise").}

\item{Sigma_noise}{Optionally, a covariance matrix describing the perceptual noise to be applied while
calculating the posterior predictive. (default: `NULL`)}
}
\description{
Get likelihood of observation(s) x given the MVG parameters mu and Sigma. This is the density of
a multivariate normal distribution over k dimensions.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
