% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIW.R, R/get-info-from-NIW-IA-stanfit.R,
%   R/get-info-from-NIW-IA.R
\name{get_NIW_categorization_function}
\alias{get_NIW_categorization_function}
\alias{get_categorization_function_from_grouped_ibbu_stanfit_draws}
\alias{get_categorization_function_from_NIW_ideal_adaptor}
\title{Get NIW categorization function}
\usage{
get_NIW_categorization_function(
  ms,
  Ss,
  kappas,
  nus,
  priors = rep(1/length(ms), length(ms)),
  lapse_rate = NULL,
  lapse_biases = rep(1/length(ms), length(ms)),
  Sigma_noise = matrix(0, nrow = if (is.null(dim(Ss[[1]]))) 1 else max(dim(Ss[[1]])),
    ncol = if (is.null(dim(Ss[[1]]))) 1 else max(dim(Ss[[1]]))),
  noise_treatment = if (is.null(Sigma_noise)) "no_noise" else "marginalize",
  logit = FALSE
)

get_categorization_function_from_grouped_ibbu_stanfit_draws(fit, ...)

get_categorization_function_from_NIW_ideal_adaptor(model, ...)
}
\arguments{
\item{ms}{Means of the multivariate normal distributions over category means.}

\item{Ss}{Scatter matrices of the inverse Wishart distribution over category covariance matrices.}

\item{kappas}{Strength of the beliefs into the distribution over category means.}

\item{nus}{Strength of the beliefs into the distribution over category covariance matrices.}

\item{priors}{Vector of categories' prior probabilities. (default: uniform prior over categories)}

\item{lapse_rate}{A lapse rate for the categorization responses.}

\item{lapse_biases}{A lapse bias for the categorization responses. (default: uniform bias over categories)}

\item{Sigma_noise}{A noise matrix. (default: a 0-matrix)}

\item{noise_treatment}{How should the noise specified in \code{Sigma_noise} be considered in the categorization function?
For details, see \code{\link{get_NIW_posterior_predictive}}.}

\item{logit}{Should the function that is returned return log-odds (TRUE) or probabilities (FALSE)? (default: TRUE)}
}
\value{
A function that takes as input cue values and returns posterior probabilities of the first category,
based on the posterior predictive of the cues given the (IBBU-derived parameters for the) categories' m, S,
kappa, nu, and prior, as well as the lapse rate.
}
\description{
Returns a categorization function for the first category, based on a set of parameters for the Normal-Inverse-Wishart (NIW)
distribution. ms, Ss, kappas, nus, and priors are assumed to be of the same length and sorted the same way, so that the first
element of ms is corresponding to the same category as the first element of Ss, kappas, nus, and priors, etc.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
