% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get-info-from-NIW-belief.R
\name{get_NIW_posterior_predictive.pmap}
\alias{get_NIW_posterior_predictive.pmap}
\title{Deprecated: get_NIW_posterior_predictive.pmap}
\usage{
get_NIW_posterior_predictive.pmap(x, m, S, kappa, nu, ...)
}
\arguments{
\item{x}{Observation(s). Can be a vector with k elements, a matrix with k columns, a data frame,
or a list of numeric vectors.}

\item{m}{Mean vector of length k.}

\item{S}{Scatter matrix of dimension k x k.}

\item{kappa}{Strength of belief on the mean (pseudocounts).}

\item{nu}{Strength of belief on the covariance matrix (pseudocounts).}

\item{...}{Additional arguments.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_NIW_posterior_predictive.pmap()} is deprecated. Use \code{\link{get_NIW_posterior_predictive}} or \code{\link{likelihood}} instead.
}
