% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIX-NIW-basics.R
\name{get_NIW_posterior_predictive}
\alias{get_NIW_posterior_predictive}
\title{Deprecated: get_NIW_posterior_predictive}
\usage{
get_NIW_posterior_predictive(
  x,
  m,
  S,
  kappa,
  nu,
  Sigma_noise = NULL,
  noise_treatment = .infer_noise_treatment(Sigma_noise),
  log = TRUE
)
}
\arguments{
\item{x}{Observation(s). Can be a vector with k elements, a matrix with k columns, a data frame,
or a list of numeric vectors.}

\item{m}{Mean vector of length k.}

\item{S}{Scatter matrix of dimension k x k.}

\item{kappa}{Strength of belief on the mean (pseudocounts).}

\item{nu}{Strength of belief on the covariance matrix (pseudocounts).}

\item{Sigma_noise}{Optional perceptual noise covariance matrix (default: \code{NULL}).}

\item{noise_treatment}{Noise treatment: \code{"no_noise"}, \code{"sample"}, or \code{"marginalize"}.}

\item{log}{Logical; if \code{TRUE}, return log-transformed density (default: \code{TRUE}).}
}
\description{
Get posterior predictive density of observations x given the Normal-Inverse-Wishart (NIW)
parameters m, S, kappa, and nu. This is the density of a multivariate Student-T distribution
\insertCite{@see @murphy2012 p. 134}{MVBeliefUpdatr}.
}
\references{
\insertRef{murphy2012}{MVBeliefUpdatr}
}
