% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIX-NIW-basics.R
\name{get_NIX_posterior_predictive}
\alias{get_NIX_posterior_predictive}
\title{Get posterior predictive for NIX distribution}
\usage{
get_NIX_posterior_predictive(
  x,
  m,
  sigma2,
  kappa,
  nu,
  Sigma_noise = NULL,
  noise_treatment = .infer_noise_treatment(Sigma_noise),
  log = TRUE
)
}
\arguments{
\item{x}{Observation(s) (numeric vector, matrix with 1 column, or 1D data frame).}

\item{m}{Prior mean location.}

\item{sigma2}{Prior expected variance parameter (or variance scale \eqn{\sigma_0^2}).}

\item{kappa}{Strength of belief on the mean (pseudocounts).}

\item{nu}{Degrees of freedom / strength of belief on variance (pseudocounts).}

\item{Sigma_noise}{Optional 1x1 noise covariance / variance (default: \code{NULL}).}

\item{noise_treatment}{Noise treatment: \code{"no_noise"}, \code{"sample"}, or \code{"marginalize"}.}

\item{log}{Logical; if \code{TRUE}, return log-transformed density (default: \code{TRUE}).}
}
\description{
Get posterior predictive density of 1D observations x given the Normal-Inverse-Chi-Squared (NIX)
parameters m, sigma2 (or S), kappa, and nu. In 1D, this evaluates a Student-T density.
}
