% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get-info-from-model.R
\name{get_nlevels_of_category_labels_from_model}
\alias{get_nlevels_of_category_labels_from_model}
\title{Deprecated: get_nlevels_of_category_labels_from_model}
\usage{
get_nlevels_of_category_labels_from_model(x)
}
\arguments{
\item{x}{A model, template, or representation.}
}
\value{
Number of categories as an integer.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_nlevels_of_category_labels_from_model()} is deprecated. Use \code{length(get_category_labels())} instead.
}
\keyword{internal}
