% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-model.R
\name{get_nlevels_of_category_labels_from_model}
\alias{get_nlevels_of_category_labels_from_model}
\title{Get number of categories from likelihood or model}
\usage{
get_nlevels_of_category_labels_from_model(x)
}
\arguments{
\item{x}{A likelihood or model object.}
}
\description{
Get the number of unique category labels from a likelihood (e.g., MVG or NIW_belief) or model (e.g., an MVG ideal observer
or NIW ideal adaptor) object.
}
