% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-information-from-stanfit.R
\name{get_number_of_draws}
\alias{get_number_of_draws}
\title{Get number of post-warmup MCMC samples from stanfit}
\usage{
get_number_of_draws(fit)
}
\description{
Get the total number of post-warmup MCMC samples in `fit`.
}
