% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{get_number_of_draws}
\alias{get_number_of_draws}
\title{Get number of post-warmup MCMC samples from stanfit}
\usage{
get_number_of_draws(fit)
}
\arguments{
\item{fit}{A \code{\link{stanfit}} object.}
}
\description{
Get the total number of post-warmup MCMC samples in \code{fit}.
}
