% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get-info-from-NIW-IA-stanfit.R
\name{get_params}
\alias{get_params}
\title{Deprecated: get_params}
\usage{
get_params(fit, original_pars = FALSE)
}
\arguments{
\item{fit}{A stanfit or MVBU_Stanfit object.}

\item{original_pars}{Whether to return original parameter names.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_params()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0. Please use \code{\link[=get_parameter_names]{get_parameter_names()}} instead.
}
\seealso{
\code{\link[=get_parameter_names]{get_parameter_names()}}
}
\keyword{internal}
