% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{get_params}
\alias{get_params}
\title{Get parameter names of stanfit}
\usage{
get_params(fit)
}
\arguments{
\item{fit}{A \code{\link{stanfit}} object.}
}
\description{
Get the names for all parameters in `fit`.
}
