% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get-info-from-model.R
\name{get_perceptual_noise_from_model}
\alias{get_perceptual_noise_from_model}
\title{Deprecated: get_perceptual_noise_from_model}
\usage{
get_perceptual_noise_from_model(model)
}
\arguments{
\item{model}{A model object.}
}
\value{
Perceptual noise covariance matrix.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_perceptual_noise_from_model()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0. Please use \code{\link[=get_noise]{get_noise()}} instead.
}
\seealso{
\code{\link[=get_noise]{get_noise()}}
}
\keyword{internal}
