% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-model.R
\name{get_perceptual_noise_from_model}
\alias{get_perceptual_noise_from_model}
\title{Get noise covariance matrix from model}
\usage{
get_perceptual_noise_from_model(model)
}
\arguments{
\item{model}{A model object.}
}
\description{
Get noise covariance matrix from model
}
