% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-basics.R
\name{get_plot_limits}
\alias{get_plot_limits}
\title{Get plot limits.}
\usage{
get_plot_limits(plot)
}
\arguments{
\item{plot}{A `ggplot` object.}
}
\value{
List with elements x and y, each of which is a vector of two values..
}
\description{
Get x and y limits from a ggplot.
}
\examples{
TBD

}
\seealso{
TBD
}
\keyword{TBD}
