% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get-info-from-MVG-IO.R,
%   R/deprecated-get-info-from-model.R
\name{get_posterior_from_MVG_ideal_observer}
\alias{get_posterior_from_MVG_ideal_observer}
\alias{get_posterior_from_model}
\title{Deprecated: get_posterior_from_MVG_ideal_observer}
\usage{
get_posterior_from_MVG_ideal_observer(
  x,
  model,
  decision_rule = "sampling",
  noise_treatment = if (decision_rule == "sampling") "sample" else "no_noise",
  lapse_treatment = if (decision_rule == "sampling") "sample" else "marginalize"
)

get_posterior_from_model(model, ...)
}
\arguments{
\item{x}{Observations.}

\item{model}{A model object.}

\item{decision_rule}{Decision rule.}

\item{noise_treatment}{Noise treatment.}

\item{lapse_treatment}{Lapse treatment.}

\item{...}{Arguments passed to \code{\link{posterior}}.}
}
\value{
Posterior data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_posterior_from_MVG_ideal_observer()} is deprecated. Use \code{\link{posterior}} instead.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_posterior_from_model()} is deprecated. Use \code{posterior()} instead.
}
\keyword{internal}
