% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-MVG-IO.R,
%   R/get-info-from-model.R
\name{get_posterior_from_MVG_ideal_observer}
\alias{get_posterior_from_MVG_ideal_observer}
\alias{get_posterior_from_model}
\title{Legacy wrapper for posterior.}
\usage{
get_posterior_from_MVG_ideal_observer(
  x,
  model,
  noise_treatment = if (decision_rule == "sampling") "sample" else
    infer_default_noise_treatment(model$Sigma_noise),
  lapse_treatment = if (decision_rule == "sampling") "sample" else "marginalize"
)

get_posterior_from_model(model, ...)
}
\description{
Deprecated. Use \code{\link{posterior}} instead.

Deprecated. Use \code{\link{posterior}} instead.
}
