% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIW.R, R/get-information-from-NIW-belief.R
\name{get_posterior_predictive}
\alias{get_posterior_predictive}
\alias{get_posterior_predictive.pmap}
\alias{get_posterior_predictive_from_NIW_belief}
\title{Get posterior predictive}
\usage{
get_posterior_predictive(x, M, S, kappa, nu, log = T)

get_posterior_predictive.pmap(x, M, S, kappa, nu, ...)

get_posterior_predictive_from_NIW_belief(
  x,
  belief,
  log = T,
  category = "category",
  category.label = NULL,
  wide = FALSE
)
}
\arguments{
\item{x}{Observations. Can be a vector with k elements for a single observation or a matrix with k
columns and n rows, in which case each row of the matrix is taken to be one observation. If x is a
tibble with k columns or a list of vectors of length k, it is reduced into a matrix with k columns.}

\item{M}{The mean of the multivariate Normal distribution of the category mean mu. Should be a
matrix or vector of length k.}

\item{S}{The scatter matrix of the inverse-Wishart distribution over the category covariance
matrix Sigma. Should be a square k x k matrix.}

\item{kappa}{The strength of the beliefs over the category mean (pseudocounts).}

\item{nu}{The strength of the beliefs over the category covariance matrix (pseudocounts).}

\item{log}{Should the log-transformed density be returned (`TRUE`)? (default: `TRUE`)}
}
\description{
Get posterior predictive of observations x given the NIW parameters M, S, kappa, and nu. This is
a multivariate Student-T distribution (Murphy, 2012, p. 135).
}
\examples{
TBD
}
\references{
Murphy, K. P. (2012). Machine learning: a probabilistic perspective. MIT press.
}
\seealso{
TBD
}
\keyword{TBD}
