% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-model.R
\name{get_priors_from_model}
\alias{get_priors_from_model}
\title{Get priors from model}
\usage{
get_priors_from_model(model, categories = model$category)
}
\arguments{
\item{model}{A model object.}

\item{categories}{A vector of category values.}
}
\value{
A vector of prior values of the same length as \code{categories}.

Legacy wrapper for get_category_prior.
}
\description{
Deprecated. Use \code{\link{get_category_prior}} instead.
}
