% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-stanfit.R
\name{get_random_draw_indices}
\alias{get_random_draw_indices}
\title{Get indices for random MCMC draws from stanfit}
\usage{
get_random_draw_indices(fit, n.draws)
}
\arguments{
\item{fit}{mv_ibbu_stanfit object.}

\item{n.draws}{Number of indices to be returned. Can't be larger than total number of
post-warmup samples across all MCMC chains in `fit`.}
}
\value{
A numeric vector.
}
\description{
Returns `n.draws` indices for random post-warmup MCMC draws (without replacement) from
`fit`.
}
