% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-core-classes.R
\name{get_registered_model_families}
\alias{get_registered_model_families}
\title{Get registered model families in MVBU.}
\usage{
get_registered_model_families()
}
\value{
Character vector of registered family names.
}
\description{
Get registered model families in MVBU.
}
