% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-classes.R, R/s7-foundation.R
\name{get_registered_model_families}
\alias{get_registered_model_families}
\title{Get registered MVBU model families}
\usage{
get_registered_model_families()

get_registered_model_families()
}
\description{
Get registered MVBU model families

Get registered MVBU model families
}
\keyword{internal}
