% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-core-classes.R
\name{get_registered_model_families}
\alias{get_registered_model_families}
\title{Get the registered MVBU model families.}
\usage{
get_registered_model_families()
}
\value{
A character vector of registered family names.
}
\description{
Get the registered MVBU model families.
}
