% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R
\name{get_representation_type}
\alias{get_representation_type}
\title{Get the representation type of a category representation}
\usage{
get_representation_type(x, ...)
}
\arguments{
\item{x}{A representation object.}
}
\value{
A character scalar identifying the representation family/type.
}
\description{
Get the representation type of a category representation
}
