% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R
\name{get_stanfit}
\alias{get_stanfit}
\title{Get the stored Stan fit from an S7 fit object.}
\usage{
get_stanfit(x, ...)
}
\arguments{
\item{x}{An S7 fit object.}
}
\value{
The stored \link[rstan:stanfit]{rstan::stanfit} object when present.
}
\description{
Get the stored Stan fit from an S7 fit object.
}
