% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{get_stanfit}
\alias{get_stanfit}
\alias{get_stanfit,ideal_adaptor_stanfit-method}
\title{Get the stanfit from an ideal adaptor stanfit}
\usage{
get_stanfit(x, ...)

\S4method{get_stanfit}{ideal_adaptor_stanfit}(x)
}
\arguments{
\item{x}{\code{\link{ideal_adaptor_stanfit}} object.}
}
\value{
An \code{\link{rstan::stanfit}} object.
}
\description{
Returns the stanfit created by \code{stan} during the creation of the \code{\link{ideal_adaptor_stanfit}}
object.
}
\seealso{
TBD
}
\keyword{TBD}
