% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R
\name{get_staninput}
\alias{get_staninput}
\title{Get the stored Stan input from an S7 fit or fit-input object.}
\usage{
get_staninput(x, ...)
}
\arguments{
\item{x}{An S7 fit or fit-input object.}
}
\value{
The stored S7 Stan input object.
}
\description{
Get the stored Stan input from an S7 fit or fit-input object.
}
