% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{get_staninput}
\alias{get_staninput}
\title{Get the input data from an ideal adaptor stanfit}
\usage{
get_staninput(x)
}
\arguments{
\item{x}{An S7 \code{\link{IdealAdaptorStanfit}} or \code{\link{IdealAdaptorStanfitInput}} object.}
}
\value{
The S7 staninput object stored in \code{x}.
}
\description{
Returns the inputs handed to \code{stan} or \code{sampling} during the creation of the \code{\link{ideal_adaptor_stanfit}}
object.
}
\seealso{
TBD
}
\keyword{TBD}
