% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{get_staninput}
\alias{get_staninput}
\alias{get_staninput,ideal_adaptor_stanfit-method}
\title{Get the input data from an ideal adaptor stanfit}
\usage{
get_staninput(x, which = c("untransformed", "transformed", "both")[1])

\S4method{get_staninput}{ideal_adaptor_stanfit}(x, which = c("untransformed", "transformed", "both")[1])
}
\arguments{
\item{x}{\code{\link{ideal_adaptor_stanfit}} object.}

\item{which}{Should "transformed" or "untransformed" staninput be returned or "both"? (default: \code{"untransformed"})}
}
\value{
A list with element names and structures determined by the type of stanfit model.
}
\description{
Returns the inputs handed to \code{stan} or \code{sampling} during the creation of the \code{\link{ideal_adaptor_stanfit}}
object.
}
\seealso{
TBD
}
\keyword{TBD}
