% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{get_staninput}
\alias{get_staninput}
\alias{get_staninput.NIW_ideal_adaptor_stanfit}
\title{Get the input data from an NIW ideal adaptor stanfit}
\usage{
get_staninput(x, ...)

\method{get_staninput}{NIW_ideal_adaptor_stanfit}(x)
}
\arguments{
\item{x}{\code{\link{NIW_ideal_adaptor_stanfit}} object.}
}
\value{
A list with element names and structures determined by the type of stanfit model.
}
\description{
Returns the inputs handed to \code{stan} or \code{sampling} during the creation of the \code{\link{NIW_ideal_adaptor_stanfit}}
object.
}
\seealso{
TBD
}
\keyword{TBD}
