% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-2025.R
\name{get_staninput_from_stanfit}
\alias{get_staninput_from_stanfit}
\title{Deprecated: get_staninput_from_stanfit}
\usage{
get_staninput_from_stanfit(...)
}
\description{
Deprecated. Use \code{\link{get_staninput}} instead.

Deprecated. Use get_staninput() instead.
}
\keyword{internal}
