% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{get_staninput_variable_levels}
\alias{get_staninput_variable_levels}
\alias{get_category_levels}
\alias{get_group_levels}
\alias{get_cue_levels}
\alias{get_staninput_variable_levels.IdealAdaptorStanfit}
\alias{get_category_levels.IdealAdaptorStanfit}
\alias{get_group_levels.IdealAdaptorStanfit}
\alias{get_cue_levels.IdealAdaptorStanfit}
\title{Get or restore the original group or category levels from an ideal adaptor stanfit.}
\usage{
get_staninput_variable_levels(x, ...)

get_category_levels(x, ...)

get_group_levels(x, ...)

get_cue_levels(x, ...)

\method{get_staninput_variable_levels}{IdealAdaptorStanfit}(
  x,
  variable = c("category", "group", "cue"),
  indices = NULL
)

\method{get_category_levels}{IdealAdaptorStanfit}(x, indices = NULL)

\method{get_group_levels}{IdealAdaptorStanfit}(x, indices = NULL, include_prior = F)

\method{get_cue_levels}{IdealAdaptorStanfit}(x, indices = NULL)
}
\arguments{
\item{x}{\code{\link{ideal_adaptor_stanfit}} object.}

\item{variable}{Either "category" or "group".}

\item{indeces}{A vector of category or group indices that should be turned into the original
category levels, or \code{NULL} if only the unique levels in their original order (as vector of characters)
should be returned. (default: \code{NULL})}
}
\value{
If no category or group indices are provided, the levels of the category/group are returned (in the
original order). Otherwise a vector of the same length as \code{indices}
}
\description{
Checks if information is available about the original values and order of the factor levels
for the category variable (for which beliefs about means and covariances are inferred) or
group variable (e.g., subject or exposure group), respectively. If available,
that information is returned. \code{get_category_levels()} and \code{get_group_levels()} are
convenience functions, calling \code{get_staninput_variable_levels()}.
}
\seealso{
\link[tidybayes:recover_types]{tidybayes::recover_types}, \code{\link{get_constructor}}
}
\keyword{TBD}
