% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get-info-from-NIW-IA-stanfit.R
\name{get_staninput_variable_levels}
\alias{get_staninput_variable_levels}
\title{Deprecated: get_staninput_variable_levels}
\usage{
get_staninput_variable_levels(
  x,
  variable = c("category", "group", "cue"),
  indices = NULL
)
}
\arguments{
\item{x}{A model object.}

\item{variable}{Either "category", "group", or "cue".}

\item{indices}{Optional indices.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_staninput_variable_levels()} is deprecated in favor of
\code{\link{get_category_labels}}, \code{\link{get_group_labels}}, and
\code{\link{get_cue_labels}}.
}
