% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get-info-from-NIW-IA-stanfit.R
\name{get_stanmodelname}
\alias{get_stanmodelname}
\title{Get the name of the stanmodel from an ideal adaptor stanfit (deprecated)}
\usage{
get_stanmodelname(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Additional arguments.}
}
\value{
A character string.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{get_stanmodelname()} is deprecated; use \code{\link{get_model_type}} instead.
}
