% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{get_stanmodelname}
\alias{get_stanmodelname}
\title{Get the name of the stanmodel from an ideal adaptor stanfit}
\usage{
get_stanmodelname(x, ...)
}
\arguments{
\item{x}{\code{\link{ideal_adaptor_stanfit}} object.}
}
\value{
A character string.
}
\description{
Returns the name of the stanfit model.
}
\seealso{
TBD
}
\keyword{TBD}
