% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{get_sufficient_category_statistics}
\alias{get_sufficient_category_statistics}
\title{Get sufficient statistics from a data set}
\usage{
get_sufficient_category_statistics(
  data,
  cues,
  category = "category",
  group = NULL,
  categories = NULL,
  groups = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} with the data. Each row should be an observation
of a category, containing the category label, the cue values of the
observation, and optionally grouping variables.}

\item{cues}{Names of columns with cue values.}

\item{category}{Name of column that contains the category label for the
exposure data. (default: "category")}

\item{group}{Name of column(s) that contains information about which
observations form a group. This could be individual subjects or conditions
in an experiment. (default: \code{NULL})}

\item{categories, groups}{Character vector of categories/groups to be
summarized. If \code{NULL}, all categories/groups will be included.
(default: \code{NULL})}

\item{verbose}{Logical; if \code{TRUE}, produce verbose output. (default: \code{FALSE})}

\item{...}{Additional arguments (currently unused).}
}
\value{
A \code{data.frame} of sufficient statistics for each combination of
category and group. This includes the count (\code{x_N}), mean vector
(\code{x_mean}), uncentered and centered sums-of-squares (\code{x_uss}, \code{x_css}), and
the covariance matrix (\code{x_cov}).
}
\description{
Get sufficient statistics from data. Calculates functions for the specified
cues for any combination of groups (optional) and categories, and returns
them as a data.frame. Rows with missing values for cues will be ignored in the
calculation of the sufficient statistics.
}
\keyword{TBD}
