% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{get_test_data}
\alias{get_test_data}
\alias{get_test_data.NIW_ideal_adaptor_stanfit}
\title{Get the test data from an NIW ideal adaptor stanfit.}
\usage{
get_test_data(x, ...)

\method{get_test_data}{NIW_ideal_adaptor_stanfit}(fit, groups = get_group_levels(fit, include_prior = FALSE))
}
\arguments{
\item{x}{\code{\link{NIW_ideal_adaptor_stanfit}} object.}

\item{groups}{Character vector of groups for which test data is requested. Typically, the levels of these factors
are automatically added to the fit during the creation of the fit. If necessary, however, it is possible to use
\code{\link[tidybayes]{recover_types}} on the stanfit object to add or change these levels later.
(default: all categories/groups will be selected)}
}
\value{
A \code{tibble} in which each row is a test token. Columns include the cues
and the response counts (one column per category) for all test tokens and all groups.
}
\description{
Returns the test data used during the creation of the \code{\link[rstan]{stanfit}}.
object.
}
\seealso{
TBD
}
\keyword{TBD}
