% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-stanfit.R
\name{get_transform_function_from_stanfit}
\alias{get_transform_function_from_stanfit}
\alias{get_untransform_function_from_stanfit}
\alias{get_transform_information_from_stanfit}
\title{Get the transform/untransform information from an NIW ideal adaptor stanfit.}
\usage{
get_transform_function_from_stanfit(model)

get_untransform_function_from_stanfit(model)

get_transform_information_from_stanfit(model)
}
\arguments{
\item{fit}{\code{\link{NIW_ideal_adaptor_stanfit}} object.}
}
\value{
A function.
}
\description{
Returns the transform/untransform information handed to \code{stan} or \code{sampling} during the creation of the \code{stanfit}
object.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
