% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R
\name{get_transform_information}
\alias{get_transform_information}
\title{Get transform metadata from an S7 fit or fit-input object.}
\usage{
get_transform_information(x, ...)
}
\arguments{
\item{x}{An S7 fit or fit-input object.}
}
\value{
A transform-information object.
}
\description{
Get transform metadata from an S7 fit or fit-input object.
}
