% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R, R/S7-stanfit-methods.R
\name{get_transform_information}
\alias{get_transform_information}
\alias{get_transform_information,any-method}
\alias{get_transform_information,MVBU_Stanfit-method}
\title{Get transform metadata from an S7 fit or fit-input object}
\usage{
get_transform_information(x, ...)

## S7 method for class <any>
get_transform_information(x)

## S7 method for class <MVBU_Stanfit>
get_transform_information(x)
}
\arguments{
\item{x}{An S7 fit or fit-input object.}
}
\value{
A transform-information object.
}
\description{
Get transform metadata from an S7 fit or fit-input object
}
