% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{get_transform_information}
\alias{get_transform_information}
\alias{get_transform_information.NIW_ideal_adaptor_stanfit}
\alias{get_transform_function.NIW_ideal_adaptor_stanfit}
\alias{get_untransform_function.NIW_ideal_adaptor_stanfit}
\title{Get the transform/untransform information from an NIW ideal adaptor stanfit}
\usage{
get_transform_information(fit, ...)

\method{get_transform_information}{NIW_ideal_adaptor_stanfit}(fit)

\method{get_transform_function}{NIW_ideal_adaptor_stanfit}(fit)

\method{get_untransform_function}{NIW_ideal_adaptor_stanfit}(fit)
}
\arguments{
\item{fit}{An \code{\link{NIW_ideal_adaptor_stanfit}} object.}
}
\value{
A function.
}
\description{
Returns the transform/untransform information handed to \code{stan} or \code{sampling} during the creation of the \code{stanfit}
object.
}
\seealso{
TBD
}
\keyword{TBD}
