% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-2025.R, R/deprecated.R
\name{get_transform_information_from_stanfit}
\alias{get_transform_information_from_stanfit}
\title{DEPRECATED: get_transform_information_from_stanfit}
\usage{
get_transform_information_from_stanfit(...)

get_transform_information_from_stanfit(...)
}
\description{
DEPRECATED: get_transform_information_from_stanfit

DEPRECATED: get_transform_information_from_stanfit
}
