% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-2025.R
\name{get_transform_information_from_stanfit}
\alias{get_transform_information_from_stanfit}
\title{Legacy wrapper for get_transform_information.ideal_adaptor_stanfit.}
\usage{
get_transform_information_from_stanfit(...)
}
\description{
Deprecated. Use \code{\link{get_transform_information.ideal_adaptor_stanfit}} instead.

Deprecated. Use get_transform_information.ideal_adaptor_stanfit() instead.
}
\keyword{internal}
