% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{get_transform_information_from_stanfit}
\alias{get_transform_information_from_stanfit}
\title{Deprecated}
\usage{
get_transform_information_from_stanfit(fit)
}
\description{
Deprecated
}
