% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-2025.R
\name{get_untransform_function_from_stanfit}
\alias{get_untransform_function_from_stanfit}
\title{Legacy wrapper for get_untransform_function.ideal_adaptor_stanfit.}
\usage{
get_untransform_function_from_stanfit(...)
}
\description{
Deprecated. Use \code{\link{get_untransform_function.ideal_adaptor_stanfit}} instead.
}
