% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-2025.R
\name{infer_prior_beliefs}
\alias{infer_prior_beliefs}
\title{Legacy wrapper for the old prior-belief inference workflow.}
\usage{
infer_prior_beliefs(
  exposure,
  test,
  cues,
  category,
  response,
  group,
  group.unique = NULL,
  center.observations = TRUE,
  scale.observations = FALSE,
  pca.observations = FALSE,
  pca.cutoff = 1,
  transform_type = NULL,
  lapse_rate = NULL,
  mu_0 = NULL,
  Sigma_0 = NULL,
  tau_scale = NULL,
  L_omega_eta = 1,
  split_loglik_per_observation = 0,
  stanmodel = NULL,
  verbose = FALSE,
  silent = 1,
  ...
)
}
\arguments{
\item{stanmodel}{Name of stanmodel that should be used. Overrides any default selection.}

\item{silent}{Verbosity level between \code{0} and \code{2}.
If \code{1} (the default), most of the
informational messages of compiler and sampler are suppressed.
If \code{2}, even more messages are suppressed. The actual
sampling progress is still printed. Set \code{refresh = 0} to turn this off
as well. If using \code{backend = "rstan"} you can also set
\code{open_progress = FALSE} to prevent opening additional progress bars.}

\item{...}{Additional parameters are passed to \link[rstan:sampling]{rstan::sampling}.}
}
\description{
Deprecated. Use \code{\link{infer_NIW_ideal_adaptor}} together with \code{\link{make_staninput_for_NIW_ideal_adaptor}} instead.
}
