% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-2025.R
\name{infer_prior_beliefs}
\alias{infer_prior_beliefs}
\title{Deprecated: infer_prior_beliefs}
\usage{
infer_prior_beliefs(
  exposure,
  test,
  cues,
  category,
  response,
  group,
  group.unique = NULL,
  center.observations = TRUE,
  scale.observations = FALSE,
  pca.observations = FALSE,
  pca.cutoff = 1,
  transform_type = NULL,
  lapse_rate = NULL,
  mu_0 = NULL,
  Sigma_0 = NULL,
  tau_scale = NULL,
  L_omega_eta = 1,
  split_loglik_per_observation = 0,
  stanmodel = NULL,
  verbose = FALSE,
  silent = 1,
  ...
)
}
\arguments{
\item{exposure}{A data frame or tibble with exposure data. Each row is assumed to contain one observation.}

\item{test}{A data frame or tibble with test data. Each row is assumed to contain one observation.}

\item{cues}{Names of columns with cue values. These columns must be present in both \code{exposure} and \code{test}.}

\item{category}{Name of the column in \code{exposure} that stores the category information. (default: "category")}

\item{response}{Name of the column in \code{test} that stores the response information. (default: "response")}

\item{group}{Name of the column in \code{exposure} and \code{test} that stores the grouping information. (default: "group")}

\item{group.unique}{Optional grouping column used to collapse repeated exposure conditions before constructing the Stan input.}

\item{stanmodel}{Character string naming the Stan model. Must be one of \code{NIX_ideal_adaptor}, \code{NIW_ideal_adaptor}, or \code{MNIX_ideal_adaptor}.}

\item{verbose}{Should verbose output be provided? (default: \code{FALSE})}

\item{silent}{Verbosity level between \code{0} and \code{2}.
If \code{1} (the default), most of the
informational messages of compiler and sampler are suppressed.
If \code{2}, even more messages are suppressed. The actual
sampling progress is still printed. Set \code{refresh = 0} to turn this off
as well. If using \code{backend = "rstan"} you can also set
\code{open_progress = FALSE} to prevent opening additional progress bars.}

\item{...}{Additional parameters are passed to \link[rstan:sampling]{rstan::sampling}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{infer_prior_beliefs()} was deprecated in MVBeliefUpdatr 0.1.0
and will be removed in 0.2.0. Please use \code{\link[=fit_ideal_adaptor]{fit_ideal_adaptor()}} instead.
}
\seealso{
\code{\link[=fit_ideal_adaptor]{fit_ideal_adaptor()}}
}
\keyword{internal}
