% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-stan.R
\name{infer_prior_beliefs}
\alias{infer_prior_beliefs}
\title{Infer prior beliefs based on adaptation behavior}
\usage{
infer_prior_beliefs(
  exposure,
  test,
  cues,
  category,
  response,
  group,
  group.unique,
  center.observations = TRUE,
  scale.observations = TRUE,
  pca.observations = FALSE,
  pca.cutoff = 1,
  tau_scale = 10,
  L_omega_scale = 1,
  useMultivariateUpdating = NULL,
  sample = TRUE,
  verbose = FALSE,
  model = NULL,
  ...
)
}
\arguments{
\item{exposure}{`tibble` or `data.frame` with the exposure data. Each row should be an observation of a category,
and contain information about the category label, the cue values of the observation, and optionally grouping variables.}

\item{test}{`tibble` or `data.frame` with the test data. Each row should be an observation, and contain information
about the cue values of the test stimulus and the participant's response.}

\item{cues}{Names of columns with cue values. Must exist in both exposure and test data.}

\item{category}{Name of column in exposure data that contains the category label. Can be `NULL` for unsupervised updating
(not yet implemented). (default: "category")}

\item{response}{Name of column in test data that contains participants' responses. (default: "response")}

\item{group}{Name of column that contains information about which observations form a group. Typically, this is
a variable identifying subjects/participants. Must exist in both exposure and test data. (default: "group")}

\item{group.unique}{Name of column that uniquely identifies each group with identical exposure. This could be a
variable indicating the different conditions in an experiment. Using group.unique is optional, but can be
substantially more efficient if many groups share the same exposure.}

\item{center.observations}{Should the data be centered? (default: `TRUE`)}

\item{scale.observations}{Should the data be standardized? (default: `TRUE`)}

\item{pca.observations}{Should the data be transformed into orthogonal principal components? (default: `FALSE`)}

\item{pca.cutoff}{Determines which principal components are handed to the MVBeliefUpdatr Stan program: all
components necessary to explain at least the pca.cutoff of the total variance. (default: .95) Ignored if
`pca.observation = FALSE`. (default: 1)}

\item{tau_scale}{Scale for the Cauchy prior for standard deviations of the covariance matrix of mu_0 and
scale for the LKJ prior for the correlations of the covariance matrix of mu_0. Set to 0 to ignore. (default: 0)}

\item{L_omega_scale}{Scale for the Cauchy prior for standard deviations of the covariance matrix of mu_0 and
scale for the LKJ prior for the correlations of the covariance matrix of mu_0. Set to 0 to ignore. (default: 0)}

\item{sample}{Should the model be fit and sampled from?}

\item{...}{Additional parameters are passed to \code{\link[rstan]{sampling}}}
}
\value{
A \code{mvg_ibbu_stanfit} object with the fitted stan model.
}
\description{
This function takes exposure and test data, the names of the cues, category, and response columns (and
optionally group and/or block columns), and uses
stan to draw samples from the prior beliefs that match the behavior.
}
\seealso{
\code{\link{is.mvg_ibbu_stanfit}} for information about mvg_ibbu_stanfit objects,
\code{\link{add_ibbu_stanfit_draws}} to draw samples from the stanfit.
}
