% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-stan.R
\name{infer_prior_beliefs}
\alias{infer_prior_beliefs}
\title{Infer prior beliefs based on adaptation behavior}
\usage{
infer_prior_beliefs(
  exposure,
  test,
  cues,
  category,
  response,
  group,
  group.unique = NULL,
  center.observations = TRUE,
  scale.observations = TRUE,
  pca.observations = FALSE,
  pca.cutoff = 1,
  m_0 = NULL,
  S_0 = NULL,
  tau_scale = 0,
  L_omega_scale = 0,
  data_list = NULL,
  transform_information = NULL,
  sample = TRUE,
  file = NULL,
  model = NULL,
  use_multivariate_updating = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{exposure}{`tibble` or `data.frame` with the exposure data. Each row should be an observation of a category,
and contain information about the category label, the cue values of the observation, and optionally grouping variables.}

\item{test}{`tibble` or `data.frame` with the test data. Each row should be an observation, and contain information
about the cue values of the test stimulus and the participant's response.}

\item{cues}{Names of columns with cue values. Must exist in both exposure and test data.}

\item{category}{Name of column in exposure data that contains the category label. Can be `NULL` for unsupervised updating
(not yet implemented). (default: "category")}

\item{response}{Name of column in test data that contains participants' responses. (default: "response")}

\item{group}{Name of column that contains information about which observations form a group. Typically, this is
a variable identifying subjects/participants. Must exist in both exposure and test data. (default: "group")}

\item{group.unique}{Name of column that uniquely identifies each group with identical exposure. This could be a
variable indicating the different conditions in an experiment. Using group.unique is optional, but can be
substantially more efficient if many groups share the same exposure. To ignore, set to `NULL`. (default: `NULL`)}

\item{center.observations}{Should the data be centered? Centering will not affect the inferred correlation or
covariance matrices but it will affect the absolute position of the inferred means. The relative position of
the inferred means remains unaffected. If `TRUE` and `m_0` is specified, `m_0` will also be centered (`S_0` is not
affected by centering and thus not changed). (default: `TRUE`)}

\item{scale.observations}{Should the data be standardized? Scaling will not affect the inferred correlation matrix,
but it will affect the inferred covariance matrix because it affects the inferred standard deviations. It will also
affect the absolute position of the inferred means. The relative position of the inferred means remains unaffected.
If `TRUE` and `m_0` and `S_0` are specified, `m_0` and `S_0` will also be scaled.
(default: `TRUE`)}

\item{pca.observations}{Should the data be transformed into orthogonal principal components? (default: `FALSE`)}

\item{pca.cutoff}{Determines which principal components are handed to the MVBeliefUpdatr Stan program: all
components necessary to explain at least the pca.cutoff of the total variance. (default: .95) Ignored if
`pca.observation = FALSE`. (default: 1)}

\item{m_0}{Optionally, prior means (m_0) and/or prior scatter matrices (S_0) for all categories. Each should be
a list, with
each element being the mean/scatter matrix for a specific category. Elements should be ordered in the same order as
the levels of the category variable in \code{exposure} and \code{test}. The means and scatter matrices could be
estimated, for example, from phonetically annotated speech recordings (see \code{\link{make_NIW_prior_from_data}}
for a convenient way to do so). To aspects should be kept in mind, however. First, an \emph{inferred} scatter
matrix include variability from perceptual and/or environmental noise, \emph{in addition} to the motor noise that
is reflected in production data. Second, the prior scatter matrix has an implicit nu associated with it, so that
the nu inferred by \code{\link{infer_prior_beliefs}} is best thought of as a multiple of the implicit nu used
during the creation of the scatter matrix S_0. For that reason, we recommend the use of nu = D + 2 in the call to
\code{\link{make_NIW_prior_from_data}} (the default), since the S_0 obtained that way is identical to the category
covariance matrix Sigma.}

\item{S_0}{Optionally, prior means (m_0) and/or prior scatter matrices (S_0) for all categories. Each should be
a list, with
each element being the mean/scatter matrix for a specific category. Elements should be ordered in the same order as
the levels of the category variable in \code{exposure} and \code{test}. The means and scatter matrices could be
estimated, for example, from phonetically annotated speech recordings (see \code{\link{make_NIW_prior_from_data}}
for a convenient way to do so). To aspects should be kept in mind, however. First, an \emph{inferred} scatter
matrix include variability from perceptual and/or environmental noise, \emph{in addition} to the motor noise that
is reflected in production data. Second, the prior scatter matrix has an implicit nu associated with it, so that
the nu inferred by \code{\link{infer_prior_beliefs}} is best thought of as a multiple of the implicit nu used
during the creation of the scatter matrix S_0. For that reason, we recommend the use of nu = D + 2 in the call to
\code{\link{make_NIW_prior_from_data}} (the default), since the S_0 obtained that way is identical to the category
covariance matrix Sigma.}

\item{data_list}{A list of the type that would be returned by \code{\link[compose_data]{compose_data_to_infer_prior_via_conjugate_ibbu_w_sufficient_stats}}.
This list can be provided instead of the individual inputs that are handed to \code{compose_data_to_infer_prior_via_conjugate_ibbu_w_sufficient_stats}.}

\item{transform_information}{Optionally, a list of transform paramters, transform function, and corresponding untransform
function of the type returned by \code{\link[transform_cues]{transform_cues}}. These functions
will be attached to the \code{NIW_ideal_adaptor_stanfit} object. This object will override any transform
information that is created by \code{\link[compose_data]{compose_data_to_infer_prior_via_conjugate_ibbu_w_sufficient_stats}}.}

\item{sample}{Should the model be fit and sampled from?}

\item{file}{Either NULL or a character string. In the latter case, the fitted model object is saved
via saveRDS in a file named after the string supplied in file. The .rds extension is added automatically.
If the file already exists, the model from that file will be loaded and returned instead of refitting the model.
As existing files won't be overwritten, you have to manually remove the file in order to refit and save the
model under an existing file name. The file name is stored in the \code{NIW_ideal_adaptor_stanfit} object
for later use. (default `NULL`)}

\item{model}{Name of stanmodel that should be used. Overrides any default selection.}

\item{use_multivariate_updating}{Should multivariate updating be used? By default this option will be
selecting if the relevant cues (after transformations, including PCA, if selected) have more than 1
dimension.}

\item{...}{Additional parameters are passed to \code{\link[rstan]{sampling}}}

\item{untransform_fit}{Logical flag indicating whether the MCMC samples of the model should be transformed back
into the original cue space by applying the untransform function. (default: `TRUE`)}
}
\value{
\code{NIW_ideal_adaptor_stanfit} object with the fitted stan model.
}
\description{
This function takes exposure and test data, the names of the cues, category, and response columns (and
optionally group and/or block columns), and uses
stan to draw samples from the prior beliefs that match the behavior.
}
\seealso{
\code{\link{is.NIW_ideal_adaptor_stanfit}} for information about NIW_ideal_adaptor_stanfit objects,
\code{\link{add_ibbu_stanfit_draws}} to draw samples from the stanfit.
}
