% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-class-exemplar-model.R
\name{is.exemplar_model}
\alias{is.exemplar_model}
\title{Deprecated: is.exemplar_model}
\usage{
is.exemplar_model(x, group = NULL, verbose = F, tolerance = MVBU_PROB_TOL)
}
\arguments{
\item{x}{Object to be checked.}

\item{group}{Name of one or more group variables, each unique combination of which describes an exemplar model. (default: NULL)}

\item{verbose}{Should verbose output be provided? (default: \code{TRUE})}

\item{category}{Name of the category variable. (default: "category")}
}
\value{
A logical.
}
\description{
Deprecated. Use the S7-based predicates and constructors instead.
}
\details{
Check whether \code{x} is an exemplar model. Optionally, one can also check whether a lapse rate
and lapse bias is part of the model.
}
\seealso{
TBD
}
\keyword{internal}
