% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-exemplars.R
\name{is.exemplars}
\alias{is.exemplars}
\title{Is this a set of exemplar categories?}
\usage{
is.exemplars(x, group = NULL, verbose = F)
}
\arguments{
\item{x}{Object to be checked.}

\item{group}{Name of one or more group variables, each unique combination of which describes a set of exemplars. (default: NULL)}

\item{category}{Name of the category variable. (default: "category")}
}
\value{
A logical.
}
\description{
Check whether \code{x} is a set of exemplar categories.
}
\seealso{
TBD
}
\keyword{TBD}
