% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-NIW-IA-stanfit.R
\name{is.ideal_adaptor_staninput}
\alias{is.ideal_adaptor_staninput}
\title{Is this a list of NIW ideal adaptor stanfit inputs?}
\usage{
is.ideal_adaptor_staninput(x, verbose = FALSE)
}
\arguments{
\item{x}{The input to be checked.}

\item{verbose}{Should verbose output be provided? (default: \code{FALSE})}
}
\value{
A logical.
}
\description{
Check whether \code{x} is a \code{ideal_adaptor_staninput}.
}
\seealso{
TBD
}
\keyword{TBD}
