% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{is.mv_ibbu_MCMC}
\alias{is.mv_ibbu_MCMC}
\title{Is this a tibble of MV IBBU draws?}
\usage{
is.mv_ibbu_MCMC(x, nested = T, long = T)
}
\description{
Check whether \code{x} is a tibble of post-warmup draws of parameters obtained from incremental
Bayesian belief-updating (IBBU).
}
