% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{is.MVBU_model}
\alias{is.MVBU_model}
\title{Is this an MVBeliefUpdatr model?}
\usage{
is.MVBU_model(x, group = NULL, verbose = F, tolerance = 1e-05)
}
\arguments{
\item{x}{Object to be checked.}

\item{group}{Name of one or more group variables, each unique combination of which describes a model. (default: NULL)}

\item{verbose}{Should verbose output be provided? (default: `TRUE`)}
}
\value{
A logical.
}
\description{
Check whether \code{x} is a model with lapse rates, biases, priors, and perceptual noise.
}
\seealso{
TBD
}
\keyword{TBD}
