% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-class.R
\name{is.MVBU_representation}
\alias{is.MVBU_representation}
\title{Deprecated: is.MVBU_representation}
\usage{
is.MVBU_representation(x, group = NULL, verbose = F, tolerance = MVBU_PROB_TOL)
}
\arguments{
\item{x}{Object to be checked.}

\item{group}{Name of one or more group variables, each unique combination of which describes a model. (default: NULL)}

\item{verbose}{Should verbose output be provided? (default: \code{TRUE})}

\item{tolerance}{Probability tolerance.}
}
\value{
A logical.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{is.MVBU_representation()} was deprecated in MVBeliefUpdatr 0.1.0 and will be removed in 0.2.0.
Please use S7 predicates and classes instead.
}
\seealso{
\link{MVBU_CategoryRepresentationTemplate}
}
\keyword{internal}
