% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-class.R
\name{is.MVBU_representation}
\alias{is.MVBU_representation}
\title{Deprecated: is.MVBU_representation}
\usage{
is.MVBU_representation(x, group = NULL, verbose = F, tolerance = MVBU_PROB_TOL)
}
\arguments{
\item{x}{Object to be checked.}

\item{group}{Name of one or more group variables, each unique combination of which describes a model. (default: NULL)}

\item{verbose}{Should verbose output be provided? (default: \code{TRUE})}
}
\value{
A logical.
}
\description{
Deprecated. Use the S7-based predicates and constructors instead.
}
\details{
Check whether \code{x} is recognized as an MVBeliefUpdatr category representation.
}
\seealso{
TBD
}
\keyword{internal}
