% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-MVG.R
\name{is.MVG}
\alias{is.MVG}
\title{Is this a set of multivariate Gaussian (MVG) categories?}
\usage{
is.MVG(x, group = NULL, category = "category", is.long = T, verbose = F)
}
\arguments{
\item{x}{Object to be checked.}

\item{group}{Name of one or more group variables, each unique combination of which describes an MVG. (default: NULL)}

\item{category}{Name of the category variable. (default: "category")}

\item{is.long}{Is this check assessing whether the ideal observer is in long format (`TRUE`) or wide format (`FALSE`)?
(default: `TRUE`)}
}
\value{
A logical.
}
\description{
Check whether \code{x} is a set of multivariate Gaussian (MVG) categories.
}
\seealso{
TBD
}
\keyword{TBD}
