% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-class-MVG-IO.R
\name{is.MVG_ideal_observer}
\alias{is.MVG_ideal_observer}
\title{Deprecated: is.MVG_ideal_observer}
\usage{
is.MVG_ideal_observer(
  x,
  group = NULL,
  category = "category",
  is.long = T,
  with.lapse = if (with.lapse_bias) T else F,
  with.lapse_bias = F,
  verbose = F,
  tolerance = MVBU_PROB_TOL
)
}
\arguments{
\item{x}{Object to be checked.}

\item{group}{Name of one or more group variables, each unique combination of which describes an MVG_ideal_observer. (default: NULL)}

\item{category}{Name of the category variable. (default: "category")}

\item{is.long}{Is this check assessing whether the ideal observer is in long format (\code{TRUE}) or wide format (\code{FALSE})?
(default: \code{TRUE})}

\item{with.lapse}{Does this ideal observer have a lapse rate? (default: \code{FALSE})}

\item{with.lapse_bias}{Does this ideal observer have a lapse bias? (default: \code{FALSE})}

\item{verbose}{Should verbose output be provided? (default: \code{TRUE})}

\item{tolerance}{Probability tolerance.}
}
\value{
A logical.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{is.MVG_ideal_observer()} was deprecated in MVBeliefUpdatr 0.1.0 and will be removed in 0.2.0.
Please use S7 validators and \link{MVG_IdealObserver} instead.
}
\seealso{
\link{MVG_IdealObserver}
}
\keyword{internal}
