% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-MVG-IO.R
\name{is.MVG_ideal_observer}
\alias{is.MVG_ideal_observer}
\title{Is this an ideal observer with multivariate Gaussian (MVG) categories?}
\usage{
is.MVG_ideal_observer(
  x,
  group = NULL,
  category = "category",
  is.long = T,
  with.lapse = if (with.lapse_bias) T else F,
  with.lapse_bias = F,
  verbose = F,
  tolerance = 1e-05
)
}
\arguments{
\item{x}{Object to be checked.}

\item{group}{Name of one or more group variables, each unique combination of which describes an MVG_ideal_observer. (default: NULL)}

\item{category}{Name of the category variable. (default: "category")}

\item{is.long}{Is this check assessing whether the ideal observer is in long format (`TRUE`) or wide format (`FALSE`)?
(default: `TRUE`)}

\item{with.lapse}{Does this ideal observer have a lapse rate? (default: `FALSE`)}

\item{with.lapse_bias}{Does this ideal observer have a lapse bias? (default: `FALSE`)}

\item{verbose}{Should verbose output be provided? (default: `TRUE`)}
}
\value{
A logical.
}
\description{
Check whether \code{x} is an ideal observer with \link[=is.MVG]{multivariate Gaussian (MVG) categories}. Optionally, one can also check whether a lapse rate
and lapse bias is part of the ideal observer.
}
\seealso{
TBD
}
\keyword{TBD}
