% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-MVG-IO.R
\name{is.MVG_ideal_observer}
\alias{is.MVG_ideal_observer}
\alias{is.MVG_ideal_observer_w_lapse}
\alias{is.MVG_ideal_observer_w_lapse_bias}
\title{Is this an ideal observer with multivariate Gaussian (MVG) categories?}
\usage{
is.MVG_ideal_observer(
  x,
  category = "category",
  is.long = T,
  with.lapse = if (with.bias) T else F,
  with.lapse_bias = F,
  verbose = F
)

is.MVG_ideal_observer_w_lapse(
  x,
  category = "category",
  is.long = T,
  with.lapse_bias = F
)

is.MVG_ideal_observer_w_lapse_bias(x, category = "category", is.long = T)
}
\arguments{
\item{x}{Object to be checked.}

\item{category}{Name of the category variable. (default: "category")}

\item{is.long}{Is this check assessing whether the ideal observer is in long format (`TRUE`) or wide format (`FALSE`)?
(default: `TRUE`)}

\item{with.lapse}{Does this ideal observer have a lapse rate? (default: `FALSE`)}

\item{with.lapse_bias}{Does this ideal observer have a lapse bias? (default: `FALSE`)}
}
\value{
A logical.
}
\description{
Check whether \code{x} is an ideal observer with \link[=is.MVG]{multivariate Gaussian (MVG) categories}. Optionally, one can also check whether a lapse rate
and lapse bias is part of the ideal observer.
}
\section{Functions}{
\itemize{
\item \code{is.MVG_ideal_observer_w_lapse}: Also checks whether the MVG has a lapse term.

\item \code{is.MVG_ideal_observer_w_lapse_bias}: Also checks whether the MVG has a lapse bias term.
}}

\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
