% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{is.NIW_belief}
\alias{is.NIW_belief}
\title{Is this a tibble of NIW beliefs?}
\usage{
is.NIW_belief(x, is.long = T, category = "category")
}
\description{
Check whether \code{x} is a tibble of NIW beliefs.
}
