% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-NIW-belief.R
\name{is.NIW_belief}
\alias{is.NIW_belief}
\title{Is this a Normal-Inverse-Wishart (NIW) belief?}
\usage{
is.NIW_belief(x, group = NULL, category = "category", is.long = T, verbose = F)
}
\arguments{
\item{x}{Object to be checked.}

\item{group}{Name of one or more group variables, each unique combination of which describes an NIW_belief. (default: NULL)}

\item{category}{Name of the category variable. (default: "category")}

\item{is.long}{Is this check assessing whether the belief is in long format (\code{TRUE}) or wide format (\code{FALSE})?
(default: \code{TRUE})}
}
\value{
A logical.
}
\description{
Check whether \code{x} is a Normal-Inverse-Wishard (NIW) belief/set of NIW beliefs. An NIW belief describes a distribution of
\link[=is.MVG]{multivariate Gaussian categories}.
}
\seealso{
TBD
}
\keyword{TBD}
