% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-class-NIW-IA.R
\name{is.NIW_ideal_adaptor}
\alias{is.NIW_ideal_adaptor}
\title{Deprecated: is.NIW_ideal_adaptor}
\usage{
is.NIW_ideal_adaptor(
  x,
  group = NULL,
  category = "category",
  is.long = T,
  with.prior = T,
  with.lapse = if (with.lapse_bias) T else F,
  with.lapse_bias = F,
  verbose = F,
  tolerance = MVBU_PROB_TOL,
  ...
)
}
\arguments{
\item{x}{Object to check.}

\item{group}{Name of one or more group variables. (default: NULL)}

\item{category}{Name of the category variable. (default: "category")}

\item{is.long}{Is this check assessing whether the ideal adaptor is in long format? (default: \code{TRUE})}

\item{with.prior}{Does this ideal adaptor have a prior? (default: \code{TRUE})}

\item{with.lapse}{Does this ideal adaptor have a lapse rate? (default: \code{FALSE})}

\item{with.lapse_bias}{Does this ideal adaptor have a lapse bias? (default: \code{FALSE})}

\item{verbose}{Should verbose output be provided? (default: \code{FALSE})}

\item{tolerance}{Tolerance for sum-to-one probability checks. (default: MVBU_PROB_TOL)}

\item{...}{Additional arguments.}
}
\value{
Logical indicating whether \code{x} is a valid NIW ideal adaptor.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{is.NIW_ideal_adaptor()} was deprecated in MVBeliefUpdatr 0.1.0 and will be removed in 0.2.0.
Please use S7 validators and \link{NIW_IdealAdaptor} instead.
}
\seealso{
\link{NIW_IdealAdaptor}
}
\keyword{internal}
