% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-NIW-IA.R
\name{is.NIW_ideal_adaptor}
\alias{is.NIW_ideal_adaptor}
\title{Is this an ideal adaptor with Normal-Inverse-Wishart (NIW) beliefs?}
\usage{
is.NIW_ideal_adaptor(
  x,
  category = "category",
  is.long = T,
  with.lapse = if (with.lapse_bias) T else F,
  with.lapse_bias = F,
  verbose = F,
  tolerance = 1e-05
)
}
\arguments{
\item{x}{Object to be checked.}

\item{category}{Name of the category variable. (default: "category")}

\item{is.long}{Is this check assessing whether the ideal adaptor is in long format (`TRUE`) or wide format (`FALSE`)?
(default: `TRUE`)}

\item{with.lapse}{Does this ideal adaptor have a lapse rate? (default: `FALSE`)}

\item{with.lapse_bias}{Does this ideal adaptor have a lapse bias? (default: `FALSE`)}
}
\value{
A logical.
}
\description{
Check whether \code{x} is an ideal adaptor with \link[=is.NIW_belief]{Normal-Inverse-Wishard (NIW) beliefs}. An ideal adaptor
describes a distribution over \link[=is.MVG_ideal_observer]{ideal observers with multivariate Gaussian categories}. In this
sense, an ideal adaptor describes uncertainty about the true ideal observer. Optionally, one can also check whether a lapse
rate and lapse bias is part of the ideal adaptor.
}
\details{
So far, the ideal adaptor is assumed to have perfect certainty about the prior, lapse rate and lapse bias (if present).
Future implementations might allow uncertainty over these parameters.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
