% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-NIW-IA-stanfit.R
\name{is.NIW_ideal_adaptor_input}
\alias{is.NIW_ideal_adaptor_input}
\title{Is this a list of NIW IBBU inputs?}
\usage{
is.NIW_ideal_adaptor_input(x)
}
\value{
A logical.
}
\description{
Check whether \code{x} is of class \code{NIW_ideal_adaptor_input}.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
