% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mvg-ibbu-stanfit.R
\name{is.NIW_ideal_adaptor_MCMC}
\alias{is.NIW_ideal_adaptor_MCMC}
\alias{is.NIW_ideal_adaptor_w_lapse_MCMC}
\alias{is.NIW_ideal_adaptor_w_bias_MCMC}
\title{Is this a tibble of MCMC draws of an NIW ideal adaptor?}
\usage{
is.NIW_ideal_adaptor_MCMC(
  x,
  is.nested = T,
  is.long = T,
  with.lapse = if (with.bias) T else F,
  with.bias = F
)

is.NIW_ideal_adaptor_w_lapse_MCMC(x, is.nested = T, is.long = T, with.bias = F)

is.NIW_ideal_adaptor_w_bias_MCMC(x, is.nested = T, is.long = T)
}
\value{
A logical.
}
\description{
Check whether \code{x} is a tibble of post-warmup draws of parameters obtained from incremental
conjugate Bayesian belief-updating (IBBU) over a Normal-Inverse-Wishart (NIW) prior.
}
\section{Functions}{
\itemize{
\item \code{is.NIW_ideal_adaptor_w_lapse_MCMC}: Also checks whether x has a lapse term.

\item \code{is.NIW_ideal_adaptor_w_bias_MCMC}: Also checks whether x has a bias term.
}}

\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
