% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{is.NIW_belief}
\alias{is.NIW_belief}
\alias{is.NIW_belief_w_lapse}
\alias{is.NIW_belief_w_bias}
\title{Is this a tibble of NIW beliefs?}
\usage{
is.NIW_belief(
  x,
  category = "category",
  is.long = T,
  with.lapse = if (with.bias) T else F,
  with.bias = F
)

is.NIW_belief_w_lapse(x, category = "category", is.long = T, with.bias = F)

is.NIW_belief_w_bias(x, category = "category", is.long = T)
}
\arguments{
\item{x}{Object to be checked.}

\item{category}{Name of the category variable. (default: "category")}

\item{is.long}{Is this check assessing whether the belief is in long format (`TRUE`) or wide format (`FALSE`)?
(default: `TRUE`)}

\item{w}{}
}
\value{
A logical.
}
\description{
Check whether \code{x} is a tibble of NIW beliefs. Optionally, one can also check whether a lapse rate
and bias is part of the belief.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
