% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-make-objects.R
\name{lift_likelihood_to_model}
\alias{lift_likelihood_to_model}
\alias{lift_exemplars_to_exemplar_model}
\alias{lift_MVG_to_MVG_ideal_observer}
\alias{lift_NIW_belief_to_NIW_ideal_adaptor}
\alias{lift_MVG_ideal_observer_to_NIW_ideal_adaptor}
\title{Turn a representation (template) into a cognitive model.}
\usage{
lift_likelihood_to_model(
  x,
  group = NULL,
  prior = NULL,
  lapse_rate = 0,
  lapse_bias = NULL,
  Sigma_noise = NULL,
  verbose = F
)

lift_exemplars_to_exemplar_model(
  x,
  group = NULL,
  prior = NULL,
  lapse_rate = 0,
  lapse_bias = NULL,
  Sigma_noise = NULL,
  verbose = F
)

lift_MVG_to_MVG_ideal_observer(
  x,
  group = NULL,
  prior = NULL,
  lapse_rate = 0,
  lapse_bias = NULL,
  Sigma_noise = NULL,
  verbose = F
)

lift_NIW_belief_to_NIW_ideal_adaptor(
  x,
  group = NULL,
  prior = NULL,
  lapse_rate = 0,
  lapse_bias = NULL,
  Sigma_noise = NULL,
  verbose = F
)

lift_MVG_ideal_observer_to_NIW_ideal_adaptor(
  x,
  group = NULL,
  kappa,
  nu,
  verbose = F
)
}
\arguments{
\item{x}{An MVG, NIW_belief, or exemplars legacy tibble.}

\item{group}{No longer supported; must be \code{NULL}.}

\item{prior}{Optional category-prior vector. (default: uniform over categories)}

\item{lapse_rate}{Optional lapse rate. (default: 0)}

\item{lapse_bias}{Optional lapse-bias vector. (default: same as \code{prior})}

\item{Sigma_noise}{Optional perceptual-noise covariance matrix. (default: \code{NULL})}

\item{verbose}{Ignored.}
}
\value{
An S7 cognitive-model object of the family matching \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{\link{new_mvg_ideal_observer}}, \code{\link{new_niw_ideal_adaptor}},
or \code{\link{new_exemplar_model}} with an S7 \code{category_template} instead.
}
\keyword{internal}
