% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-objects.R
\name{lift_likelihood_to_model}
\alias{lift_likelihood_to_model}
\alias{lift_MVG_to_MVG_ideal_observer}
\alias{lift_NIW_belief_to_NIW_ideal_adaptor}
\title{Turn an MVG/NIW_belief object into an ideal observer/adaptor}
\usage{
lift_likelihood_to_model(
  x,
  group = NULL,
  category = "category",
  prior = NA_real_,
  lapse_rate = NA_real_,
  bias = NA_real_,
  Sigma_noise = NULL
)

lift_MVG_to_MVG_ideal_observer(
  x,
  group = NULL,
  category = "category",
  prior = NA_real_,
  lapse_rate = NA_real_,
  bias = NA_real_,
  Sigma_noise = NULL
)

lift_NIW_belief_to_NIW_ideal_adaptor(
  x,
  group = NULL,
  category = "category",
  prior = NA_real_,
  lapse_rate = NA_real_,
  bias = NA_real_,
  Sigma_noise = NULL
)
}
\arguments{
\item{x}{Either an MVG or NIW_belief object.}

\item{group}{Optionally, a grouping structure can be specified. If group structure is not NULL, one
NIW belief or ideal adaptor will be derived for each level of \code{group_structure}. (default: NULL)}

\item{category}{Name of variable in \code{data} that contains the category information. (default: "category")}

\item{prior}{Optionally specify a prior probability for each category (in each group). (default: a uniform
prior over all categories).}

\item{lapse_rate}{Optionally specify a lapse rate. (default: \code{NA})}

\item{bias}{Optionally specify a response bias. (default: \code{NA})}

\item{Sigma_noise}{Optionally specify a (multivariate Gaussian) noise covariance matrix. This argument will be
ignored if NULL. (default: NULL)}

\item{kappa}{The strength of the beliefs over the category mean (pseudocounts). (default: same as nu)}

\item{nu}{The strength of the beliefs over the category covariance matrix (pseudocounts). (default: number of
cues + 2)}

\item{verbose}{If true provides more information. (default: FALSE)}

\item{keep.category_parameters}{Should categories' mu and Sigma be included in the output (in addition to m
and S of the prior)? (default: FALSE)}
}
\value{
A tibble that is an NIW_belief or NIW_ideal_adaptor object.
}
\description{
Make an ideal observer or adaptor out of an MVG or NIW_belief object, respectively, by providing the missing
information about the prior, lapse rate, bias, and perceptual noise (if any).
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
