% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-objects.R
\name{lift_likelihood_to_model}
\alias{lift_likelihood_to_model}
\alias{lift_exemplars_to_exemplar_model}
\alias{lift_MVG_to_MVG_ideal_observer}
\alias{lift_NIW_belief_to_NIW_ideal_adaptor}
\alias{lift_MVG_ideal_observer_to_NIW_ideal_adaptor}
\title{Turn an MVG/NIW_belief object into an ideal observer/adaptor}
\usage{
lift_likelihood_to_model(
  x,
  group = NULL,
  prior = rep(1/get_nlevels_of_category_labels_from_model(x),
    get_nlevels_of_category_labels_from_model(x)),
  lapse_rate = 0,
  lapse_bias = rep(1/get_nlevels_of_category_labels_from_model(x),
    get_nlevels_of_category_labels_from_model(x)),
  Sigma_noise = NULL,
  verbose = F
)

lift_exemplars_to_exemplar_model(
  x,
  group = NULL,
  prior = rep(1/(n.cat <- get_nlevels_of_category_labels_from_model(x)), n.cat),
  lapse_rate = 0,
  lapse_bias = rep(1/(n.cat <- get_nlevels_of_category_labels_from_model(x)), n.cat),
  Sigma_noise = NULL,
  verbose = F
)

lift_MVG_to_MVG_ideal_observer(
  x,
  group = NULL,
  prior = rep(1/(n.cat <- get_nlevels_of_category_labels_from_model(x)), n.cat),
  lapse_rate = 0,
  lapse_bias = rep(1/(n.cat <- get_nlevels_of_category_labels_from_model(x)), n.cat),
  Sigma_noise = NULL,
  verbose = F
)

lift_NIW_belief_to_NIW_ideal_adaptor(
  x,
  group = NULL,
  prior = rep(1/(n.cat <- get_nlevels_of_category_labels_from_model(x)), n.cat),
  lapse_rate = 0,
  lapse_bias = rep(1/(n.cat <- get_nlevels_of_category_labels_from_model(x)), n.cat),
  Sigma_noise = NULL,
  verbose = F
)

lift_MVG_ideal_observer_to_NIW_ideal_adaptor(
  x,
  group = NULL,
  kappa,
  nu,
  verbose = F
)
}
\arguments{
\item{x}{Either an MVG or NIW_belief object.}

\item{group}{Optionally, a grouping structure can be specified. If group structure is not NULL, one
NIW belief or ideal adaptor will be derived for each level of \code{group_structure}. (default: NULL)}

\item{prior}{Optionally specify a prior probability for each category (in each group). (default: a uniform
prior over all categories).}

\item{lapse_rate}{Optionally specify a lapse rate. (default: 0)}

\item{lapse_bias}{Optionally specify a lapse bias for each category. (default: uniform bias for all categories)}

\item{Sigma_noise}{Optionally specify a (multivariate Gaussian) covariance matrix of perceptual noise. This argument will be
ignored if NULL. (default: NULL)}

\item{verbose}{If true provides more information. (default: FALSE)}

\item{kappa}{The strength of the beliefs over the category mean (pseudocounts). (default: same as nu)}

\item{nu}{The strength of the beliefs over the category covariance matrix (pseudocounts). (default: number of
cues + 2)}

\item{keep.category_parameters}{Should categories' mu and Sigma be included in the output (in addition to m
and S of the prior)? (default: FALSE)}
}
\value{
A tibble that is an NIW_belief or NIW_ideal_adaptor object.
}
\description{
Make an ideal observer or adaptor out of an MVG or NIW_belief object, respectively, by providing the missing
information about the prior, lapse rate, bias, and perceptual noise (if any).
}
\seealso{
TBD
}
\keyword{TBD}
