% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-core-classes.R
\name{list_model_families}
\alias{list_model_families}
\title{List registered model families in MVBU.}
\usage{
list_model_families()
}
\value{
Character vector of registered family names.
}
\description{
List registered model families in MVBU.
}
