% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-make-objects.R
\name{make_exemplar_model_from_data}
\alias{make_exemplar_model_from_data}
\title{Deprecated: make_exemplar_model_from_data}
\usage{
make_exemplar_model_from_data(
  data,
  group = NULL,
  category = "category",
  cues,
  sim_function = NULL,
  prior = NULL,
  lapse_rate = 0,
  lapse_bias = NULL,
  Sigma_noise = NULL,
  verbose = F
)
}
\arguments{
\item{data}{The tibble or data.frame from which to construct the exemplar representations.}

\item{group}{No longer supported; must be \code{NULL}.}

\item{category}{Name of variable in \code{data} that contains the category information. (default: "category")}

\item{cues}{Name(s) of variables in \code{data} that contain the cue information.}

\item{sim_function}{No longer supported; custom similarity functions are ignored.}

\item{prior}{Optional category-prior vector, forwarded as \code{category_prior}. (default: uniform over categories)}

\item{lapse_rate}{Optional lapse rate. (default: 0)}

\item{lapse_bias}{Optional lapse-bias vector. (default: same as \code{prior})}

\item{Sigma_noise}{Optional perceptual-noise covariance matrix. (default: \code{NULL})}

\item{verbose}{If true provides more information. (default: \code{FALSE})}
}
\value{
An \code{Exemplar_Model} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{\link{new_exemplar_model_from_data}} instead.
}
\keyword{internal}
