% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-objects.R
\name{make_exemplars_from_data}
\alias{make_exemplars_from_data}
\alias{make_exemplar_model_from_data}
\title{Make exemplar models from data.}
\usage{
make_exemplars_from_data(
  data,
  group = NULL,
  category = "category",
  cues,
  sim_function = function(x, y) {
     j <- 2
     k <- 1
     distance <- (x - y)^j
    
    similarity <- exp(-distance * k)
 },
  verbose = F
)

make_exemplar_model_from_data(
  data,
  group = NULL,
  category = "category",
  cues,
  sim_function = function(x, y) {
     j <- 2
     k <- 1
     distance <- (x - y)^j
    
    similarity <- exp(-distance * k)
 },
  ...,
  verbose = F
)
}
\arguments{
\item{data}{The tibble or data.frame from which to construct the exemplar model.}

\item{group}{Optionally, a vector of one or more grouping variables. If group is not NULL, one MVG or
ideal observers will be derived for each level of \code{group}. (default: NULL)}

\item{category}{Name of variable in \code{data} that contains the category information. (default: "category")}

\item{cues}{Name(s) of variables in \code{data} that contain the cue information.}

\item{sim_function}{Similarity function that is used to calculate exemplar-to-examplar similarity. Defaults
to \code{exp(-(x - y)^2)}.}

\item{verbose}{If true provides more information. (default: FALSE)}

\item{prior}{Optionally specify a prior probability for each category (in each group). (default: a uniform
prior over all categories).}

\item{lapse_rate}{Optionally specify a lapse rate. (default: \code{NA})}

\item{lapse_bias}{Optionally specify a lapse bias. (default: \code{NA})}

\item{Sigma_noise}{Optionally specify a (multivariate Gaussian) covariance matrix of perceptual noise. This
argument will be ignored if NULL. (default: NULL)}
}
\value{
A tibble that is an MVG or MVG ideal observer object.
}
\description{
Constructs an exemplar model for all categories found in the data.
}
\seealso{
TBD
}
\keyword{TBD}
