% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-make-objects.R
\name{make_MVG_from_data}
\alias{make_MVG_from_data}
\title{Deprecated: make_MVG_from_data}
\usage{
make_MVG_from_data(
  data,
  group = NULL,
  category = "category",
  cues,
  verbose = F
)
}
\arguments{
\item{data}{The tibble or data.frame from which to construct the exemplar representations.}

\item{group}{No longer supported; must be \code{NULL}.}

\item{category}{Name of variable in \code{data} that contains the category information. (default: "category")}

\item{cues}{Name(s) of variables in \code{data} that contain the cue information.}

\item{verbose}{If true provides more information. (default: \code{FALSE})}
}
\value{
An \code{MVBU_CategoryRepresentationTemplate} of MVG category representations.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{make_MVG_from_data()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0. Please use \code{\link[=new_mvg_category_representation_template_from_data]{new_mvg_category_representation_template_from_data()}} instead.
}
\seealso{
\code{\link[=new_mvg_category_representation_template_from_data]{new_mvg_category_representation_template_from_data()}}
}
\keyword{internal}
