% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-objects.R
\name{make_MVG_from_data}
\alias{make_MVG_from_data}
\alias{make_MVG_ideal_observer_from_data}
\title{Make multivariate Gaussian(s) from data.}
\usage{
make_MVG_from_data(
  data,
  groups = NULL,
  category = "category",
  cues,
  keep.category_parameters = F,
  verbose = F
)

make_MVG_ideal_observer_from_data(
  data,
  groups = NULL,
  category = "category",
  cues,
  prior = NA_real_,
  lapse_rate = NA_real_,
  bias = NA_real_,
  Sigma_noise = NULL,
  keep.category_parameters = F,
  verbose = F
)
}
\arguments{
\item{data}{The tibble or data.frame from which to construct the prior.}

\item{groups}{Optionally, a group variable can be specified. If group is not NULL, one set of multivariate
Gaussian categories will be derived for each level of group. (default: NULL)}

\item{category}{Name of variable in \code{data} that contains the category information. (default: "category")}

\item{cues}{Name(s) of variables in \code{data} that contain the cue information.}

\item{keep.category_parameters}{Should categories' mu and Sigma be included in the output (in addition to m
and S of the prior)? (default: FALSE)}

\item{verbose}{If true provides more information. (default: FALSE)}

\item{prior}{Optionally specify a prior probability for each category (in each group). (default: a uniform
prior over all categories).}

\item{lapse_rate}{Optionally specify a lapse rate. (default: \code{NA})}

\item{bias}{Optionally specify a response bias. (default: \code{NA})}

\item{Sigma_noise}{Optionally specify a (multivariate Gaussian) noise covariance matrix. This argument will be}
}
\value{
A tibble that is an MVG object.
}
\description{
Constructs an \code{\link[=is.MVG]{MVG}} or \code{\link[=is.MVG_ideal_observer]{MVG_ideal_observer}} object with category
information for all categories found in the data.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
