% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-nest-model.R
\name{make_named_square_matrix}
\alias{make_named_square_matrix}
\title{Deprecated: make_named_square_matrix}
\usage{
make_named_square_matrix(x, names)
}
\arguments{
\item{x}{Values for the matrix.}

\item{names}{Character vector of row and column names.}
}
\value{
A named square matrix.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{make_named_square_matrix()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0.
}
\keyword{internal}
