% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-nest-model.R
\name{make_named_vector}
\alias{make_named_vector}
\title{Deprecated: make_named_vector}
\usage{
make_named_vector(x, names)
}
\arguments{
\item{x}{A vector.}

\item{names}{Character vector of names.}
}
\value{
A named vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{make_named_vector()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0.
}
\seealso{
\code{\link[stats:setNames]{stats::setNames()}}
}
\keyword{internal}
