% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-staninput.R
\name{make_staninput}
\alias{make_staninput}
\title{Prepare data to fit ideal_adaptor_stanfit via rstan}
\usage{
make_staninput(
  exposure,
  test,
  cues,
  category = "category",
  response = "response",
  group = "group",
  group.unique = NULL,
  lapse_rate = NULL,
  mu_0 = NULL,
  Sigma_0 = NULL,
  control = control_staninput(),
  stanmodel = "NIW_ideal_adaptor",
  verbose = F
)
}
\arguments{
\item{exposure}{`tibble` or `data.frame` with the exposure data. Each row should be an observation of a category,
and contain information about the category label, the cue values of the observation, and optionally grouping variables.}

\item{test}{`tibble` or `data.frame` with the test data. Each row should be an observation, and contain information
about the cue values of the test stimulus and the participant's response.}

\item{cues}{Names of columns with cue values. Must exist in both exposure and test data.}

\item{category}{Name of column in exposure data that contains the category label. Can be \code{NULL} for unsupervised updating
(not yet implemented). (default: "category")}

\item{response}{Name of column in test data that contains participants' responses. (default: "response")}

\item{group}{Name of column that contains information about which observations form a group. Typically, this is
a variable identifying subjects/participants. Must exist in both exposure and test data. (default: "group")}

\item{group.unique}{Name of column that uniquely identifies each group with identical exposure. This could be a
variable indicating the different conditions in an experiment. Using group.unique is optional, but can be
substantially more efficient if many groups share the same exposure. To ignore, set to \code{NULL}. (default: \code{NULL})}

\item{lapse_rate, mu_0, Sigma_0}{Optionally, lapse rate, prior expected category means (\code{mu_0}) and/or prior expected
category covariance matrices (\code{Sigma_0}) for all categories. Lapse rate should be a number between 0 and 1. For \code{mu_0}
and \code{Sigma_0}, each should be a list, with each element being the expected mean/covariance matrix for a specific
category prior to updating. Elements of \code{mu_0} and \code{Sigma_0} should be ordered in the same order as the levels of the
category variable in \code{exposure} and \code{test}. These prior expected means and covariance matrices could be
estimated, for example, from phonetically annotated speech recordings (see \code{\link{make_MVG_from_data}}
for a convenient way to do so). Internally, m_0 is then set to \code{mu_0} (so that the expected value of the prior
distribution of means is mu_0) and S_0 is set so that the expected value of the inverse-Wishart is \code{Sigma_0} given nu_0.
Importantly, \strong{Sigma_0 should be convolved with perceptual noise (i.e., add perceptual noise covariance matrix to
the category variability covariance matrices when you specify \code{Sigma_0})} since the stancode for the inference of the
NIW ideal adaptor does \emph{not} infer category and noise variability separately.}

\item{control}{A list of control parameters that only experienced users should change since it can change the fitting and
interpretation of the model. See \code{\link{control_staninput}} for details.}
}
\value{
A list consisting of:
\itemize{
  \item{`data`: }{A data.frame with the exposure and test data after exclusion of NAs and other checks.}
  \item{`staninput`: }{A named list of variables and values to be handed to Stan.}
  \item{`transform_information`: }{A list with information about the transformation that was applied to the data.}
}
}
\description{
Take exposure and test data as input, and prepare the data for input into an MVBeliefUpdatr Stan program.
}
\details{
It is important to use \code{group} to identify individuals that had a specific exposure (or no exposure at all)
and specific test trials. You should \emph{not} use \code{group} to identify exposure conditions. Setting \code{group} to an exposure condition
results in an exposure that concatenates the exposure observations from all subjects in that condition. Typically, this
is not what users intend, as it models exposure to the combination of exposure tokens across all subjects, rather than
exposure to one set of those exposure tokens. To achieve this intended outcome, use
\code{group.unique} to identify groups with identical exposure. This will correctly use only one unique instance of the
observations that any level of \code{group} receives during exposure.
}
\seealso{
\code{\link{is.ideal_adaptor_staninput}}
}
\keyword{TBD}
