% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-2025.R
\name{make_staninput_deprecated}
\alias{make_staninput_deprecated}
\title{Deprecated: make_staninput_deprecated}
\usage{
make_staninput_deprecated(
  exposure,
  test,
  cues,
  category = "category",
  response = "response",
  group = "group",
  group.unique = NULL,
  center.observations = T,
  scale.observations = F,
  pca.observations = F,
  pca.cutoff = 1,
  lapse_rate = NULL,
  mu_0 = NULL,
  Sigma_0 = NULL,
  tau_scale = NULL,
  L_omega_eta = 1,
  split_loglik_per_observation = 0,
  use_univariate_updating = FALSE,
  verbose = F
)
}
\arguments{
\item{center.observations}{Should the data be centered based on cues' means during exposure? Note that the cues' means
used for centering are calculated after aggregating the data to all unique combinations specified by \code{group.unique}.
These means are only expected to be the same as the standard deviations over the entire exposure data if the exposure data
are perfectly balanced with regard to \code{group.unique}. Centering will not affect the inferred correlation
or covariance matrices but it will affect the absolute position of the inferred means. The relative position of the inferred
means remains unaffected.
If \code{TRUE} and \code{mu_0} is specified, \code{mu_0} will also be centered (\code{Sigma_0} is not affected by centering and thus not changed).
(default: \code{TRUE})}

\item{scale.observations}{Should the data be standardized based on cues' standard deviation during exposure? Note that the
cues' standard deviations used for scaling are calculated after aggregating the data to all unique combinations specified
by \code{group.unique}. These standard deviations are only expected to be the same as the standard deviations over the entire
exposure data if the exposure data are perfectly balanced with regard to \code{group.unique}.
Scaling will not affect the inferred correlation matrix,
but it will affect the inferred covariance matrix because it affects the inferred standard deviations. It will also
affect the absolute position of the inferred means. The relative position of the inferred means remains unaffected.
If \code{TRUE} and \code{mu_0} and \code{Sigma_0} are specified, \code{mu_0} and \code{Sigma_0} will also be scaled.
(default: \code{FALSE})}

\item{pca.observations}{Should the data be transformed into orthogonal principal components? (default: \code{FALSE})}

\item{pca.cutoff}{Determines which principal components are handed to the MVBeliefUpdatr Stan program: all
components necessary to explain at least the pca.cutoff of the total variance. (default: .95) Ignored if
\code{pca.observation = FALSE}. (default: 1)}
}
\description{
Deprecated. Use \code{\link{fit_ideal_adaptor}} or \code{\link{new_ideal_adaptor_staninput}} instead.
}
