% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{make_vector_column}
\alias{make_vector_column}
\title{Combine a number of columns into a new vector column}
\usage{
make_vector_column(data, cols, vector_col, .keep = "all")
}
\arguments{
\item{data}{A tibble or data.frame.}

\item{cols}{A character vector of variable names to combine.}

\item{vector_col}{Name of the new vector-valued column.}

\item{.keep}{A tidyselect option passed to dplyr::mutate.}
}
\value{
Same as \code{data}.
}
\description{
Combine a number of columns into a new column in which each cell is the vector of values from the original columns.
}
\keyword{TBD}
