% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{make_vector_column}
\alias{make_vector_column}
\title{Combine a number of columns into a new vector column}
\usage{
make_vector_column(data, cols, vector_col, transmute = F)
}
\arguments{
\item{data}{`tibble` or `data.frame`.}

\item{cols}{Vector of characters with names of variables to combine.}

\item{vector_col}{Name of new column of vectors.}
}
\value{
Same as \code{data}.
}
\description{
Combine a number of columns into a new column in which each cell is the vector of values from the original columns.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
