% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-classes.R, R/s7-foundation.R
\name{MNIX-IdealAdaptorDistribution-class}
\alias{MNIX-IdealAdaptorDistribution-class}
\alias{MNIX_IdealAdaptorDistribution}
\title{MNIX Ideal Adaptor Distribution Class}
\usage{
MNIX_IdealAdaptorDistribution(
  model_family = character(0),
  cache = list(),
  metadata = list(),
  group_label = character(0)
)

MNIX_IdealAdaptorDistribution(
  model_family = character(0),
  cache = list(),
  metadata = list(),
  group_label = character(0)
)
}
\description{
Family-typed model-distribution class requiring \code{model_family = "MNIX"}.

Family-typed model-distribution class requiring \code{model_family = "MNIX"}.
}
\seealso{
\link{MNIX_IdealAdaptor}
}
\keyword{internal}
