% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVBeliefUpdatr-package.R
\docType{package}
\name{MVBeliefUpdatr-package}
\alias{MVBeliefUpdatr}
\alias{MVBeliefUpdatr-package}
\title{MVBeliefUpdatr: Fitting, Summarizing, and Visualizing of Multivariate Gaussian Belief-Updating}
\description{
This package provides a number of convenience functions to interact with Stan through rstan, in order to infer prior beliefs about the cue distributions of an arbitrary number of (independent) multivariate Gaussian categories based on categorization responses during a `test phase' following category-labeled inputs during the `exposure phase'. Functions are provided to (1) prepare data as input for the stan program that implements the multivariate Bayesian belief-updating, (2) to summarize and visualize the prior and posterior beliefs represented by the resulting fit, and (3) to generate behavioral test data under a known ground truth (e.g., in order to test the functions for (1) and (2), or simply in order to simulate the consequences of incremental Bayesian belief-updating.
}
\section{Package options}{
 TBD
}

\references{
TBD
}
\author{
\strong{Maintainer}: Florian Jaeger \email{fjaeger@ur.rochester.edu} (\href{https://orcid.org/0000-0002-1158-7308}{ORCID})

}
\keyword{internal}
