% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-transform-information.R
\docType{class}
\name{MVBU-TransformInformation-class}
\alias{MVBU-TransformInformation-class}
\title{S7 base class for transformation information for Stan models.}
\description{
S7 base class for transformation information for Stan models.
}
