% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit.R
\docType{class}
\name{MVBU_Stanfit}
\alias{MVBU_Stanfit}
\alias{IdealAdaptorStanfit}
\title{MVBeliefUpdatr Stanfit Classes and Constructors}
\usage{
MVBU_Stanfit(
  data = data.frame(),
  staninput = NULL,
  stanvars = NULL,
  backend = "rstan",
  save_pars = NULL,
  stan_args = list(),
  stanfit = NULL,
  basis = NULL,
  transform_information = NULL,
  criteria = list(),
  file = NULL,
  version = NULL,
  metadata = list()
)

IdealAdaptorStanfit(
  data = data.frame(),
  staninput = NULL,
  stanvars = NULL,
  backend = "rstan",
  save_pars = NULL,
  stan_args = list(),
  stanfit = NULL,
  basis = NULL,
  transform_information = NULL,
  criteria = list(),
  file = NULL,
  version = NULL,
  metadata = list()
)
}
\description{
An S7 base class for Stan fit objects
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{data.frame} containing the data used to fit the model.}

\item{\code{staninput}}{A Stan input object containing the data handed to rstan through
\code{\link{make_staninput}}. The staninput object contains at
least two components: \code{transformed} and \code{untransformed}.}

\item{\code{stanvars}}{A \code{\link{stanvars}} object or \code{NULL}.}

\item{\code{backend}}{The name of the backend used to fit the model (character).}

\item{\code{save_pars}}{Optional storage for saved parameter names.}

\item{\code{stan_args}}{Named list of additional control arguments that were passed
to the Stan backend directly. NOT YET USED}

\item{\code{stanfit}}{An object of class
\code{\link[rstan:stanfit-class]{stanfit}} containing posterior draws.}

\item{\code{basis}}{An object that contains a small subset of the Stan data
created at fitting time, needed to process new data. NOT YET USED}

\item{\code{transform_information}}{An object of type \code{\link{MVBU_TransformInformation}}.}

\item{\code{criteria}}{An empty \code{list} for adding model fit criteria
after estimation of the model. NOT YET USED}

\item{\code{file}}{Optional name of a file in which the model object was stored in
or loaded from.}

\item{\code{version}}{The versions of \pkg{MVBeliefUpdatr} and \pkg{rstan} with
which the model was fitted.}

\item{\code{metadata}}{List containing auxiliary information including
\code{label_information}.}
}}

