% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-stanfit-input.R
\docType{class}
\name{control_staninput}
\alias{control_staninput}
\alias{MVBU_StanfitInput}
\alias{IdealAdaptorStanfitInput}
\title{Specify control parameters for new_ideal_adaptor_stanfit_input()}
\usage{
control_staninput(
  tau_scale = 5,
  L_omega_eta = 1,
  split_loglik_per_observation = 0,
  transform_type = c("identity", "center", "standardize", "PCA whiten", "ZCA whiten")[3]
)

IdealAdaptorStanfitInput(
  data = data.frame(),
  staninput = NULL,
  transform_information = NULL,
  metadata = list()
)
}
\arguments{
\item{tau_scale}{A vector of scales for the Cauchy priors for each cue's
standard deviations.}

\item{L_omega_eta}{LKJ prior parameter for correlation matrices.}

\item{split_loglik_per_observation}{Split log likelihood per observation.}

\item{transform_type}{Affine transformation type applied to the data.}
}
\value{
A list of control parameters.
}
\description{
This function is used to specify control parameters for the \code{new_ideal_adaptor_stanfit_input()} function, and to provide
reasonable defaults for any of the unspecified parameters.

An S7 Class for Ideal-Adaptor Fit-Input Objects
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{data.frame} containing the input data.}

\item{\code{staninput}}{An \code{\link{IdealAdaptorStaninput}} object.}

\item{\code{transform_information}}{An \code{\link{MVBU_TransformInformation}} object.}

\item{\code{metadata}}{A named \code{list} containing \code{label_information}
with character vectors (\code{cue}, \code{category}, and \code{group}).}
}}

