% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-nest-model.R
\name{nest_cue_information_in_model}
\alias{nest_cue_information_in_model}
\alias{unnest_cue_information_in_model}
\title{Deprecated: nest_cue_information_in_model}
\usage{
nest_cue_information_in_model(model)

unnest_cue_information_in_model(model)
}
\arguments{
\item{model}{A model object or tibble.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{nest_cue_information_in_model()} and \code{unnest_cue_information_in_model()} were deprecated
in MVBeliefUpdatr 0.1.0 and will be removed in 0.2.0. S7 cognitive models store native
parameters directly within category representations.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{unnest_cue_information_in_model()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0.
}
\keyword{internal}
