% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-model.R
\name{nest_cue_information_in_model}
\alias{nest_cue_information_in_model}
\alias{unnest_cue_information_in_model}
\title{Nest/unnest the cue information in a model}
\usage{
nest_cue_information_in_model(model)

unnest_cue_information_in_model(model)
}
\arguments{
\item{model}{A model object.}
}
\description{
Take the centrality (e.g., mu or m) and scatter parameters (e.g., Sigma or S) of the model and nest them
into vector/matrix form or unnest them. The unnested format introduces two cue columns 'cue' and 'cue2',
which indicate the cue label for the cue value in the mu/m and Sigma/S columns.
}
