% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-foundation.R
\name{new_category_representation}
\alias{new_category_representation}
\title{Construct a base representation object}
\usage{
new_category_representation(
  category_labels,
  cue_labels,
  category_likelihood = function(...) stop("category_likelihood not implemented.", call.
    = FALSE),
  metadata = list()
)
}
\description{
Construct a base representation object
}
\keyword{internal}
