% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-core-classes.R
\name{new_category_representation}
\alias{new_category_representation}
\title{Construct a base representation object}
\usage{
new_category_representation(
  category_labels,
  cue_labels,
  category_likelihood_function = NULL,
  metadata = list()
)
}
\description{
Construct a base representation object
}
