% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-foundation.R
\name{new_category_representation_template}
\alias{new_category_representation_template}
\title{Construct a category representation set}
\usage{
new_category_representation_template(representations, metadata = list())
}
\description{
Construct a category representation set
}
\keyword{internal}
