% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-classes.R, R/s7-foundation.R
\name{new_exemplar_category_representation}
\alias{new_exemplar_category_representation}
\title{Construct an Exemplar representation object}
\usage{
new_exemplar_category_representation(
  category_labels,
  cue_labels,
  exemplars,
  exemplar_weights = NULL,
  metadata = list()
)

new_exemplar_category_representation(
  category_labels,
  cue_labels,
  exemplars,
  exemplar_weights = NULL,
  metadata = list()
)
}
\description{
Construct an Exemplar representation object

Construct an Exemplar representation object
}
\keyword{internal}
