% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-foundation.R
\name{new_exemplar_model}
\alias{new_exemplar_model}
\title{Construct an Exemplar model object}
\usage{
new_exemplar_model(
  category_likelihood,
  decision_rule = "sampling",
  category_prior = NULL,
  lapse_rate = 0,
  lapse_bias = NULL,
  metadata = list()
)
}
\description{
Construct an Exemplar model object
}
\keyword{internal}
