% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-core-classes.R
\name{new_mnix_category_representation}
\alias{new_mnix_category_representation}
\title{Construct a MNIX representation object}
\usage{
new_mnix_category_representation(
  category_labels,
  cue_labels,
  component_m,
  component_kappa,
  component_nu,
  component_sigma2,
  component_weights = NULL,
  metadata = list()
)
}
\description{
Construct a MNIX representation object
}
