% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-foundation.R
\name{new_mnix_ideal_adaptor}
\alias{new_mnix_ideal_adaptor}
\title{Construct a MNIX ideal adaptor object}
\usage{
new_mnix_ideal_adaptor(
  category_likelihood,
  decision_rule = "sampling",
  category_prior = NULL,
  lapse_rate = 0,
  lapse_bias = NULL,
  metadata = list()
)
}
\description{
Construct a MNIX ideal adaptor object
}
\keyword{internal}
