% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-classes.R, R/s7-foundation.R
\name{new_mnix_model_distribution}
\alias{new_mnix_model_distribution}
\title{Construct a MNIX model distribution object}
\usage{
new_mnix_model_distribution(
  cache = list(),
  metadata = list(),
  group_label = ""
)

new_mnix_model_distribution(
  cache = list(),
  metadata = list(),
  group_label = ""
)
}
\description{
Construct a MNIX model distribution object

Construct a MNIX model distribution object
}
\keyword{internal}
