% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-foundation.R
\name{new_model_distribution}
\alias{new_model_distribution}
\title{Construct a base inferred-model object}
\usage{
new_model_distribution(
  model_family,
  cache = list(),
  metadata = list(),
  group_label = ""
)
}
\description{
Construct a base inferred-model object
}
\keyword{internal}
