% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-classes.R, R/s7-foundation.R
\name{new_muvg_category_representation}
\alias{new_muvg_category_representation}
\title{Construct a MUVG representation object}
\usage{
new_muvg_category_representation(
  category_labels,
  cue_labels,
  component_mu,
  component_sigma2,
  component_weights = NULL,
  metadata = list()
)

new_muvg_category_representation(
  category_labels,
  cue_labels,
  component_mu,
  component_sigma2,
  component_weights = NULL,
  metadata = list()
)
}
\description{
Construct a MUVG representation object

Construct a MUVG representation object
}
\keyword{internal}
