% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-foundation.R
\name{new_muvg_model_distribution}
\alias{new_muvg_model_distribution}
\title{Construct a MUVG model distribution object}
\usage{
new_muvg_model_distribution(
  cache = list(),
  metadata = list(),
  group_label = ""
)
}
\description{
Construct a MUVG model distribution object
}
\keyword{internal}
