% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-classes.R, R/s7-foundation.R
\name{new_mvbu_object}
\alias{new_mvbu_object}
\title{Construct a base MVBU object}
\usage{
new_mvbu_object()

new_mvbu_object()
}
\description{
Construct a base MVBU object

Construct a base MVBU object
}
\keyword{internal}
