% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-core-classes.R
\name{new_mvg_category_representation}
\alias{new_mvg_category_representation}
\title{Construct a MVG representation object}
\usage{
new_mvg_category_representation(
  category_labels,
  cue_labels,
  mu,
  Sigma,
  metadata = list()
)
}
\description{
Construct a MVG representation object
}
