% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-core-classes.R
\name{new_mvg_ideal_observer}
\alias{new_mvg_ideal_observer}
\title{Construct a MVG ideal observer object}
\usage{
new_mvg_ideal_observer(
  category_template = NULL,
  decision_rule = "sampling",
  category_prior = NULL,
  lapse_rate = 0,
  lapse_bias = NULL,
  Sigma_noise = NULL,
  noise_treatment = "no_noise",
  lapse_treatment = "no_lapses",
  metadata = list()
)
}
\description{
Construct a MVG ideal observer object
}
