% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-classes.R
\name{new_niw_category_representation}
\alias{new_niw_category_representation}
\title{Construct a NIW representation object}
\usage{
new_niw_category_representation(
  category_labels,
  cue_labels,
  m,
  kappa,
  nu,
  S,
  metadata = list()
)
}
\description{
Construct a NIW representation object
}
\keyword{internal}
