% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-core-classes.R
\name{new_niw_ideal_adaptor}
\alias{new_niw_ideal_adaptor}
\title{Construct a NIW ideal adaptor object}
\usage{
new_niw_ideal_adaptor(
  category_template = NULL,
  decision_rule = "sampling",
  category_prior = NULL,
  lapse_rate = 0,
  lapse_bias = NULL,
  Sigma_noise = NULL,
  noise_treatment = "no_noise",
  lapse_treatment = "no_lapses",
  metadata = list()
)
}
\description{
Construct a NIW ideal adaptor object
}
