% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-core-classes.R
\name{new_nix_category_representation}
\alias{new_nix_category_representation}
\title{Construct a NIX representation object}
\usage{
new_nix_category_representation(
  category_labels,
  cue_labels,
  m,
  kappa,
  nu,
  sigma2,
  metadata = list()
)
}
\description{
Construct a NIX representation object
}
