% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-classes.R, R/s7-foundation.R
\name{new_nix_ideal_adaptor}
\alias{new_nix_ideal_adaptor}
\title{Construct a NIX ideal adaptor object}
\usage{
new_nix_ideal_adaptor(
  category_likelihood,
  decision_rule = "sampling",
  category_prior = NULL,
  lapse_rate = 0,
  lapse_bias = NULL,
  metadata = list()
)

new_nix_ideal_adaptor(
  category_likelihood,
  decision_rule = "sampling",
  category_prior = NULL,
  lapse_rate = 0,
  lapse_bias = NULL,
  metadata = list()
)
}
\description{
Construct a NIX ideal adaptor object

Construct a NIX ideal adaptor object
}
\keyword{internal}
