% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-core-classes.R
\name{new_nix_model_distribution}
\alias{new_nix_model_distribution}
\title{Construct a NIX model distribution object.}
\usage{
new_nix_model_distribution(cache = list(), metadata = list(), group_label = "")
}
\arguments{
\item{cache}{Optional cache list.}

\item{metadata}{Optional metadata list.}

\item{group_label}{Optional group label.}
}
\value{
A NIX S7 model-distribution object.
}
\description{
Construct a NIX model distribution object.
}
