% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-classes.R
\name{new_uvg_category_representation}
\alias{new_uvg_category_representation}
\title{Construct a UVG representation object}
\usage{
new_uvg_category_representation(
  category_labels,
  cue_labels,
  mu,
  sigma2,
  metadata = list()
)
}
\description{
Construct a UVG representation object
}
\keyword{internal}
