% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-foundation.R
\name{new_uvg_ideal_observer}
\alias{new_uvg_ideal_observer}
\title{Construct a UVG ideal observer object}
\usage{
new_uvg_ideal_observer(
  category_likelihood,
  decision_rule = "sampling",
  category_prior = NULL,
  lapse_rate = 0,
  lapse_bias = NULL,
  metadata = list()
)
}
\description{
Construct a UVG ideal observer object
}
\keyword{internal}
