% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-classes.R
\name{new_uvg_model_distribution}
\alias{new_uvg_model_distribution}
\title{Construct a UVG model distribution object}
\usage{
new_uvg_model_distribution(cache = list(), metadata = list(), group_label = "")
}
\description{
Construct a UVG model distribution object
}
\keyword{internal}
