% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-NIW-IA.R
\docType{class}
\name{NIW_ideal_adaptor-class}
\alias{NIW_ideal_adaptor-class}
\alias{NIW_ideal_adaptor}
\title{An S4 class for NIW ideal adaptor objects.}
\description{
An S4 class for NIW ideal adaptor objects.
}
\details{
See \code{methods(class = "NIW_ideal_adaptor")} for an overview of available methods.
}
