% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-foundation.R
\name{NIX-IdealAdaptorDistribution-class}
\alias{NIX-IdealAdaptorDistribution-class}
\alias{NIX_IdealAdaptorDistribution}
\title{NIX Ideal Adaptor Distribution Class}
\usage{
NIX_IdealAdaptorDistribution(
  model_family = character(0),
  cache = list(),
  metadata = list(),
  group_label = character(0)
)
}
\description{
Family-typed model-distribution class requiring \code{model_family = "NIX"}.
}
\seealso{
\link{NIX_IdealAdaptor}
}
\keyword{internal}
