% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-expected-categories.R
\name{plot_expected_categories}
\alias{plot_expected_categories}
\alias{plot_expected_ibbu_stanfit_categories}
\alias{plot_expected_categories_contour}
\alias{plot_expected_categories_density}
\alias{plot_expected_categories_contour2D}
\alias{plot_expected_categories_contour_2D}
\alias{plot_expected_categories_density1D}
\alias{plot_expected_categories_density_1D}
\alias{plot_expected_categories_density2D}
\alias{plot_expected_categories_density_2D}
\alias{plot_expected_categories.NIW_ideal_adaptor_stanfit}
\alias{plot_expected_categories_contour.NIW_ideal_adaptor_stanfit}
\alias{plot_expected_ibbu_stanfit_categories_contour}
\alias{plot_expected_categories_density.NIW_ideal_adaptor_stanfit}
\alias{plot_expected_ibbu_stanfit_categories_density}
\alias{plot_expected_categories_contour2D.NIW_ideal_adaptor}
\alias{plot_expected_categories_contour2D.NIW_ideal_adaptor_stanfit}
\alias{plot_expected_ibbu_stanfit_categories_contour2D}
\alias{plot_expected_categories_density1D.NIW_ideal_adaptor}
\alias{plot_expected_categories_density1D.NIW_ideal_adaptor_stanfit}
\alias{plot_expected_ibbu_stanfit_categories_density1D}
\alias{plot_expected_categories_density2D.NIW_ideal_adaptor_stanfit}
\alias{plot_expected_ibbu_stanfit_categories_density2D}
\title{Plot expected categories of ideal adaptor stanfit}
\usage{
plot_expected_categories(model, ...)

plot_expected_categories_contour(model, ...)

plot_expected_categories_density(model, ...)

plot_expected_categories_contour2D(model, ...)

plot_expected_categories_density1D(model, ...)

plot_expected_categories_density2D(model, ...)

\method{plot_expected_categories}{NIW_ideal_adaptor_stanfit}(model, type, ...)

\method{plot_expected_categories_contour}{NIW_ideal_adaptor_stanfit}(model, cues = get_cue_levels(model), ...)

\method{plot_expected_categories_density}{NIW_ideal_adaptor_stanfit}(model, cues = get_cue_levels(model), ...)

\method{plot_expected_categories_contour2D}{NIW_ideal_adaptor}(
  x,
  levels = c(1/2, 2/3, 4/5, 9/10, 19/20),
  data.exposure = NULL,
  data.test = NULL,
  facet_rows_by = NULL,
  facet_cols_by = NULL,
  facet_wrap_by = NULL,
  animate_by = NULL,
  animation_follow = F,
  category.ids = NULL,
  category.labels = NULL,
  category.colors = NULL,
  ...
)

\method{plot_expected_categories_contour2D}{NIW_ideal_adaptor_stanfit}(
  model,
  categories = get_category_levels(model),
  groups = get_group_levels(model, include_prior = T),
  cues = get_cue_levels(model),
  plot.test = T,
  plot.exposure = F,
  annotate_inferred_category_means = c("rug", "text"),
  untransform_cues = TRUE,
  levels = plogis(seq(-15, qlogis(0.95), length.out = 20)),
  category.colors = get_default_colors("category", categories)
)

\method{plot_expected_categories_density1D}{NIW_ideal_adaptor}(
  x,
  data.exposure = NULL,
  data.test = NULL,
  facet_rows_by = NULL,
  facet_cols_by = NULL,
  facet_wrap_by = NULL,
  animate_by = NULL,
  animation_follow = F,
  xlim,
  ylim = NULL,
  x.expand = c(0, 0),
  category.ids = NULL,
  category.labels = NULL,
  category.colors = NULL,
  category.linetypes = NULL,
  ...
)

\method{plot_expected_categories_density1D}{NIW_ideal_adaptor_stanfit}(
  model,
  categories = get_category_levels(model),
  groups = get_group_levels(model, include_prior = T),
  cues = get_cue_levels(model),
  ndraws = NULL,
  plot.test = T,
  plot.exposure = F,
  annotate_inferred_category_means = c("rug", "text"),
  untransform_cues = TRUE,
  category.colors = get_default_colors("category", categories),
  xlim = NULL,
  resolution = 101
)

\method{plot_expected_categories_density2D}{NIW_ideal_adaptor_stanfit}(
  model,
  categories = get_category_levels(model),
  groups = get_group_levels(model, include_prior = T),
  cues = get_cue_levels(model),
  ndraws = NULL,
  plot.test = T,
  plot.exposure = F,
  annotate_inferred_category_means = c("rug", "text"),
  untransform_cues = TRUE,
  category.colors = get_default_colors("category", categories),
  xlim = NULL,
  ylim = NULL,
  resolution = 25
)
}
\arguments{
\item{model}{An ideal adaptor stanfit object.}

\item{...}{additional arguments to geom_line.}

\item{type}{Either `"contour"` or `"density"`, specifying the type of plot. Note that the contour plot is *much*
faster. It simply gets the expected values of \code{mu} (based on the NIW parameter \code{m}) and \code{Sigma}
(based on the NIW parameters \code{S} and \code{nu}) at each MCMC draw, and then averages over
all MCMC draws. The plotted categories represent those means of the expected \code{mu} and \code{Sigma}. The
density plot instead calculates the posterior predictive for each MCMC draw (i.e, the multivariate Student-T
density based on the NIW parameters \code{m, S, kappa, nu}), and then averages those densities. Since this is
done for *all* points defined by the data.grid this can be rather computationally expensive and slow.}

\item{x}{An \code{\link{NIW_belief}} or \code{\link{NIW_beliefs}} object.}

\item{levels}{Levels of the confidence ellipses. (default: .5, .66, .8, .9., and .95)}

\item{data.exposure}{Optional \code{tibble} or \code{data.frame} that contains exposure data to be plotted. (default: `NULL`)}

\item{data.test}{Optional \code{tibble} or \code{data.frame} that contains test data to be plotted. (default: `NULL`)}

\item{facet_rows_by, facet_cols_by, facet_wrap_by, animate_by}{Which group variables, if any, should be used for faceting and/or
animation? (defaults: `NULL`)}

\item{animation_follow}{Should the animation follow the data (zoom in and out)? (default: `FALSE`)}

\item{category.ids}{Vector of category IDs to be plotted or leave `NULL` to plot all groups. (default: `NULL`)}

\item{category.labels}{Vector of group labels of same length as `category.ids` or `NULL` to use defaults. (default: `NULL`)}

\item{category.colors}{Vector of colors of same length as category.ids or `NULL` to use defaults. (default: `NULL`)}

\item{categories, groups, cues}{Character vector of categories, groups, and cues to be plotted.
(default: all categories, groups, and cues in the model will be plotted)}

\item{plot.test, plot.exposure}{Should the test and/or exposure stimuli be plotted? (default: `TRUE` for `plot.test`,
`FALSE` for `plot.exposure`) The test items are plotted as black points. The exposure mean is plotted as point,
and the .95 interval of cue distributions during exposure are plotted as dashed ellipse in the same color as the
expected categories.}

\item{annotate_inferred_category_means}{Character vector indicating whether the location and value of the mean be
indicated through data rugs (`"rug"`) and/or text labels (`"text"`)? Set to NULL to ignore. (default: `c("rug", "text")`)}

\item{untransform_cues}{Should m_0 and S_0 be transformed back into the original cue space? (default: `TRUE`)}

\item{xlim, ylim}{Limits for the x- and y-axis.}

\item{category.linetypes}{Vector of linetypes of same length as category.ids or `NULL` to use defaults. (default: `NULL`)
Currently being ignored.}

\item{ndraws}{Number of draws from posterior to use for plot, or `NULL` if all draws are to be returned. (default: `NULL`)}

\item{resolution}{For density plots. How many steps along x and y should be calculated? Note that computational
complexity increases quadratically with resolution. (default: 25)}

\item{data.grid.xlim, data.grid.ylim, data.grid.resolution}{Used only if `type` is `"density"`. Limits for x- and y-axis as
well as resolution of the data.grid, defining the range over which the posterior predictive (multivariate Student-T density)
is calculated. Note that the number of densities to calculate is a *quadratic* function of `data.grid.resolution`. The default
for `data.grid.resolution` is 10, corresponding to 100 densities to be calculated for each MCMC draw.}
}
\value{
ggplot object.

ggplot object.

ggplot object.
}
\description{
Plot bivariate Gaussian categories expected given the parameters inferred by incremental Bayesian belief-
updating (IBBU) for an ideal adaptor. Specifically, the categories are derived by marginalizing over the uncertainty represented
by the posterior MCMC samples. Two methods are available (specified by `type`), which differ in their
computational demands and speed.

Plot bivariate Gaussian categories expected given NIW belief(s). One NIW belief describes the uncertainty about the
category statistics of all categories. This includes the m (the mean of category means \eqn{\mu}), S (the scattermatrix),
kappa (the strength of the belief in m) and nu (the strength of the belief in S).

Plot univariate Gaussian categories expected given NIW belief(s). One NIW belief describes the uncertainty about the
category statistics of all categories. This includes the m (the mean of category means \eqn{\mu}), S (the scattermatrix),
kappa (the strength of the belief in m) and nu (the strength of the belief in S). For the univariate case, m and S are
scalars \insertCite{@see @murphy2012 p. 136}{MVBeliefUpdatr}.
}
\details{
Typically, the categories, groups, and cues
are automatically added to the fit during the creation of the fit. If necessary, however, it is possible to use
\code{\link[tidybayes]{recover_types}} on the stanfit object to add or change these levels later.

It is possible to hand more than one NIW belief to this function, and to facet or animate by variables that uniquely
identify the different beliefs. For example, one can plot
different priors for different talkers (grouping by talker), or different posteriors for different exposure conditions
(grouping by exposure condition), the incremental updating of NIW beliefs (grouping by observations), or any combinations
of these.

It is possible to hand more than one NIW belief to this function, and to facet or animate by variables that uniquely
identify the different beliefs. For example, one can plot
different priors for different talkers (grouping by talker), or different posteriors for different exposure conditions
(grouping by exposure condition), the incremental updating of NIW beliefs (grouping by observations), or any combinations
of these.
}
\references{
\insertRef{murphy2012}{MVBeliefUpdatr}
}
\seealso{
TBD

TBD

TBD
}
\keyword{TBD}
