% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-fit.R
\name{plot_expected_categories_contour2D}
\alias{plot_expected_categories_contour2D}
\title{Plot expected bivariate (2D) categories.}
\usage{
plot_expected_categories_contour2D(
  x,
  grouping.var = NULL,
  panel.group = !animate.group,
  animate.group = !panel.group,
  levels = plogis(seq(-15, qlogis(0.95), length.out = 20)),
  category.ids = NULL,
  category.labels = NULL,
  category.colors = NULL,
  category.linetypes = NULL
)
}
\arguments{
\item{x}{NIW belief object.}

\item{category.ids}{Vector of category IDs to be plotted or leave `NULL` to plot all groups. (default: `NULL`) It is possible
to use \code{\link[tidybayes]{recover_types}} on the stanfit object prior to handing it to this plotting function.}

\item{category.labels}{Vector of group labels of same length as `category.ids` or `NULL` to use defaults. (default: `NULL`)}

\item{category.colors}{Vector of colors of same length as category.ids or `NULL` to use defaults. (default: `NULL`)}

\item{category.linetypes}{Vector of linetypes of same length as category.ids or `NULL` to use defaults. (default: `NULL`)
Currently being ignored.}
}
\value{
ggplot object.
}
\description{
Plot bivariate Gaussian categories expected given NIW parameters.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
