% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-forward-ibbu.R
\name{plot_expected_categories_contour2D}
\alias{plot_expected_categories_contour2D}
\title{Plot expected bivariate (2D) categories}
\usage{
plot_expected_categories_contour2D(
  x,
  facet_rows_by = NULL,
  facet_cols_by = NULL,
  animate_by = NULL,
  animation_follow = F,
  levels = c(1/2, 2/3, 4/5, 9/10, 19/20),
  data.exposure = NULL,
  data.test = NULL,
  category.ids = NULL,
  category.labels = NULL,
  category.colors = NULL,
  category.linetypes = NULL
)
}
\arguments{
\item{x}{NIW belief object.}

\item{facet_rows_by, facet_cols_by, animate_by}{Which group variables, if any, should be used for faceting and/or
animation? (defaults: `NULL`)}

\item{levels}{Levels of the confidence ellipses. (default: .5, .66, .8, .9., and .95)}

\item{data.exposure}{Optional \code{tibble} or \code{data.frame} that contains exposure data to be plotted. (default: `NULL`)}

\item{data.test}{Optional \code{tibble} or \code{data.frame} that contains test data to be plotted. (default: `NULL`)}

\item{category.ids}{Vector of category IDs to be plotted or leave `NULL` to plot all groups. (default: `NULL`) It is possible
to use \code{\link[tidybayes]{recover_types}} on the stanfit object prior to handing it to this plotting function.}

\item{category.labels}{Vector of group labels of same length as `category.ids` or `NULL` to use defaults. (default: `NULL`)}

\item{category.colors}{Vector of colors of same length as category.ids or `NULL` to use defaults. (default: `NULL`)}

\item{category.linetypes}{Vector of linetypes of same length as category.ids or `NULL` to use defaults. (default: `NULL`)
Currently being ignored.}
}
\value{
ggplot object.
}
\description{
Plot bivariate Gaussian categories expected given NIW belief(s). One NIW belief describes the uncertainty about the
category statistics of all categories. This includes the M (the mean of category means \eqn{\mu}), S (the scattermatrix),
kappa (the strength of the belief in M) and nu (the strength of the belief in S).
}
\details{
It is possible to hand more than one NIW belief to this function, and to facet or animate by variables that uniquely
identify the different beliefs. For example, one can plot
different priors for different talkers (grouping by talker), or different posteriors for different exposure conditions
(grouping by exposure condition), the incremental updating of NIW beliefs (grouping by observations), or any combinations
of these.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
