% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-visualize-NIW-IA.R
\name{plot_expected_categorization_function_1D}
\alias{plot_expected_categorization_function_1D}
\title{Deprecated: plot_expected_categorization_function_1D}
\usage{
plot_expected_categorization_function_1D(x, ...)
}
\arguments{
\item{x}{Model object.}

\item{...}{Arguments passed to \code{\link[=plot_categorization_function]{plot_categorization_function()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{plot_expected_categorization_function_1D()} was deprecated in MVBeliefUpdatr 0.1.0 and will be
removed in 0.2.0. Please use \code{\link[=plot_categorization_function]{plot_categorization_function()}} instead.
}
\seealso{
\code{\link[=plot_categorization_function]{plot_categorization_function()}}
}
\keyword{internal}
