% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-forward-ibbu.R
\name{plot_expected_categorization_function_1D}
\alias{plot_expected_categorization_function_1D}
\title{Plot expected categorization function for univariate (1D) categories.}
\usage{
plot_expected_categorization_function_1D(
  x,
  data.exposure = NULL,
  data.test = NULL,
  target_category = 1,
  logit = F,
  xlim,
  ylim = NULL,
  facet_rows_by = NULL,
  facet_cols_by = NULL,
  facet_wrap_by = NULL,
  animate_by = NULL,
  animation_follow = F,
  category.ids = NULL,
  category.labels = NULL,
  category.colors = NULL,
  category.linetypes = NULL,
  ...
)
}
\arguments{
\item{x}{An \code{\link{NIW_belief}} or \code{\link{NIW_beliefs}} object.}

\item{data.exposure}{Optional \code{tibble} or \code{data.frame} that contains exposure data to be plotted. (default: `NULL`)}

\item{data.test}{Optional \code{tibble} or \code{data.frame} that contains test data to be plotted. (default: `NULL`)}

\item{target_category}{The index of the category for which categorization should be shown. (default: `1`)}

\item{logit}{Should the categorization function be plotted in logit (`TRUE`) or probabilities (`FALSE`)?
(default: `FALSE`)}

\item{xlim, ylim}{Limits for the x- and y-axis.}

\item{facet_rows_by}{Which group variables, if any, should be used for faceting and/or
animation? (defaults: `NULL`)}

\item{facet_cols_by}{Which group variables, if any, should be used for faceting and/or
animation? (defaults: `NULL`)}

\item{facet_wrap_by}{Which group variables, if any, should be used for faceting and/or
animation? (defaults: `NULL`)}

\item{animate_by}{Which group variables, if any, should be used for faceting and/or
animation? (defaults: `NULL`)}

\item{animation_follow}{Should the animation follow the data (zoom in and out)? (default: `FALSE`)}

\item{category.ids}{Vector of category IDs to be plotted or leave `NULL` to plot all groups. (default: `NULL`) It is possible
to use \code{\link[tidybayes]{recover_types}} on the stanfit object prior to handing it to this plotting function.}

\item{category.labels}{Vector of group labels of same length as `category.ids` or `NULL` to use defaults. (default: `NULL`)}

\item{category.colors}{Vector of colors of same length as category.ids or `NULL` to use defaults. (default: `NULL`)}

\item{category.linetypes}{Vector of linetypes of same length as category.ids or `NULL` to use defaults. (default: `NULL`)
Currently being ignored.}

\item{...}{additional arguments handed to geom_polygon.}
}
\value{
ggplot object.
}
\description{
Plot categorization function for univariate Gaussian categories expected given NIW parameters.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
