% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-NIW-IA.R
\name{plot_expected_categorization_function_2D}
\alias{plot_expected_categorization_function_2D}
\title{Plot expected categorization function for bivariate (2D) categories.}
\usage{
plot_expected_categorization_function_2D(
  x,
  data.exposure = NULL,
  data.test = NULL,
  noise_treatment = infer_default_noise_treatment(x$Sigma_noise),
  lapse_treatment = c("no_lapses", "marginalize")[2],
  target_category = 1,
  logit = F,
  xlim,
  ylim,
  resolution = 25,
  facet_rows_by = NULL,
  facet_cols_by = NULL,
  facet_wrap_by = NULL,
  animate_by = NULL,
  animation_follow = F,
  category.ids = NULL,
  category.labels = NULL,
  category.colors = NULL,
  ...
)
}
\arguments{
\item{x}{An \code{\link{NIW_belief}} or \code{\link{NIW_beliefs}} object.}

\item{data.exposure}{Optional \code{tibble} or \code{data.frame} that contains exposure data to be plotted. (default: \code{NULL})}

\item{data.test}{Optional \code{tibble} or \code{data.frame} that contains test data to be plotted. (default: \code{NULL})}

\item{noise_treatment}{Should the consequences of perceptual noise be included in the categorization function
("marginalize") or not ("no_noise")? (default: "marginalize")}

\item{lapse_treatment}{Should the consequences of attentional lapses be included in the categorization function
("marginalize") or not ("no_lapses")? (default: "marginalize")}

\item{target_category}{The index of the category for which categorization should be shown. (default: \code{1})}

\item{logit}{Should the categorization function be plotted in logit (\code{TRUE}) or probabilities (\code{FALSE})?
(default: \code{FALSE})}

\item{xlim, ylim}{Limits for the x- and y-axis.}

\item{resolution}{How many steps along x and y should be calculated? Note that computational
complexity increases quadratically with resolution. (default: 25)}

\item{facet_rows_by, facet_cols_by, facet_wrap_by, animate_by}{Which group variables, if any, should be used for faceting and/or
animation? (defaults: \code{NULL})}

\item{animation_follow}{Should the animation follow the data (zoom in and out)? (default: \code{FALSE})}

\item{category.ids}{Vector of category IDs to be plotted or leave \code{NULL} to plot all groups. (default: \code{NULL})}

\item{category.labels}{Vector of group labels of same length as \code{category.ids} or \code{NULL} to use defaults. (default: \code{NULL})}

\item{category.colors}{Vector of colors of same length as category.ids or \code{NULL} to use defaults. (default: \code{NULL})}

\item{...}{additional arguments to geom_line.}
}
\value{
ggplot object.
}
\description{
Plot categorization function for bivariate Gaussian categories expected given NIW parameters.
}
\seealso{
TBD
}
\keyword{TBD}
