% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-forward-ibbu.R
\name{plot_expected_categorization_function_2D}
\alias{plot_expected_categorization_function_2D}
\title{Plot expected categorization function for bivariate (2D) categories.}
\usage{
plot_expected_categorization_function_2D(
  x,
  facet_rows_by = NULL,
  facet_cols_by = NULL,
  animate_by = NULL,
  data.exposure = NULL,
  data.test = NULL,
  target_category = 1,
  xlim,
  ylim,
  resolution = 25,
  category.ids = NULL,
  category.labels = NULL,
  category.colors = NULL,
  category.linetypes = NULL
)
}
\arguments{
\item{x}{NIW belief object.}

\item{facet_rows_by}{Which group variables, if any, should be used for faceting and/or
animation? (defaults: `NULL`)}

\item{facet_cols_by}{Which group variables, if any, should be used for faceting and/or
animation? (defaults: `NULL`)}

\item{animate_by}{Which group variables, if any, should be used for faceting and/or
animation? (defaults: `NULL`)}

\item{data.exposure}{Optional \code{tibble} or \code{data.frame} that contains exposure data to be plotted. (default: `NULL`)}

\item{data.test}{Optional \code{tibble} or \code{data.frame} that contains test data to be plotted. (default: `NULL`)}

\item{target_category}{The index of the category for which categorization should be shown. (default: `1`)}

\item{xlim, ylim}{Limits for the x- and y-axis.}

\item{resolution}{How many steps along x and y should be calculated? Note that computational
complexity increases quadratically with resolution. (default: 25)}

\item{category.ids}{Vector of category IDs to be plotted or leave `NULL` to plot all groups. (default: `NULL`) It is possible
to use \code{\link[tidybayes]{recover_types}} on the stanfit object prior to handing it to this plotting function.}

\item{category.labels}{Vector of group labels of same length as `category.ids` or `NULL` to use defaults. (default: `NULL`)}

\item{category.colors}{Vector of colors of same length as category.ids or `NULL` to use defaults. (default: `NULL`)}

\item{category.linetypes}{Vector of linetypes of same length as category.ids or `NULL` to use defaults. (default: `NULL`)
Currently being ignored.}
}
\value{
ggplot object.
}
\description{
Plot categorization function for bivariate Gaussian categories expected given NIW parameters.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
