% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-forward-ibbu.R
\name{plot_expected_categorization_function_2D}
\alias{plot_expected_categorization_function_2D}
\title{Plot expected categorization function for bivariate (2D) categories.}
\usage{
plot_expected_categorization_function_2D(
  x,
  grouping.var = NULL,
  panel.group = F,
  animate.group = F,
  data.exposure = NULL,
  data.test = NULL,
  category.ids = NULL,
  category.labels = NULL,
  category.colors = NULL,
  category.linetypes = NULL
)
}
\arguments{
\item{x}{NIW belief object.}

\item{grouping.var}{Grouping variable that is used to create separate instances of the categories. These instances
can be plotted in separate panels (if panel.group is `TRUE`) or used to create animates (if animate.group is `TRUE`).}

\item{panel.group, animate.group}{Determines whether the grouping variable is for paneling or animation. (both defaults: `FALSE`)}

\item{data.exposure}{Optional \code{tibble} or \code{data.frame} that contains exposure data to be plotted. (default: `NULL`)}

\item{data.test}{Optional \code{tibble} or \code{data.frame} that contains test data to be plotted. (default: `NULL`)}

\item{category.ids}{Vector of category IDs to be plotted or leave `NULL` to plot all groups. (default: `NULL`) It is possible
to use \code{\link[tidybayes]{recover_types}} on the stanfit object prior to handing it to this plotting function.}

\item{category.labels}{Vector of group labels of same length as `category.ids` or `NULL` to use defaults. (default: `NULL`)}

\item{category.colors}{Vector of colors of same length as category.ids or `NULL` to use defaults. (default: `NULL`)}

\item{category.linetypes}{Vector of linetypes of same length as category.ids or `NULL` to use defaults. (default: `NULL`)
Currently being ignored.}
}
\value{
ggplot object.
}
\description{
Plot categorization function for bivariate Gaussian categories expected given NIW parameters.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
