% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-NIW-IA-stanfit.R
\name{plot_ibbu_stanfit_parameter_correlations}
\alias{plot_ibbu_stanfit_parameter_correlations}
\title{Plot correlations between IBBU parameters.}
\usage{
plot_ibbu_stanfit_parameter_correlations(
  model,
  categories = get_category_levels_from_stanfit(model),
  groups = "prior",
  cues = get_cue_levels_from_stanfit(model),
  pars = NULL,
  ndraws = NULL,
  untransform_cues = TRUE,
  category.colors = get_default_colors("category", categories)
)
}
\arguments{
\item{model}{mv-ibbu-stanfit object.}

\item{categories, groups, cues, pars}{Character vector of categories, groups, cues, and/or parameters to be plotted. To
subset to specific parameters, note the following naming conventions:

(1) all `m` parameters are named `m_{cue_name}`
(2) the `S` parameter is decomposed into a vector of standard deviations `tau` and a correlation matrix `rho`.
(3) all `tau` parameters are named `tau_{cue_name}`
(4) all `rho` parameters are named `rho_{cue_name1}__x__{cue_name2}`

(default: all categories, cues, and parameters in the model; `"prior"` for group since those are the only free parameters).}

\item{ndraws}{Number of draws to plot (or use to calculate the CIs), or `NULL` if all draws are to be returned. (default: `NULL`)}

\item{untransform_cues}{Should m_0 and S_0 be transformed back into the original cue space? (default: `TRUE`)}

\item{category.colors}{Vector of fill colors of same length as category or `NULL` to use defaults. (default: `NULL`)}
}
\value{
ggplot object.
}
\description{
Plot correlations between post-warmup MCMC samples for all parameters representing the prior and/or posterior beliefs.
}
\details{
Typically, the categories, groups, and cues are automatically added to the fit during the creation of the fit. If necessary,
however, it is possible to use \code{\link[tidybayes]{recover_types}} on the stanfit object to add or change these levels later.
}
\seealso{
TBD
}
\keyword{TBD}
