% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-NIW-IA-stanfit.R
\name{plot_ibbu_stanfit_parameter_correlations}
\alias{plot_ibbu_stanfit_parameter_correlations}
\title{Plot correlations between IBBU parameters.}
\usage{
plot_ibbu_stanfit_parameter_correlations(
  model,
  categories = get_category_levels_from_stanfit(model),
  groups = "prior",
  cues = get_cue_levels_from_stanfit(model),
  ndraws = NULL,
  untransform_cues = TRUE,
  category.colors = get_default_colors("category", categories)
)
}
\arguments{
\item{model}{mv-ibbu-stanfit object.}

\item{categories, groups, cues}{Character vector of categories/groups/cues to be plotted. Typically, the levels of these factors
are automatically added to the fit during the creation of the fit. If necessary, however, it is possible to use
\code{\link[tidybayes]{recover_types}} on the stanfit object to add or change these levels later.
(default: all categories and cues; `"prior"` for group since those are the only free parameters)}

\item{ndraws}{Number of draws to plot (or use to calculate the CIs), or `NULL` if all draws are to be returned. (default: `NULL`)}

\item{untransform_cues}{Should m_0 and S_0 be transformed back into the original cue space? (default: `TRUE`)}

\item{category.colors}{Vector of fill colors of same length as category or `NULL` to use defaults. (default: `NULL`)}
}
\value{
ggplot object.
}
\description{
Plot correlations between post-warmup MCMC samples for all parameters representing the prior and/or posterior beliefs.
}
\examples{
TBD

}
\seealso{
TBD
}
\keyword{TBD}
