% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-stanfit-ibbu.R
\name{plot_ibbu_stanfit_parameters}
\alias{plot_ibbu_stanfit_parameters}
\title{Plot distribution of IBBU parameters.}
\usage{
plot_ibbu_stanfit_parameters(
  fit,
  which = "both",
  n.draws = NULL,
  group.ids = NULL,
  group.labels = NULL,
  group.colors = NULL,
  panel_scaling = F
)
}
\arguments{
\item{fit}{mv-ibbu-stanfit object.}

\item{which}{Should parameters for the prior, posterior, or both be plotted? (default: `"both"`)}

\item{n.draws}{Number of draws to plot (or use to calculate the CIs), or `NULL` if all draws are to be returned. (default: `NULL`)}

\item{group.ids}{Vector of group IDs to be plotted or leave `NULL` to plot all groups. (default: `NULL`) It is possible
to use \code{\link[tidybayes]{recover_types}} on the stanfit object prior to handing it to this plotting function.}

\item{group.labels}{Vector of group labels of same length as group.ids or `NULL` to use defaults. (default: `NULL`)
The default labels each categorization function based on whether it is showing prior or posterior categorization,
and by its group ID.}

\item{group.colors}{Vector of fill colors of same length as group.ids or `NULL` to use defaults. (default: `NULL`)}

\item{panel_scaling}{Should the relative scaling be calculated separately for each panel? If not the scaling is calculated
globally. (default: `FALSE`)}
}
\value{
ggplot object.
}
\description{
Plot distribution of post-warmup MCMC samples for all parameters representing the
prior and/or posterior beliefs.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
