% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-stanfit-ibbu.R
\name{plot_ibbu_stanfit_test_categorization}
\alias{plot_ibbu_stanfit_test_categorization}
\title{Plot prior and posterior categorization of test tokens.}
\usage{
plot_ibbu_stanfit_test_categorization(
  model,
  data.test = NULL,
  which = "both",
  summarize = T,
  ndraws = NULL,
  untransform_cues = TRUE,
  confidence.intervals = c(0.66, 0.95),
  target_category = 1,
  panel.group = if (plot_in_cue_space) TRUE else FALSE,
  group.ids = NULL,
  group.labels = NULL,
  group.colors = NULL,
  group.linetypes = NULL,
  category.ids = NULL,
  category.colors = NULL,
  all_test_locations = TRUE,
  plot_in_cue_space = FALSE,
  sort_by = if (plot_in_cue_space) NULL else "prior"
)
}
\arguments{
\item{model}{\code{\link{NIW_ideal_adaptor_stanfit}} object.}

\item{data.test}{Optionally, a \code{tibble} or \code{data.frame} with test data.
If `NULL` the input will be extracted from fit. (default: `NULL`).}

\item{which}{Should categorization for the prior, posterior, or both be plotted? (default: `"both"`)}

\item{summarize}{Should one categorization function (optionally with CIs) be plotted (`TRUE`) or should separate
unique categorization function be plotted for each MCMC draw (`FALSE`)? (default: `TRUE`)}

\item{ndraws}{Number of draws to plot (or use to calculate the CIs), or `NULL` if all draws are to be returned.
(default: `NULL`)}

\item{confidence.intervals}{The two confidence intervals that should be plotted (using `geom_ribbon`) around the mean.
(default: `c(.66, .95)`)}

\item{target_category}{The index of the category for which categorization should be shown. (default: `1`)}

\item{panel.group}{Should the groups be plotted in separate panels? (default: `FALSE`)}

\item{group.ids}{Vector of group IDs to be plotted or leave `NULL` to plot all groups. (default: `NULL`)}

\item{group.labels}{Vector of group labels of same length as `group.ids` or `NULL` to use defaults. (default: `NULL`)
The default labels each categorization function based on whether it is showing prior or posterior categorization,
and by its group ID.}

\item{group.colors, group.linetypes}{Vector of colors and linetypes of same length as group.ids or `NULL` to use defaults.
(default: `NULL`)}

\item{category.colors}{Vector of colors and linetypes of same length as category.ids or `NULL` to use defaults. Only
relevant when `plot_in_cue_space = TRUE`. (default: `NULL`)}

\item{all_test_locations}{Should predictions be shown for all combinations of test locations and group, or should only
combinations be shown that actually occurred in the data? (default: `FALSE`)}

\item{plot_in_cue_space}{Should predictions be plotted in the cue space? If not, test tokens are essentially treated
as factors and sorted along the x-axis based on `sort_by`. (default: `TRUE`)}

\item{sort_by}{Which group, if any, should the x-axis be sorted by (in increasing order of posterior probability
from left to right). Set to 0 for sorting by prior (default). Set to `NULL` if no sorting is desired. (default: `"prior"`)}
}
\value{
ggplot object.
}
\description{
Categorize test tokens by prior and/or posterior beliefs, and plot the resulting categorization function along
a one-dimensional continuum (regardless of the dimensionality of the cue space in which categorization takes
place). This provides the type of categorization plot typical for, for example, perceptual recalibration or
phonetic tuning studies.
}
\details{
Tokens are sorted based on the increasing probability of a \code{target_category} response for the condition
(\code{group}, e.g., prior or a specific exposure group) specified in \code{sort_by}. By default both the mean
categorization and confidence intervals are plotted.
If `summarize=TRUE`, the function marginalizes over all posterior samples. The number of samples
is determined by ndraws. If ndraws is NULL, all samples are used. Otherwise ndraws random
samples will be used. If `summarize=FALSE`, separate categorization plots for all ndraws
individual samples will be plotted in separate panels.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
