% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-basics.R
\name{plot_pairwise_cue_correlation_matrix}
\alias{plot_pairwise_cue_correlation_matrix}
\title{Plot pairwise cue correlation matrix}
\usage{
plot_pairwise_cue_correlation_matrix(
  data,
  cues,
  category = category,
  category.colors = 1:length(unique(data$category))
)
}
\arguments{
\item{data}{A \code{tibble} or \code{data.frame} that contains the \code{cues} and \code{category}.}

\item{cues}{Variables in \code{data} -containing the cue values that are to be plotted.}

\item{category}{Variable in \code{data} that contains the category information. (currently required)}

\item{category.colors}{The colors to be mapped onto the category values.}
}
\value{
A ggplot2 plot.
}
\description{
Plots a n x n correlation matrix, using \link[ggforce:facet_matrix]{ggforce::facet_matrix}, with univariate densities for each category on
the diagonal, scatter plots in the lower triangle and 95\\% bivariate Gaussian ellipsoids on the upper diagonal.
}
\seealso{
TBD
}
\keyword{TBD}
