% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-basics.R
\name{plot_pairwise_cue_correlation_matrix}
\alias{plot_pairwise_cue_correlation_matrix}
\title{Plot pairwise cue correlation matrix}
\usage{
plot_pairwise_cue_correlation_matrix(
  data,
  cues,
  category = category,
  category.colors = 1:length(unique(data$category))
)
}
\arguments{
\item{data}{A `tibble` or `data.frame` that contains the `cues` and `category`.}

\item{cues}{Variables in `data` -containing the cue values that are to be plotted.}

\item{category}{Variable in `data` that contains the category information. (currently required)}

\item{category.colors}{The colors to be mapped onto the category values.}
}
\value{
A ggplot2 plot.
}
\description{
Plots a n x n correlation matrix, using ggforce::facet_matrix, with univariate densities for each category on
the diagonal, scatter plots in the lower triangle and 95% bivariate Gaussian ellipsoids on the upper diagonal.
}
\examples{
TBD

}
\seealso{
TBD
}
\keyword{TBD}
