% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-class.R
\name{print.MVBU_model}
\alias{print.MVBU_model}
\title{Deprecated: print.MVBU_model}
\usage{
\method{print}{MVBU_model}(x, ...)
}
\arguments{
\item{x}{An \code{MVBeliefUpdatr_model} object.}

\item{...}{Additional arguments.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{print.MVBU_model()} was deprecated in MVBeliefUpdatr 0.1.0 and will be removed in 0.2.0.
Please use S7 print methods instead.
}
\seealso{
\link{MVBU_CognitiveModel}
}
\keyword{internal}
