% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-classes.R, R/s7-foundation.R
\name{register_model_family}
\alias{register_model_family}
\title{Register an MVBU model family}
\usage{
register_model_family(
  family,
  category_representation_class,
  cognitive_model_class,
  model_distribution_class
)

register_model_family(
  family,
  category_representation_class,
  cognitive_model_class,
  model_distribution_class
)
}
\description{
Register an MVBU model family

Register an MVBU model family
}
\keyword{internal}
