% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-classes.R
\name{register_stan_family_hooks}
\alias{register_stan_family_hooks}
\title{Register Stan-family extension hooks}
\usage{
register_stan_family_hooks(
  family,
  stanfit_class = NULL,
  staninput_class = NULL,
  bridge_methods = character(0),
  dependency_rationale = character(0)
)
}
\description{
Register Stan-family extension hooks
}
\keyword{internal}
