% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R
\name{set_stanfit}
\alias{set_stanfit}
\title{Set the stored Stan fit on an S7 fit object}
\usage{
set_stanfit(x, stanfit, ...)
}
\arguments{
\item{x}{An S7 fit object.}

\item{stanfit}{An \link[rstan:stanfit]{rstan::stanfit} object.}
}
\value{
The updated S7 fit object.
}
\description{
Set the stored Stan fit on an S7 fit object
}
