% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{set_stanfit}
\alias{set_stanfit}
\title{Set the stanfit of an ideal adaptor stanfit}
\usage{
set_stanfit(x, stanfit = NULL)
}
\arguments{
\item{x}{\code{\link{ideal_adaptor_stanfit}} object.}

\item{stanfit}{An \link[rstan:stanfit]{rstan::stanfit} object of adequate structure.}
}
\value{
An \code{\link{ideal_adaptor_stanfit}} object with the updated stanfit.
}
\description{
Sets the stanfit of an the \code{\link{ideal_adaptor_stanfit}} object.
}
\seealso{
TBD
}
\keyword{TBD}
