% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{set_stanfit}
\alias{set_stanfit}
\alias{set_stanfit.NIW_ideal_adaptor_stanfit}
\title{Set the stanfit of an NIW ideal adaptor stanfit}
\usage{
set_stanfit(x, ...)

\method{set_stanfit}{NIW_ideal_adaptor_stanfit}(x, stanfit = NULL)
}
\arguments{
\item{x}{\code{\link{NIW_ideal_adaptor_stanfit}} object.}

\item{stanfit}{An \code{\link{rstan::stanfit}} object of adequate structure.}
}
\value{
An \code{\link{NIW_ideal_adaptor_stanfit}} object with the updated stanfit.
}
\description{
Sets the stanfit of an the \code{\link{NIW_ideal_adaptor_stanfit}} object.
}
\seealso{
TBD
}
\keyword{TBD}
