% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info-from-NIW-IA-stanfit.R
\name{set_staninput}
\alias{set_staninput}
\alias{set_staninput,ideal_adaptor_stanfit-method}
\title{Set the input data from an ideal adaptor stanfit}
\usage{
set_staninput(
  x,
  staninput,
  which = c("untransformed", "transformed", "both")[3]
)

\S4method{set_staninput}{ideal_adaptor_stanfit}(
  x,
  staninput,
  which = c("untransformed", "transformed", "both")[3]
)
}
\arguments{
\item{x}{\code{\link{ideal_adaptor_stanfit}} object.}

\item{staninput}{A list with element names and structures determined by the type of stanfit model.}

\item{which}{Should "transformed" or "untransformed" staninput be set or "both"? (default: \code{"both"})}
}
\value{
A list with element names and structures determined by the type of stanfit model.

An \code{\link{ideal_adaptor_stanfit}} object with the updated staninput.
}
\description{
Sets the inputs handed to \code{stan} or \code{sampling} during the creation of the \code{\link{ideal_adaptor_stanfit}}
object.
}
\seealso{
TBD
}
\keyword{TBD}
