% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-NIW-IA-stanfit.R
\name{summary.ideal_adaptor_stanfit}
\alias{summary.ideal_adaptor_stanfit}
\title{Summarize NIW ideal adaptor stanfit}
\usage{
\method{summary}{ideal_adaptor_stanfit}(
  x,
  pars = NULL,
  indices_as_names = TRUE,
  sufficient_only = FALSE,
  include_transformed_pars = F,
  ...
)
}
\arguments{
\item{x}{An \code{\link{ideal_adaptor_stanfit}} object.}

\item{pars}{A character vector of parameter names to be summarized. If \code{NULL}, all parameters of the
NIW_ideal_adaptor are summarized. (default: \code{NULL})}

\item{indices_as_names}{Should the indices of the parameters be translated into level names in the summary?
This output is substantially more readable, but might lend itself less for post-processing. The
names will be shown in separate columns (distribution, group, category, cue, ...). (default: \code{TRUE})}

\item{sufficient_only}{Should only the sufficient parameters be summarized? (default: \code{FALSE})}

\item{include_transformed_pars}{Should transformed parameters be included in the summary? (default: \code{FALSE})}

\item{...}{Additional arguments passed to \link[rstan:stanfit-class]{rstan::summary}.}}

\description{
\code{summary} method for \code{\link{ideal_adaptor_stanfit}} objects. Specifies reasonable defaults
for the parameters to be summarized for the stanfit object.
}
