% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-NIW-IA.R
\name{summary.NIW_ideal_adaptor_stanfit}
\alias{summary.NIW_ideal_adaptor_stanfit}
\title{Summarize NIW ideal adaptor stanfit}
\usage{
\method{summary}{NIW_ideal_adaptor_stanfit}(x, pars = NULL, ...)
}
\arguments{
\item{x}{An \code{\link{NIW_ideal_adaptor_stanfit}} object.}
}
\description{
\code{summary} method for \code{\link{NIW_ideal_adaptor_stanfit}} objects. Specifies reasonable defaults for the parameters to be summarized for the stanfit object.
}
