% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-staninput.R
\name{to_array}
\alias{to_array}
\title{Turn list into array (for Stan input)}
\usage{
to_array(x, inner_dims = 1, outer_dims = NULL)
}
\arguments{
\item{x}{The input to be turned into an array.}

\item{inner_dims}{Intended dimensions of x. Will be used for checks and to enforce the correct dimensions when `x` is `NULL`.
(default: 1)}

\item{outer_dims}{A vector of outer dimensions for the array. If `x` is not a list, an array of `x` will be repeated for each
outer dimension. If `x` is list, the outer dimensions will each index one element of `x`.
The total length of `x` must match the product of outer dimensions. The first dimension will be iterated
over first (and thus alternating fastest), the second dimension will be iterated over second, etc. See details.}
}
\value{
An array.
}
\description{
Takes `NULL`, scalars, or lists as inputs and turns them into numeric arrays.
}
\details{
\itemize{
  \item{`NULL` will be turned in a `array(numeric(0), dim = c(.outer_dims, .inner_dims))`. This is sometimes required for
  optional Stan inputs.}
  \item{Scalars and vectors will be turned into arrays of appropriate dimensions.}
  \item{Lists will be turned into arrays of appropriate dimensions.}
}

E.g., if the input is a list of six 2x2 covariance matrices, and `...` is `3, 2`, then output will
be an array of dimensionality `c(3, 2, 2, 2)`.
}
\keyword{TBD}
