% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose-input-for-stan.R
\name{transform_cues}
\alias{transform_cues}
\alias{untransform_cues}
\title{Transform and untransform cues by applying or undoing PCA, centering, and/or scaling. If the `transform` argument
is specified, the transforms in that object will be applied. This can be useful when the goal is to transform one
data set (e.g., test data) based on the statistics of the another data set (e.g.., training data). If no `transform`
is specified, then the transformations specified by the `center`, `scale`, and `PCA` flags will be applied.}
\usage{
transform_cues(
  data,
  cues,
  transform = NULL,
  return.transform = F,
  center = T,
  scale = T,
  PCA = F
)

untransform_cues(
  data,
  cues,
  transform = NULL,
  center = NULL,
  scale = NULL,
  PCA = NULL
)
}
\arguments{
\item{data}{`tibble` or `data.frame`.}

\item{cues}{Vector of characters with names of cue variables.}

\item{transform}{List of transforms (default: `NULL`)}

\item{return.transform}{Should the list of transforms be returned along with the data? (default: `FALSE`)}

\item{center}{Should the data be centered? (default: `TRUE`)}

\item{scale}{Should the data be standardized? (default: `TRUE`)}

\item{PCA}{Should the data be transformed into orthogonal principal components? (default: `FALSE`)}
}
\value{
Data frame, unless `return.transform = T`. In that case, a list with two elements (`data` and `transform`).
}
\description{
Transform and untransform cues by applying or undoing PCA, centering, and/or scaling. If the `transform` argument
is specified, the transforms in that object will be applied. This can be useful when the goal is to transform one
data set (e.g., test data) based on the statistics of the another data set (e.g.., training data). If no `transform`
is specified, then the transformations specified by the `center`, `scale`, and `PCA` flags will be applied.
}
\examples{
TBD
}
\seealso{
TBD
}
\keyword{TBD}
