% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{transform_model}
\alias{transform_model}
\alias{untransform_model}
\alias{transform_category_mean}
\alias{untransform_category_mean}
\alias{transform_category_cov}
\alias{untransform_category_cov}
\title{Transform and untransform category means and covariance matrices of a model}
\usage{
transform_model(model, transform)

untransform_model(model, transform)

transform_category_mean(m, transform)

untransform_category_mean(m, transform)

transform_category_cov(S, transform)

untransform_category_cov(S, transform)
}
\arguments{
\item{model}{A model with columns that specify category means (mu or m) and covariance information (Sigma or S).}

\item{transform}{A transform or transform parameter object of the type returned by \code{\link{transform_cues}}.}

\item{m}{A single category mean or alike.}

\item{S}{A single covariance matrix or alike.}
}
\value{
A model, category mean, or covariance matrix of the same type as the input.
}
\description{
Applies the transformation specified in \code{transform} object (e.g., centering, scaling, PCA) to the category
mean(s) and/or covariance(s) in the model.
}
\details{
Note that \code{(un)transform_category_cov} can also be used for centered sums-of-squares matrices, but not for
uncentered sums-of-squares matrices. The latter require additional adjustments that might be implemented in the
future (see https://chatgpt.com/c/683b3ed6-4924-800c-8f22-e61680f3360f).
}
\keyword{TBD}
