% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{update_NIW_belief_by_sufficient_statistics}
\alias{update_NIW_belief_by_sufficient_statistics}
\alias{update_NIW_belief_by_one_observation}
\title{Update NIW prior beliefs about multivariate Gaussian category based on sufficient statistics of observations.}
\usage{
update_NIW_belief_by_sufficient_statistics(
  prior,
  x_category,
  x_mean,
  x_S,
  x_N,
  add_noise = NULL,
  method = "label-certain"
)

update_NIW_belief_by_one_observation(
  prior,
  x_category,
  x,
  add_noise = NULL,
  method = "label-certain"
)
}
\arguments{
\item{prior}{An \code{\link[=is.NIW_belief]{NIW_belief}} object, specifying the prior beliefs.}

\item{x_category}{Character. The label of the category that is to be updated.}

\item{x_mean}{Mean of the observations.}

\item{x_S}{Centered sum of squares matrix of the observations.}

\item{x_N}{Number of observations that went into the mean and sum of squares matrix.}

\item{add_noise}{Determines whether multivariate Gaussian noise is added to the input.
If `NULL`, no noise is added during the updating. If "sample" then a sample of
noise is added to the input. If "marginalize" then each observation is transformed into the marginal distribution
that results from convolving the input with noise. This latter option might be helpful, for example, if one is
interested in estimating the consequences of noise across individuals. If add_noise is not `NULL` a Sigma_noise
column must be present in the NIW_belief object specified as the priors argument. (default: `NULL`)}

\item{method}{Which updating method should be used? See details. (default: "supervised-certain")}

\item{x}{A single observation.}
}
\value{
An NIW_belief object.
}
\description{
Returns updated/posterior beliefs about the Gaussian categories based on conjugate NIW prior.
}
\details{
The priors for the categories are specified through the \code{priors} argument, an
\code{\link[=is.NIW_belief]{NIW_belief}} object. Which category is updated is determined by x_category.
Updating proceeds as in \insertCite{@see @murphy2012 p. 134;textual}{MVBeliefUpdatr}. The prior kappa
and nu will be incremented by the number of observations (x_N). The prior M and S will be updated based on the
prior kappa, prior nu, x_N and, of course, the sample mean (x_mean) and sum of squares (x_S) of the observations.

A number of different updating schemes are supported, including supervised updating based on labeled data and
unsupervised updating based on unlabeled data. Except for "nolabel-sampling", the unsupervised updating rules
were originally presented in \insertCite{yan:jaeger2018;textual}{MVBeliefUpdatr}.
\itemize{
  \item "no-updating" doesn't update the prior. Combined with keep_history = T, this allows the creation of baseline
  against which to compare the updated beliefs. This option is likely most useful when used as part of a call to
  \code{\link{update_NIW_beliefs_incrementally}}.
  \item "label-certain" assumes that the label is provided and known to the observer with 100% certainty, resulting
  in fully Bayesian supervised belief-updating. This update method is \emph{not order sensitive}. Under this update
  method, the order of a batch of observations does not affect the final posterior belief after all observations
  have been made (it does, of course, effect the beliefs held in the interim). This reflects the "exchangeability"
  assumption of Bayesian belief-updating under the assumption that the observations are drawn from a stationary
  random process.
  \item "nolabel-criterion" implements a winner-takes-all update based on the prior beliefs. The
  posterior probability of all categories under the prior is calculated, and only the category with the
  highest posterior probabilty is updated (see decision_rule = "criterion" in \code{\link{get_categorization_from_NIW_belief}}).
  The update for this category proceed the same ways as under the "label-certain" method. Method "nolabel-criterion" is
  thus \emph{order sensitive}, but deterministic (i.e., it yields the same result on each repeated run), provided
  there is no noise (specifically, no noise = "sample").
  \item "nolabel-sampling" implements another winner-takes-all update based on the prior beliefs. The
  posterior probability of all categories under the prior is calculated. Then a single category is \emph{sampled}
  based on this posterior (see decision_rule = "sampling" in \code{\link{get_categorization_from_NIW_belief}}), and only this
  sampled category is  updated with the observation.
  Like "nolabel-criterion", this updating method is \emph{order sensitive}. Unlike "nolabel-criterion", this method had a
  random element and thus does \emph{not replicate the same result on each run}.
  \item "nolabel-proportional" captures the uncertainty about the category label (and is this sense fully Bayesian). The
  posterior probability of all categories under the prior is calculated (see decision_rule = "proportional" in
  \code{\link{get_categorization_from_NIW_belief}}). The input is then distributed across all categories
  based on their posterior probability under the prior beliefs. That is, \emph{all categories are updated}, but to different
  degrees. Like "nolabel-criterion" and "nolabel-sampling", this updating method is \emph{order sensitive}. Like "nolabel-criterion",
  it is \emph{deterministic}.
  \item "nolabel-uniform" implements unsupervised updating under maximal \emph{uninformed} uncertainty. The input is attributed
  to equal parts to all categories. Like all the "nolabel-*" methods, this updating method is \emph{order sensitive}. Like
  "nolabel-criterion" and "nolabel-proportional, it is \emph{deterministic}. Unlike any of these methods, it completely ignores
  the knowledge listeners have (i.e., the prior).
}
\strong{Please consider the following important limitations of these functions:}
\itemize{
  \item the category priors---p(category)---are currently assumed to be uniform. This affects the categorization of the input
based on the NIW prior and thus the way that the input observations are distributed across the categories during
unsupervised updating.
  \item the updating models introduced here treat the cue distributions of the different categories as \emph{independent} of
  each other. That is, the updating does not consider knowledge about, for example, the covariance of the category means, although
  there is evidence that listeners have knowledge of this covariance and draw on it during categorization and adaptation.
}
Please feel free to suggest additional features.
}
\examples{
TBD
}
\references{
\insertRef{murphy2012}{MVBeliefUpdatr}
}
\seealso{
\code{\link{update_NIW_belief_kappa}}, \code{\link{update_NIW_belief_nu}}, \code{\link{update_NIW_belief_M}},
\code{\link{update_NIW_belief_S}}, all of which are called by \code{update_NIW_belief_by_sufficient_statistics}.
}
\keyword{NIW}
\keyword{Normal-Inverse}
\keyword{Wishart}
\keyword{belief-updating}
