% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-update-NIW-IA.R
\name{update_NIW_belief_by_sufficient_statistics_of_one_category}
\alias{update_NIW_belief_by_sufficient_statistics_of_one_category}
\title{Deprecated: update_NIW_belief_by_sufficient_statistics_of_one_category}
\usage{
update_NIW_belief_by_sufficient_statistics_of_one_category(
  prior_model,
  x_category = NULL,
  x_mean,
  x_SS,
  x_N,
  ...
)
}
\arguments{
\item{prior_model}{An NIW S7 category representation.}

\item{x_category}{Deprecated and ignored for S7 representations.}

\item{x_mean}{Observation mean.}

\item{x_SS}{Centered observation sum-of-squares matrix.}

\item{x_N}{Number of observations.}

\item{...}{Deprecated compatibility arguments.}
}
\value{
An updated NIW category representation.
}
\description{
Deprecated: update_NIW_belief_by_sufficient_statistics_of_one_category
}
