% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{update_NIW_beliefs}
\alias{update_NIW_beliefs}
\title{Update NIW prior beliefs about multivariate Gaussian category based on exposure data.}
\usage{
update_NIW_beliefs(
  data,
  priors,
  category = "category",
  cues = names(priors$M[[1]]),
  store.history = T
)
}
\arguments{
\item{data}{\code{data.frame} or \code{tibble} with exposure data. Each row is assumed to contain one observation.}

\item{priors}{A tibble with information about the prior. See Details for expected format of the \code{priors} argument.}

\item{category}{Name of variable in \code{data} that contains the category information. (default: "category")}

\item{cues}{Name(s) of variables in \code{data} that contain the cue information. By default these cue names are
extracted from the prior object.}
}
\value{
A tibble.
}
\description{
Returns updated/posterior beliefs about the Gaussian categories based on conjugate NIW prior.
}
\details{
The priors for the categories are specified through the \code{priors} argument. This is expected to be a tibble
of the same format as the posterior draws stored in an MV IBBU stanfit object. Each row of the tibble specifies
the prior for one category (specified in the \code{category} column). The four parameters of the NIW are the
pseudocounts indicating the strength of the prior beliefs into the mean (\code{kappa}) and covariance (\code{
nu}), as well as the prior mean of means (\code{M}, same as \code{M_0} in Murphy, 2012) and the prior scatter
matrix (\code{S}, same as \code{S_0} in Murphy, 2012).
}
\examples{
TBD
}
\references{
Murphy, K. P. (2012). Machine learning: a probabilistic perspective. MIT press.
}
\seealso{
TBD
}
\keyword{TBD}
