% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{update_NIW_belief_kappa}
\alias{update_NIW_belief_kappa}
\title{Update parameters of NIW prior beliefs about multivariate Gaussian category.}
\usage{
update_NIW_belief_kappa(kappa_0, x_N)
}
\description{
Returns updated/posterior M, S, kappa, or nu based on \insertCite{@see @murphy2012 p. 134;textual}{MVBeliefUpdatr}.
}
