% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-NIW-IA.R
\name{update_NIW_belief_kappa}
\alias{update_NIW_belief_kappa}
\alias{update_NIW_belief_nu}
\alias{update_NIW_belief_m}
\alias{update_NIW_belief_S}
\title{Update parameters of NIW prior beliefs about multivariate Gaussian category.}
\usage{
update_NIW_belief_kappa(kappa_0, x_N)

update_NIW_belief_nu(nu_0, x_N)

update_NIW_belief_m(kappa_0, m_0, x_N, x_mean)

update_NIW_belief_S(kappa_0, m_0, S_0, x_N, x_mean, x_SS)
}
\arguments{
\item{kappa_0, nu_0, m_0, S_0}{Prior parameter values of NIW.}

\item{x_n, x_mean, x_SS}{Number of observations, mean of those observations, and \emph{centered} sum-of-square matrix of those
observations based on which the parameters of the NIW should be updated.}
}
\value{
Depending on the updated parameter, a numeric scalar (kappa, nu), vector (m) or square matrix (S).
}
\description{
Returns updated/posterior m, S, kappa, or nu based on \insertCite{@see @murphy2012 p. 134;textual}{MVBeliefUpdatr}.
These functions are called, for example, by \code{\link{update_NIW_belief_by_sufficient_statistics_of_one_category}},
\code{\link{update_NIW_belief_by_one_observation}}, and \code{\link{update_NIW_beliefs_incrementally}}.
}
