% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7-generics.R
\name{update_template}
\alias{update_template}
\title{Update a model's category-representation template from observations}
\usage{
update_template(x, observations, ...)
}
\arguments{
\item{x}{A model object.}

\item{observations}{Observations used for updating.}
}
\value{
An updated model, or a history of updated models when requested by a method.
}
\description{
Currently implemented for \code{NIW_IdealAdaptor}; methods for other model types
will be added as their update workflows are migrated to S7.
}
