% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-foundation.R
\name{UVG-IdealObserverDistribution-class}
\alias{UVG-IdealObserverDistribution-class}
\alias{UVG_IdealObserverDistribution}
\title{UVG Ideal Observer Distribution Class}
\usage{
UVG_IdealObserverDistribution(
  model_family = character(0),
  cache = list(),
  metadata = list(),
  group_label = character(0)
)
}
\description{
Family-typed model-distribution class requiring \code{model_family = "UVG"}.
}
\seealso{
\link{UVG_IdealObserver}
}
\keyword{internal}
